% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessibility_commands.R
\name{Accessibility.getFullAXTree}
\alias{Accessibility.getFullAXTree}
\title{Send the command Accessibility.getFullAXTree}
\usage{
Accessibility.getFullAXTree(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Fetches the entire accessibility tree
}
