% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessibility_commands.R
\name{Accessibility.getPartialAXTree}
\alias{Accessibility.getPartialAXTree}
\title{Send the command Accessibility.getPartialAXTree

Fetches the accessibility node and partial accessibility tree for this DOM node, if it exists.}
\usage{
Accessibility.getPartialAXTree(promise, nodeId = NULL,
  backendNodeId = NULL, objectId = NULL, fetchRelatives = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{Optional. A DOM.NodeId.
Identifier of the node to get the partial accessibility tree for.}

\item{backendNodeId}{Optional. A DOM.BackendNodeId.
Identifier of the backend node to get the partial accessibility tree for.}

\item{objectId}{Optional. A Runtime.RemoteObjectId.
JavaScript object id of the node wrapper to get the partial accessibility tree for.}

\item{fetchRelatives}{Optional. A logical.
Whether to fetch this nodes ancestors, siblings and children. Defaults to true.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Accessibility.getPartialAXTree

Fetches the accessibility node and partial accessibility tree for this DOM node, if it exists.
}
