% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animation_commands.R
\name{Animation.disable}
\alias{Animation.disable}
\title{Send the command Animation.disable

Disables animation domain notifications.}
\usage{
Animation.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Animation.disable

Disables animation domain notifications.
}
