% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animation_commands.R
\name{Animation.getPlaybackRate}
\alias{Animation.getPlaybackRate}
\title{Send the command Animation.getPlaybackRate}
\usage{
Animation.getPlaybackRate(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Gets the playback rate of the document timeline.
}
