% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animation_commands.R
\name{Animation.releaseAnimations}
\alias{Animation.releaseAnimations}
\title{Send the command Animation.releaseAnimations}
\usage{
Animation.releaseAnimations(promise, animations)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{animations}{A list of string.
List of animation ids to seek.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Releases a set of animations to no longer be manipulated.
}
