% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animation_commands.R
\name{Animation.resolveAnimation}
\alias{Animation.resolveAnimation}
\title{Send the command Animation.resolveAnimation

Gets the remote object of the Animation.}
\usage{
Animation.resolveAnimation(promise, animationId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{animationId}{A character string.
Animation id.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Animation.resolveAnimation

Gets the remote object of the Animation.
}
