% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animation_commands.R
\name{Animation.seekAnimations}
\alias{Animation.seekAnimations}
\title{Send the command Animation.seekAnimations

Seek a set of animations to a particular time within each animation.}
\usage{
Animation.seekAnimations(promise, animations, currentTime)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{animations}{A list of string.
List of animation ids to seek.}

\item{currentTime}{A numeric.
Set the current time of each animation.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Animation.seekAnimations

Seek a set of animations to a particular time within each animation.
}
