% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animation_commands.R
\name{Animation.setPaused}
\alias{Animation.setPaused}
\title{Send the command Animation.setPaused}
\usage{
Animation.setPaused(promise, animations, paused)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{animations}{A list of string.
Animations to set the pause state of.}

\item{paused}{A logical.
Paused state to set to.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sets the paused state of a set of animations.
}
