% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animation_commands.R
\name{Animation.setPlaybackRate}
\alias{Animation.setPlaybackRate}
\title{Send the command Animation.setPlaybackRate}
\usage{
Animation.setPlaybackRate(promise, playbackRate)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{playbackRate}{A numeric.
Playback rate for animations on page}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sets the playback rate of the document timeline.
}
