% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animation_commands.R
\name{Animation.setTiming}
\alias{Animation.setTiming}
\title{Send the command Animation.setTiming

Sets the timing of an animation node.}
\usage{
Animation.setTiming(promise, animationId, duration, delay)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{animationId}{A character string.
Animation id.}

\item{duration}{A numeric.
Duration of the animation.}

\item{delay}{A numeric.
Delay of the animation.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Animation.setTiming

Sets the timing of an animation node.
}
