% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationCache_commands.R
\name{ApplicationCache.enable}
\alias{ApplicationCache.enable}
\title{Send the command ApplicationCache.enable

Enables application cache domain notifications.}
\usage{
ApplicationCache.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command ApplicationCache.enable

Enables application cache domain notifications.
}
