% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationCache_commands.R
\name{ApplicationCache.getApplicationCacheForFrame}
\alias{ApplicationCache.getApplicationCacheForFrame}
\title{Send the command ApplicationCache.getApplicationCacheForFrame}
\usage{
ApplicationCache.getApplicationCacheForFrame(promise, frameId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{frameId}{A Page.FrameId.
Identifier of the frame containing document whose application cache is retrieved.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns relevant application cache data for the document in given frame.
}
