% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationCache_commands.R
\name{ApplicationCache.getFramesWithManifests}
\alias{ApplicationCache.getFramesWithManifests}
\title{Send the command ApplicationCache.getFramesWithManifests

Returns array of frame identifiers with manifest urls for each frame containing a document
associated with some application cache.}
\usage{
ApplicationCache.getFramesWithManifests(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command ApplicationCache.getFramesWithManifests

Returns array of frame identifiers with manifest urls for each frame containing a document
associated with some application cache.
}
