% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationCache_commands.R
\name{ApplicationCache.getManifestForFrame}
\alias{ApplicationCache.getManifestForFrame}
\title{Send the command ApplicationCache.getManifestForFrame

Returns manifest URL for document in the given frame.}
\usage{
ApplicationCache.getManifestForFrame(promise, frameId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{frameId}{A Page.FrameId.
Identifier of the frame containing document whose manifest is retrieved.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command ApplicationCache.getManifestForFrame

Returns manifest URL for document in the given frame.
}
