% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_connect.R
\name{as.promise.devtoolscnx}
\alias{as.promise.devtoolscnx}
\title{Coerce to a promise}
\usage{
as.promise.devtoolscnx(x)
}
\arguments{
\item{obj}{A DevTools connexion.}
}
\value{
A promise.
}
\description{
Coerce to a promise
}
