% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevToolsConnexion.R
\name{as.promise.DevToolsConnexion}
\alias{as.promise.DevToolsConnexion}
\title{Coerce to a promise}
\usage{
\method{as.promise}{DevToolsConnexion}(x)
}
\arguments{
\item{obj}{A DevTools connexion.}
}
\value{
A promise.
}
\description{
Coerce to a promise
}
