% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevToolsConnexion.R
\name{as.promise}
\alias{as.promise}
\alias{as.promise.DevToolsConnexion}
\title{Coerce to a promise}
\usage{
\method{as.promise}{DevToolsConnexion}(x)
}
\description{
See \code{promises::\link[promises]{is.promise}} for details.
}
\keyword{internal}
