% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Audits_commands.R
\name{Audits.getEncodedResponse}
\alias{Audits.getEncodedResponse}
\title{Send the command Audits.getEncodedResponse

Returns the response body and size if it were re-encoded with the specified settings. Only
applies to images.}
\usage{
Audits.getEncodedResponse(promise, requestId, encoding, quality = NULL,
  sizeOnly = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{requestId}{A Network.RequestId.
Identifier of the network request to get content for.}

\item{encoding}{A character string.
The encoding to use. Accepted values: webp, jpeg, png.}

\item{quality}{Optional. A numeric.
The quality of the encoding (0-1). (defaults to 1)}

\item{sizeOnly}{Optional. A logical.
Whether to only return the size information (defaults to false).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 3.
}
\description{
Send the command Audits.getEncodedResponse

Returns the response body and size if it were re-encoded with the specified settings. Only
applies to images.
}
