% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/await.R
\name{await}
\alias{await}
\title{Await a promise is resolved}
\usage{
await(x)
}
\arguments{
\item{x}{A \code{\link[promises]{promise}} object.}
}
\value{
The value of the promise once resolved.
}
\description{
The \code{await} function is a helper to transform an asynchronous value to a
synchronous one. The R session returns once a \code{\link[promises]{promise}}
object is resolved or rejected. If the the promise is rejected an error is
thrown.
}
\details{
The \code{await} function cannot be used with Shiny.
}
\examples{
\dontrun{

library(promises)
library(crrri)
library(later)

pr <- promise(~ later(~ resolve("result of the async task"), 1))
value <- await(pr)
cat(value, "\\n")
}
}
