% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.close}
\alias{Browser.close}
\title{Send the command Browser.close}
\usage{
Browser.close(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Close browser gracefully.
}
