% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.crash}
\alias{Browser.crash}
\title{Send the command Browser.crash

Crashes browser on the main thread.}
\usage{
Browser.crash(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Browser.crash

Crashes browser on the main thread.
}
