% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.getHistogram}
\alias{Browser.getHistogram}
\title{Send the command Browser.getHistogram

Get a Chrome histogram by name.}
\usage{
Browser.getHistogram(promise, name, delta = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{name}{A character string.
Requested histogram name.}

\item{delta}{Optional. A logical.
If true, retrieve delta since last call.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Browser.getHistogram

Get a Chrome histogram by name.
}
