% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.getHistograms}
\alias{Browser.getHistograms}
\title{Send the command Browser.getHistograms

Get Chrome histograms.}
\usage{
Browser.getHistograms(promise, query = NULL, delta = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{query}{Optional. A character string.
Requested substring in name. Only histograms which have query as a
substring in their name are extracted. An empty or absent query returns
all histograms.}

\item{delta}{Optional. A logical.
If true, retrieve delta since last call.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Browser.getHistograms

Get Chrome histograms.
}
