% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.getVersion}
\alias{Browser.getVersion}
\title{Send the command Browser.getVersion}
\usage{
Browser.getVersion(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 5.
}
\description{
Returns version information.
}
