% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.getWindowBounds}
\alias{Browser.getWindowBounds}
\title{Send the command Browser.getWindowBounds

Get position and size of the browser window.}
\usage{
Browser.getWindowBounds(promise, windowId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{windowId}{A WindowID.
Browser window id.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Browser.getWindowBounds

Get position and size of the browser window.
}
