% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.grantPermissions}
\alias{Browser.grantPermissions}
\title{Send the command Browser.grantPermissions}
\usage{
Browser.grantPermissions(promise, origin, permissions,
  browserContextId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{origin}{A character string.}

\item{permissions}{A list of PermissionType.}

\item{browserContextId}{Optional. A Target.BrowserContextID.
BrowserContext to override permissions. When omitted, default browser context is used.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Grant specific permissions to the given origin and reject all others.
}
