% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.resetPermissions}
\alias{Browser.resetPermissions}
\title{Send the command Browser.resetPermissions}
\usage{
Browser.resetPermissions(promise, browserContextId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{browserContextId}{Optional. A Target.BrowserContextID.
BrowserContext to reset permissions. When omitted, default browser context is used.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Reset all permission management for all origins.
}
