% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.setDockTile}
\alias{Browser.setDockTile}
\title{Send the command Browser.setDockTile

Set dock tile details, platform-specific.}
\usage{
Browser.setDockTile(promise, badgeLabel = NULL, image = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{badgeLabel}{Optional. A character string.}

\item{image}{Optional. A character string.
Png encoded image.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Browser.setDockTile

Set dock tile details, platform-specific.
}
