% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.setWindowBounds}
\alias{Browser.setWindowBounds}
\title{Send the command Browser.setWindowBounds}
\usage{
Browser.setWindowBounds(promise, windowId, bounds)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{windowId}{A WindowID.
Browser window id.}

\item{bounds}{A Bounds.
New window bounds. The 'minimized', 'maximized' and 'fullscreen' states cannot be combined
with 'left', 'top', 'width' or 'height'. Leaves unspecified fields unchanged.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Set position and/or size of the browser window.
}
