% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheStorage_commands.R
\name{CacheStorage.deleteCache}
\alias{CacheStorage.deleteCache}
\title{Send the command CacheStorage.deleteCache}
\usage{
CacheStorage.deleteCache(promise, cacheId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{cacheId}{A CacheId.
Id of cache for deletion.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Deletes a cache.
}
