% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheStorage_commands.R
\name{CacheStorage.deleteEntry}
\alias{CacheStorage.deleteEntry}
\title{Send the command CacheStorage.deleteEntry

Deletes a cache entry.}
\usage{
CacheStorage.deleteEntry(promise, cacheId, request)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{cacheId}{A CacheId.
Id of cache where the entry will be deleted.}

\item{request}{A character string.
URL spec of the request.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command CacheStorage.deleteEntry

Deletes a cache entry.
}
