% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheStorage_commands.R
\name{CacheStorage.requestCachedResponse}
\alias{CacheStorage.requestCachedResponse}
\title{Send the command CacheStorage.requestCachedResponse}
\usage{
CacheStorage.requestCachedResponse(promise, cacheId, requestURL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{cacheId}{A CacheId.
Id of cache that contains the enty.}

\item{requestURL}{A character string.
URL spec of the request.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Fetches cache entry.
}
