% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheStorage_commands.R
\name{CacheStorage.requestCacheNames}
\alias{CacheStorage.requestCacheNames}
\title{Send the command CacheStorage.requestCacheNames

Requests cache names.}
\usage{
CacheStorage.requestCacheNames(promise, securityOrigin)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{securityOrigin}{A character string.
Security origin.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command CacheStorage.requestCacheNames

Requests cache names.
}
