% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheStorage_commands.R
\name{CacheStorage.requestEntries}
\alias{CacheStorage.requestEntries}
\title{Send the command CacheStorage.requestEntries

Requests data from cache.}
\usage{
CacheStorage.requestEntries(promise, cacheId, skipCount, pageSize)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{cacheId}{A CacheId.
ID of cache to get entries from.}

\item{skipCount}{An integer.
Number of records to skip.}

\item{pageSize}{An integer.
Number of records to fetch.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command CacheStorage.requestEntries

Requests data from cache.
}
