% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDPSession.R
\name{CDPSession}
\alias{CDPSession}
\title{Connect to a remote instance implementing the Chrome Debugging Protocol}
\usage{
CDPSession(host = "localhost", port = 9222, secure = FALSE,
  ws_url = NULL, local = FALSE, callback = NULL)
}
\arguments{
\item{host}{Character scalar, the host name of the application.}

\item{port}{Numeric scalar, the remote debugging port.}

\item{secure}{Logical scalar, indicating whether the https/wss protocols
shall be used for connecting to the remote application.}

\item{ws_url}{Character scalar, the websocket URL. If provided, \code{host} and
\code{port} and \code{secure} arguments are ignored.}

\item{local}{Logical scalar, indicating whether the local version of the
protocol (embedded in \code{crrri}) must be used or the protocol must be
fetched \emph{remotely}.}

\item{callback}{Function with one argument, executed when the R session is
connected to Chrome. The connection object is passed to this function.}
}
\value{
A connection object of class \code{CDPSession}.
}
\description{
This function creates a websocket connection to a remote instance using
the Chrome Debugging Protocol.
}
