% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_connect.R
\name{chr_connect}
\alias{chr_connect}
\title{Connect R to a new Chrome instance}
\usage{
chr_connect(bin = Sys.getenv("HEADLESS_CHROME"), debug_port = 9222,
  extra_args = NULL, headless = TRUE, auto_connect = TRUE,
  async = TRUE)
}
\arguments{
\item{bin}{Command or path to the Chrome exe.}

\item{debug_port}{Port of the Chrome instance.}

\item{extra_args}{Extra arguments passed to command line.}

\item{headless}{Is Chrome launch in headless mode?}

\item{auto_connect}{Is the connexion to Chrome launched immediately?}

\item{async}{Is the returned object a promise?}
}
\value{
A connexion object that is a named list of two elements. The object
has a \code{result} element that is \code{NULL} for a new connexion and a \code{ws}
element that is a WebSocket object.
}
\description{
Connect R to a new Chrome instance
}
