% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_connect.R
\name{chr_disconnect}
\alias{chr_disconnect}
\title{Disconnect R from a Chrome instance and close it}
\usage{
chr_disconnect(promise)
}
\arguments{
\item{promise}{A previous asynchronous result.}
}
\value{
\code{TRUE} invisibly.
}
\description{
Disconnect R from a Chrome instance and close it
}
