% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Console_commands.R
\name{Console.clearMessages}
\alias{Console.clearMessages}
\title{Send the command Console.clearMessages

Does nothing.}
\usage{
Console.clearMessages(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Console.clearMessages

Does nothing.
}
