% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Console_commands.R
\name{Console.enable}
\alias{Console.enable}
\title{Send the command Console.enable}
\usage{
Console.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables console domain, sends the messages collected so far to the client by means of the
\code{messageAdded} notification.
}
