% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.addRule}
\alias{CSS.addRule}
\title{Send the command CSS.addRule}
\usage{
CSS.addRule(promise, styleSheetId, ruleText, location)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{styleSheetId}{A StyleSheetId.
The css style sheet identifier where a new rule should be inserted.}

\item{ruleText}{A character string.
The text of a new rule.}

\item{location}{A SourceRange.
Text position of a new rule in the target style sheet.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Inserts a new rule with the given \code{ruleText} in a stylesheet with given \code{styleSheetId}, at the
position specified by \code{location}.
}
