% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.collectClassNames}
\alias{CSS.collectClassNames}
\title{Send the command CSS.collectClassNames

Returns all class names from specified stylesheet.}
\usage{
CSS.collectClassNames(promise, styleSheetId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{styleSheetId}{A StyleSheetId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command CSS.collectClassNames

Returns all class names from specified stylesheet.
}
