% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.createStyleSheet}
\alias{CSS.createStyleSheet}
\title{Send the command CSS.createStyleSheet}
\usage{
CSS.createStyleSheet(promise, frameId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{frameId}{A Page.FrameId.
Identifier of the frame where "via-inspector" stylesheet should be created.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Creates a new special "via-inspector" stylesheet in the frame with given \code{frameId}.
}
