% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.disable}
\alias{CSS.disable}
\title{Send the command CSS.disable

Disables the CSS agent for the given page.}
\usage{
CSS.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command CSS.disable

Disables the CSS agent for the given page.
}
