% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.enable}
\alias{CSS.enable}
\title{Send the command CSS.enable

Enables the CSS agent for the given page. Clients should not assume that the CSS agent has been
enabled until the result of this command is received.}
\usage{
CSS.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command CSS.enable

Enables the CSS agent for the given page. Clients should not assume that the CSS agent has been
enabled until the result of this command is received.
}
