% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.forcePseudoState}
\alias{CSS.forcePseudoState}
\title{Send the command CSS.forcePseudoState}
\usage{
CSS.forcePseudoState(promise, nodeId, forcedPseudoClasses)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A DOM.NodeId.
The element id for which to force the pseudo state.}

\item{forcedPseudoClasses}{A list of string.
Element pseudo classes to force when computing the element's style.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Ensures that the given node will have specified pseudo-classes whenever its style is computed by
the browser.
}
