% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.getBackgroundColors}
\alias{CSS.getBackgroundColors}
\title{Send the command CSS.getBackgroundColors}
\usage{
CSS.getBackgroundColors(promise, nodeId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A DOM.NodeId.
Id of the node to get background colors for.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 4.
}
\description{
Send the command CSS.getBackgroundColors
}
