% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.getComputedStyleForNode}
\alias{CSS.getComputedStyleForNode}
\title{Send the command CSS.getComputedStyleForNode

Returns the computed style for a DOM node identified by \code{nodeId}.}
\usage{
CSS.getComputedStyleForNode(promise, nodeId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A DOM.NodeId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command CSS.getComputedStyleForNode

Returns the computed style for a DOM node identified by \code{nodeId}.
}
