% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.getInlineStylesForNode}
\alias{CSS.getInlineStylesForNode}
\title{Send the command CSS.getInlineStylesForNode}
\usage{
CSS.getInlineStylesForNode(promise, nodeId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A DOM.NodeId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Returns the styles defined inline (explicitly in the "style" attribute and implicitly, using DOM
attributes) for a DOM node identified by \code{nodeId}.
}
