% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.getMatchedStylesForNode}
\alias{CSS.getMatchedStylesForNode}
\title{Send the command CSS.getMatchedStylesForNode}
\usage{
CSS.getMatchedStylesForNode(promise, nodeId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A DOM.NodeId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 6.
}
\description{
Returns requested styles for a DOM node identified by \code{nodeId}.
}
