% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.getMediaQueries}
\alias{CSS.getMediaQueries}
\title{Send the command CSS.getMediaQueries

Returns all media queries parsed by the rendering engine.}
\usage{
CSS.getMediaQueries(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command CSS.getMediaQueries

Returns all media queries parsed by the rendering engine.
}
