% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.getPlatformFontsForNode}
\alias{CSS.getPlatformFontsForNode}
\title{Send the command CSS.getPlatformFontsForNode}
\usage{
CSS.getPlatformFontsForNode(promise, nodeId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A DOM.NodeId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Requests information about platform fonts which we used to render child TextNodes in the given
node.
}
