% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.getStyleSheetText}
\alias{CSS.getStyleSheetText}
\title{Send the command CSS.getStyleSheetText

Returns the current textual content for a stylesheet.}
\usage{
CSS.getStyleSheetText(promise, styleSheetId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{styleSheetId}{A StyleSheetId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command CSS.getStyleSheetText

Returns the current textual content for a stylesheet.
}
