% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.setEffectivePropertyValueForNode}
\alias{CSS.setEffectivePropertyValueForNode}
\title{Send the command CSS.setEffectivePropertyValueForNode}
\usage{
CSS.setEffectivePropertyValueForNode(promise, nodeId, propertyName, value)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A DOM.NodeId.
The element id for which to set property.}

\item{propertyName}{A character string.}

\item{value}{A character string.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Find a rule with the given active property for the given node and set the new value for this
property
}
