% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.setKeyframeKey}
\alias{CSS.setKeyframeKey}
\title{Send the command CSS.setKeyframeKey}
\usage{
CSS.setKeyframeKey(promise, styleSheetId, range, keyText)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{styleSheetId}{A StyleSheetId.}

\item{range}{A SourceRange.}

\item{keyText}{A character string.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Modifies the keyframe rule key text.
}
