% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.setMediaText}
\alias{CSS.setMediaText}
\title{Send the command CSS.setMediaText}
\usage{
CSS.setMediaText(promise, styleSheetId, range, text)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{styleSheetId}{A StyleSheetId.}

\item{range}{A SourceRange.}

\item{text}{A character string.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Modifies the rule selector.
}
