% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.setStyleSheetText}
\alias{CSS.setStyleSheetText}
\title{Send the command CSS.setStyleSheetText}
\usage{
CSS.setStyleSheetText(promise, styleSheetId, text)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{styleSheetId}{A StyleSheetId.}

\item{text}{A character string.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Sets the new stylesheet text.
}
