% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.setStyleTexts}
\alias{CSS.setStyleTexts}
\title{Send the command CSS.setStyleTexts}
\usage{
CSS.setStyleTexts(promise, edits)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{edits}{A list of StyleDeclarationEdit.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Applies specified style edits one after another in the given order.
}
