% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.startRuleUsageTracking}
\alias{CSS.startRuleUsageTracking}
\title{Send the command CSS.startRuleUsageTracking

Enables the selector recording.}
\usage{
CSS.startRuleUsageTracking(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command CSS.startRuleUsageTracking

Enables the selector recording.
}
