% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.stopRuleUsageTracking}
\alias{CSS.stopRuleUsageTracking}
\title{Send the command CSS.stopRuleUsageTracking}
\usage{
CSS.stopRuleUsageTracking(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Stop tracking rule usage and return the list of rules that were used since last call to
\code{takeCoverageDelta} (or since start of coverage instrumentation)
}
