% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_commands.R
\name{CSS.takeCoverageDelta}
\alias{CSS.takeCoverageDelta}
\title{Send the command CSS.takeCoverageDelta}
\usage{
CSS.takeCoverageDelta(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Obtain list of rules that became used since last call to this method (or since start of coverage
instrumentation)
}
