% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_commands.R
\name{Database.disable}
\alias{Database.disable}
\title{Send the command Database.disable

Disables database tracking, prevents database events from being sent to the client.}
\usage{
Database.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Database.disable

Disables database tracking, prevents database events from being sent to the client.
}
