% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_commands.R
\name{Database.enable}
\alias{Database.enable}
\title{Send the command Database.enable}
\usage{
Database.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables database tracking, database events will now be delivered to the client.
}
