% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_commands.R
\name{Database.executeSQL}
\alias{Database.executeSQL}
\title{Send the command Database.executeSQL}
\usage{
Database.executeSQL(promise, databaseId, query)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{databaseId}{A DatabaseId.}

\item{query}{A character string.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 3.
}
\description{
Send the command Database.executeSQL
}
