% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_commands.R
\name{Database.getDatabaseTableNames}
\alias{Database.getDatabaseTableNames}
\title{Send the command Database.getDatabaseTableNames}
\usage{
Database.getDatabaseTableNames(promise, databaseId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{databaseId}{A DatabaseId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Database.getDatabaseTableNames
}
