% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.enable}
\alias{Debugger.enable}
\title{Send the command Debugger.enable

Enables debugger for the given page. Clients should not assume that the debugging has been
enabled until the result for this command is received.}
\usage{
Debugger.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Debugger.enable

Enables debugger for the given page. Clients should not assume that the debugging has been
enabled until the result for this command is received.
}
