% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.getScriptSource}
\alias{Debugger.getScriptSource}
\title{Send the command Debugger.getScriptSource

Returns source for the script with given id.}
\usage{
Debugger.getScriptSource(promise, scriptId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{scriptId}{A Runtime.ScriptId.
Id of the script to get source for.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Debugger.getScriptSource

Returns source for the script with given id.
}
