% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.pauseOnAsyncCall}
\alias{Debugger.pauseOnAsyncCall}
\title{Send the command Debugger.pauseOnAsyncCall}
\usage{
Debugger.pauseOnAsyncCall(promise, parentStackTraceId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{parentStackTraceId}{A Runtime.StackTraceId.
Debugger will pause when async call with given stack trace is started.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Debugger.pauseOnAsyncCall
}
