% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.removeBreakpoint}
\alias{Debugger.removeBreakpoint}
\title{Send the command Debugger.removeBreakpoint

Removes JavaScript breakpoint.}
\usage{
Debugger.removeBreakpoint(promise, breakpointId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{breakpointId}{A BreakpointId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Debugger.removeBreakpoint

Removes JavaScript breakpoint.
}
