% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.restartFrame}
\alias{Debugger.restartFrame}
\title{Send the command Debugger.restartFrame}
\usage{
Debugger.restartFrame(promise, callFrameId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{callFrameId}{A CallFrameId.
Call frame identifier to evaluate on.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 3.
}
\description{
Restarts particular call frame from the beginning.
}
