% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.resume}
\alias{Debugger.resume}
\title{Send the command Debugger.resume}
\usage{
Debugger.resume(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Resumes JavaScript execution.
}
