% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.searchInContent}
\alias{Debugger.searchInContent}
\title{Send the command Debugger.searchInContent}
\usage{
Debugger.searchInContent(promise, scriptId, query, caseSensitive = NULL,
  isRegex = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{scriptId}{A Runtime.ScriptId.
Id of the script to search in.}

\item{query}{A character string.
String to search for.}

\item{caseSensitive}{Optional. A logical.
If true, search is case sensitive.}

\item{isRegex}{Optional. A logical.
If true, treats string parameter as regex.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Searches for given string in script content.
}
