% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.setAsyncCallStackDepth}
\alias{Debugger.setAsyncCallStackDepth}
\title{Send the command Debugger.setAsyncCallStackDepth}
\usage{
Debugger.setAsyncCallStackDepth(promise, maxDepth)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{maxDepth}{An integer.
Maximum depth of async call stacks. Setting to \code{0} will effectively disable collecting async
call stacks (default).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables or disables async call stacks tracking.
}
