% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.setBreakpoint}
\alias{Debugger.setBreakpoint}
\title{Send the command Debugger.setBreakpoint}
\usage{
Debugger.setBreakpoint(promise, location, condition = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{location}{A Location.
Location to set breakpoint in.}

\item{condition}{Optional. A character string.
Expression to use as a breakpoint condition. When specified, debugger will only stop on the
breakpoint if this expression evaluates to true.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Sets JavaScript breakpoint at a given location.
}
