% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.setBreakpointByUrl}
\alias{Debugger.setBreakpointByUrl}
\title{Send the command Debugger.setBreakpointByUrl}
\usage{
Debugger.setBreakpointByUrl(promise, lineNumber, url = NULL,
  urlRegex = NULL, scriptHash = NULL, columnNumber = NULL,
  condition = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{lineNumber}{An integer.
Line number to set breakpoint at.}

\item{url}{Optional. A character string.
URL of the resources to set breakpoint on.}

\item{urlRegex}{Optional. A character string.
Regex pattern for the URLs of the resources to set breakpoints on. Either \code{url} or
\code{urlRegex} must be specified.}

\item{scriptHash}{Optional. A character string.
Script hash of the resources to set breakpoint on.}

\item{columnNumber}{Optional. An integer.
Offset in the line to set breakpoint at.}

\item{condition}{Optional. A character string.
Expression to use as a breakpoint condition. When specified, debugger will only stop on the
breakpoint if this expression evaluates to true.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Sets JavaScript breakpoint at given location specified either by URL or URL regex. Once this
command is issued, all existing parsed scripts will have breakpoints resolved and returned in
\code{locations} property. Further matching script parsing will result in subsequent
\code{breakpointResolved} events issued. This logical breakpoint will survive page reloads.
}
