% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Debugger.R
\name{Debugger.setBreakpointOnFunctionCall}
\alias{Debugger.setBreakpointOnFunctionCall}
\title{Send the command Debugger.setBreakpointOnFunctionCall}
\usage{
Debugger.setBreakpointOnFunctionCall(promise, objectId, condition = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{objectId}{A Runtime.RemoteObjectId.
Function object id.}

\item{condition}{Optional. A character string.
Expression to use as a breakpoint condition. When specified, debugger will
stop on the breakpoint if this expression evaluates to true.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Sets JavaScript breakpoint before each call to the given function.
If another function was created from the same source as a given one,
calling it will also trigger the breakpoint.
}
