% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.setBreakpointsActive}
\alias{Debugger.setBreakpointsActive}
\title{Send the command Debugger.setBreakpointsActive}
\usage{
Debugger.setBreakpointsActive(promise, active)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{active}{A logical.
New value for breakpoints active state.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Activates / deactivates all breakpoints on the page.
}
