% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.setPauseOnExceptions}
\alias{Debugger.setPauseOnExceptions}
\title{Send the command Debugger.setPauseOnExceptions}
\usage{
Debugger.setPauseOnExceptions(promise, state)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{state}{A character string.
Pause on exceptions mode. Accepted values: none, uncaught, all.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Defines pause on exceptions state. Can be set to stop on all exceptions, uncaught exceptions or
no exceptions. Initial pause on exceptions state is \code{none}.
}
