% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.setReturnValue}
\alias{Debugger.setReturnValue}
\title{Send the command Debugger.setReturnValue

Changes return value in top frame. Available only at return break position.}
\usage{
Debugger.setReturnValue(promise, newValue)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{newValue}{A Runtime.CallArgument.
New return value.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Debugger.setReturnValue

Changes return value in top frame. Available only at return break position.
}
