% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.setScriptSource}
\alias{Debugger.setScriptSource}
\title{Send the command Debugger.setScriptSource}
\usage{
Debugger.setScriptSource(promise, scriptId, scriptSource, dryRun = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{scriptId}{A Runtime.ScriptId.
Id of the script to edit.}

\item{scriptSource}{A character string.
New content of the script.}

\item{dryRun}{Optional. A logical.
If true the change will not actually be applied. Dry run may be used to get result
description without actually modifying the code.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 5.
}
\description{
Edits JavaScript source live.
}
