% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.setSkipAllPauses}
\alias{Debugger.setSkipAllPauses}
\title{Send the command Debugger.setSkipAllPauses

Makes page not interrupt on any pauses (breakpoint, exception, dom exception etc).}
\usage{
Debugger.setSkipAllPauses(promise, skip)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{skip}{A logical.
New value for skip pauses state.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Debugger.setSkipAllPauses

Makes page not interrupt on any pauses (breakpoint, exception, dom exception etc).
}
