% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.stepInto}
\alias{Debugger.stepInto}
\title{Send the command Debugger.stepInto}
\usage{
Debugger.stepInto(promise, breakOnAsyncCall = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{breakOnAsyncCall}{Experimental. Optional. A logical.
Debugger will issue additional Debugger.paused notification if any async task is scheduled
before next pause.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Steps into the function call.
}
