% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.stepOut}
\alias{Debugger.stepOut}
\title{Send the command Debugger.stepOut

Steps out of the function call.}
\usage{
Debugger.stepOut(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Debugger.stepOut

Steps out of the function call.
}
