% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.stepOver}
\alias{Debugger.stepOver}
\title{Send the command Debugger.stepOver

Steps over the statement.}
\usage{
Debugger.stepOver(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Debugger.stepOver

Steps over the statement.
}
