% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeviceOrientation_commands.R
\name{DeviceOrientation.clearDeviceOrientationOverride}
\alias{DeviceOrientation.clearDeviceOrientationOverride}
\title{Send the command DeviceOrientation.clearDeviceOrientationOverride

Clears the overridden Device Orientation.}
\usage{
DeviceOrientation.clearDeviceOrientationOverride(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DeviceOrientation.clearDeviceOrientationOverride

Clears the overridden Device Orientation.
}
