% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevToolsConnexion.R
\name{DevToolsConnexion}
\alias{DevToolsConnexion}
\title{Create a websocket connexion using the Chrome DevTools Protocol}
\description{
Create a websocket connexion using the Chrome DevTools Protocol
}
