% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.collectClassNamesFromSubtree}
\alias{DOM.collectClassNamesFromSubtree}
\title{Send the command DOM.collectClassNamesFromSubtree

Collects class names for the node with given id and all of it's child nodes.}
\usage{
DOM.collectClassNamesFromSubtree(promise, nodeId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the node to collect class names.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command DOM.collectClassNamesFromSubtree

Collects class names for the node with given id and all of it's child nodes.
}
