% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.copyTo}
\alias{DOM.copyTo}
\title{Send the command DOM.copyTo

Creates a deep copy of the specified node and places it into the target container before the
given anchor.}
\usage{
DOM.copyTo(promise, nodeId, targetNodeId, insertBeforeNodeId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the node to copy.}

\item{targetNodeId}{A NodeId.
Id of the element to drop the copy into.}

\item{insertBeforeNodeId}{Optional. A NodeId.
Drop the copy before this node (if absent, the copy becomes the last child of
\code{targetNodeId}).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command DOM.copyTo

Creates a deep copy of the specified node and places it into the target container before the
given anchor.
}
