% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.describeNode}
\alias{DOM.describeNode}
\title{Send the command DOM.describeNode}
\usage{
DOM.describeNode(promise, nodeId = NULL, backendNodeId = NULL,
  objectId = NULL, depth = NULL, pierce = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{Optional. A NodeId.
Identifier of the node.}

\item{backendNodeId}{Optional. A BackendNodeId.
Identifier of the backend node.}

\item{objectId}{Optional. A Runtime.RemoteObjectId.
JavaScript object id of the node wrapper.}

\item{depth}{Optional. An integer.
The maximum depth at which children should be retrieved, defaults to 1. Use -1 for the
entire subtree or provide an integer larger than 0.}

\item{pierce}{Optional. A logical.
Whether or not iframes and shadow roots should be traversed when returning the subtree
(default is false).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Describes node given its id, does not require domain to be enabled. Does not start tracking any
objects, can be used for automation.
}
