% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.discardSearchResults}
\alias{DOM.discardSearchResults}
\title{Send the command DOM.discardSearchResults}
\usage{
DOM.discardSearchResults(promise, searchId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{searchId}{A character string.
Unique search session identifier.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Discards search results from the session with the given id. \code{getSearchResults} should no longer
be called for that search.
}
