% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.enable}
\alias{DOM.enable}
\title{Send the command DOM.enable

Enables DOM agent for the given page.}
\usage{
DOM.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOM.enable

Enables DOM agent for the given page.
}
