% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.getAttributes}
\alias{DOM.getAttributes}
\title{Send the command DOM.getAttributes}
\usage{
DOM.getAttributes(promise, nodeId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the node to retrieve attibutes for.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns attributes for the specified node.
}
