% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.getDocument}
\alias{DOM.getDocument}
\title{Send the command DOM.getDocument

Returns the root DOM node (and optionally the subtree) to the caller.}
\usage{
DOM.getDocument(promise, depth = NULL, pierce = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{depth}{Optional. An integer.
The maximum depth at which children should be retrieved, defaults to 1. Use -1 for the
entire subtree or provide an integer larger than 0.}

\item{pierce}{Optional. A logical.
Whether or not iframes and shadow roots should be traversed when returning the subtree
(default is false).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command DOM.getDocument

Returns the root DOM node (and optionally the subtree) to the caller.
}
