% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.getFrameOwner}
\alias{DOM.getFrameOwner}
\title{Send the command DOM.getFrameOwner}
\usage{
DOM.getFrameOwner(promise, frameId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{frameId}{A Page.FrameId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Returns iframe node that owns iframe with the given domain.
}
