% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.getNodeForLocation}
\alias{DOM.getNodeForLocation}
\title{Send the command DOM.getNodeForLocation}
\usage{
DOM.getNodeForLocation(promise, x, y, includeUserAgentShadowDOM = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{x}{An integer.
X coordinate.}

\item{y}{An integer.
Y coordinate.}

\item{includeUserAgentShadowDOM}{Optional. A logical.
False to skip to the nearest non-UA shadow root ancestor (default: false).}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 2.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Returns node id at given location. Depending on whether DOM domain is enabled, nodeId is
either returned or not.
}
