% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.getRelayoutBoundary}
\alias{DOM.getRelayoutBoundary}
\title{Send the command DOM.getRelayoutBoundary}
\usage{
DOM.getRelayoutBoundary(promise, nodeId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the node.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns the id of the nearest ancestor that is a relayout boundary.
}
