% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.getSearchResults}
\alias{DOM.getSearchResults}
\title{Send the command DOM.getSearchResults

Returns search results from given \code{fromIndex} to given \code{toIndex} from the search with the given
identifier.}
\usage{
DOM.getSearchResults(promise, searchId, fromIndex, toIndex)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{searchId}{A character string.
Unique search session identifier.}

\item{fromIndex}{An integer.
Start index of the search result to be returned.}

\item{toIndex}{An integer.
End index of the search result to be returned.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command DOM.getSearchResults

Returns search results from given \code{fromIndex} to given \code{toIndex} from the search with the given
identifier.
}
