% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.highlightRect}
\alias{DOM.highlightRect}
\title{Send the command DOM.highlightRect

Highlights given rectangle.}
\usage{
DOM.highlightRect(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOM.highlightRect

Highlights given rectangle.
}
