% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.moveTo}
\alias{DOM.moveTo}
\title{Send the command DOM.moveTo

Moves node into the new container, places it before the given anchor.}
\usage{
DOM.moveTo(promise, nodeId, targetNodeId, insertBeforeNodeId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the node to move.}

\item{targetNodeId}{A NodeId.
Id of the element to drop the moved node into.}

\item{insertBeforeNodeId}{Optional. A NodeId.
Drop node before this one (if absent, the moved node becomes the last child of
\code{targetNodeId}).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command DOM.moveTo

Moves node into the new container, places it before the given anchor.
}
