% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.performSearch}
\alias{DOM.performSearch}
\title{Send the command DOM.performSearch}
\usage{
DOM.performSearch(promise, query, includeUserAgentShadowDOM = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{query}{A character string.
Plain text or query selector or XPath search query.}

\item{includeUserAgentShadowDOM}{Optional. A logical.
True to search in user agent shadow DOM.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 2.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Searches for a given string in the DOM tree. Use \code{getSearchResults} to access search results or
\code{cancelSearch} to end this search session.
}
