% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.pushNodeByPathToFrontend}
\alias{DOM.pushNodeByPathToFrontend}
\title{Send the command DOM.pushNodeByPathToFrontend}
\usage{
DOM.pushNodeByPathToFrontend(promise, path)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{path}{A character string.
Path to node in the proprietary format.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Requests that the node is sent to the caller given its path. // FIXME, use XPath
}
