% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.querySelector}
\alias{DOM.querySelector}
\title{Send the command DOM.querySelector}
\usage{
DOM.querySelector(promise, nodeId, selector)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the node to query upon.}

\item{selector}{A character string.
Selector string.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Executes \code{querySelector} on a given node.
}
