% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.redo}
\alias{DOM.redo}
\title{Send the command DOM.redo

Re-does the last undone action.}
\usage{
DOM.redo(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOM.redo

Re-does the last undone action.
}
