% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.removeAttribute}
\alias{DOM.removeAttribute}
\title{Send the command DOM.removeAttribute

Removes attribute with given name from an element with given id.}
\usage{
DOM.removeAttribute(promise, nodeId, name)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the element to remove attribute from.}

\item{name}{A character string.
Name of the attribute to remove.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOM.removeAttribute

Removes attribute with given name from an element with given id.
}
