% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.resolveNode}
\alias{DOM.resolveNode}
\title{Send the command DOM.resolveNode}
\usage{
DOM.resolveNode(promise, nodeId = NULL, backendNodeId = NULL,
  objectGroup = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{Optional. A NodeId.
Id of the node to resolve.}

\item{backendNodeId}{Optional. A DOM.BackendNodeId.
Backend identifier of the node to resolve.}

\item{objectGroup}{Optional. A character string.
Symbolic group name that can be used to release multiple objects.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Resolves the JavaScript node object for a given NodeId or BackendNodeId.
}
