% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.setAttributesAsText}
\alias{DOM.setAttributesAsText}
\title{Send the command DOM.setAttributesAsText}
\usage{
DOM.setAttributesAsText(promise, nodeId, text, name = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{nodeId}{A NodeId.
Id of the element to set attributes for.}

\item{text}{A character string.
Text with a number of attributes. Will parse this text using HTML parser.}

\item{name}{Optional. A character string.
Attribute name to replace with new attributes derived from text in case text parsed
successfully.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Sets attributes on element with given id. This method is useful when user edits some existing
attribute value and types in several attribute name/value pairs.
}
