% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.setAttributeValue}
\alias{DOM.setAttributeValue}
\title{Send the command DOM.setAttributeValue}
\usage{
DOM.setAttributeValue(promise, nodeId, name, value)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the element to set attribute for.}

\item{name}{A character string.
Attribute name.}

\item{value}{A character string.
Attribute value.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sets attribute for an element with given id.
}
