% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.setInspectedNode}
\alias{DOM.setInspectedNode}
\title{Send the command DOM.setInspectedNode}
\usage{
DOM.setInspectedNode(promise, nodeId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
DOM node id to be accessible by means of $x command line API.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables console to refer to the node with given id via $x (see Command Line API for more details
$x functions).
}
