% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.setNodeName}
\alias{DOM.setNodeName}
\title{Send the command DOM.setNodeName}
\usage{
DOM.setNodeName(promise, nodeId, name)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the node to set name for.}

\item{name}{A character string.
New node's name.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Sets node name for a node with given id.
}
