% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.setNodeValue}
\alias{DOM.setNodeValue}
\title{Send the command DOM.setNodeValue}
\usage{
DOM.setNodeValue(promise, nodeId, value)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the node to set value for.}

\item{value}{A character string.
New node's value.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sets node value for a node with given id.
}
