% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.setOuterHTML}
\alias{DOM.setOuterHTML}
\title{Send the command DOM.setOuterHTML

Sets node HTML markup, returns new node id.}
\usage{
DOM.setOuterHTML(promise, nodeId, outerHTML)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the node to set markup for.}

\item{outerHTML}{A character string.
Outer HTML markup to set.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOM.setOuterHTML

Sets node HTML markup, returns new node id.
}
