% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.undo}
\alias{DOM.undo}
\title{Send the command DOM.undo

Undoes the last performed action.}
\usage{
DOM.undo(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOM.undo

Undoes the last performed action.
}
