% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMDebugger_commands.R
\name{DOMDebugger.getEventListeners}
\alias{DOMDebugger.getEventListeners}
\title{Send the command DOMDebugger.getEventListeners

Returns event listeners of the given object.}
\usage{
DOMDebugger.getEventListeners(promise, objectId, depth = NULL,
  pierce = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{objectId}{A Runtime.RemoteObjectId.
Identifier of the object to return listeners for.}

\item{depth}{Optional. An integer.
The maximum depth at which Node children should be retrieved, defaults to 1. Use -1 for the
entire subtree or provide an integer larger than 0.}

\item{pierce}{Optional. A logical.
Whether or not iframes and shadow roots should be traversed when returning the subtree
(default is false). Reports listeners for all contexts if pierce is enabled.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command DOMDebugger.getEventListeners

Returns event listeners of the given object.
}
