% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMDebugger_commands.R
\name{DOMDebugger.removeDOMBreakpoint}
\alias{DOMDebugger.removeDOMBreakpoint}
\title{Send the command DOMDebugger.removeDOMBreakpoint}
\usage{
DOMDebugger.removeDOMBreakpoint(promise, nodeId, type)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A DOM.NodeId.
Identifier of the node to remove breakpoint from.}

\item{type}{A DOMBreakpointType.
Type of the breakpoint to remove.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Removes DOM breakpoint that was set using \code{setDOMBreakpoint}.
}
