% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMDebugger_commands.R
\name{DOMDebugger.removeEventListenerBreakpoint}
\alias{DOMDebugger.removeEventListenerBreakpoint}
\title{Send the command DOMDebugger.removeEventListenerBreakpoint

Removes breakpoint on particular DOM event.}
\usage{
DOMDebugger.removeEventListenerBreakpoint(promise, eventName,
  targetName = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{eventName}{A character string.
Event name.}

\item{targetName}{Experimental. Optional. A character string.
EventTarget interface name.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOMDebugger.removeEventListenerBreakpoint

Removes breakpoint on particular DOM event.
}
