% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMDebugger_commands.R
\name{DOMDebugger.removeXHRBreakpoint}
\alias{DOMDebugger.removeXHRBreakpoint}
\title{Send the command DOMDebugger.removeXHRBreakpoint}
\usage{
DOMDebugger.removeXHRBreakpoint(promise, url)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{url}{A character string.
Resource URL substring.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Removes breakpoint from XMLHttpRequest.
}
