% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMDebugger_commands.R
\name{DOMDebugger.setDOMBreakpoint}
\alias{DOMDebugger.setDOMBreakpoint}
\title{Send the command DOMDebugger.setDOMBreakpoint}
\usage{
DOMDebugger.setDOMBreakpoint(promise, nodeId, type)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A DOM.NodeId.
Identifier of the node to set breakpoint on.}

\item{type}{A DOMBreakpointType.
Type of the operation to stop upon.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sets breakpoint on particular operation with DOM.
}
