% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMDebugger_commands.R
\name{DOMDebugger.setEventListenerBreakpoint}
\alias{DOMDebugger.setEventListenerBreakpoint}
\title{Send the command DOMDebugger.setEventListenerBreakpoint

Sets breakpoint on particular DOM event.}
\usage{
DOMDebugger.setEventListenerBreakpoint(promise, eventName,
  targetName = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{eventName}{A character string.
DOM Event name to stop on (any DOM event will do).}

\item{targetName}{Experimental. Optional. A character string.
EventTarget interface name to stop on. If equal to \code{"*"} or not provided, will stop on any
EventTarget.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOMDebugger.setEventListenerBreakpoint

Sets breakpoint on particular DOM event.
}
