% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMDebugger_commands.R
\name{DOMDebugger.setInstrumentationBreakpoint}
\alias{DOMDebugger.setInstrumentationBreakpoint}
\title{Send the command DOMDebugger.setInstrumentationBreakpoint

Sets breakpoint on particular native event.}
\usage{
DOMDebugger.setInstrumentationBreakpoint(promise, eventName)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{eventName}{A character string.
Instrumentation name to stop on.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOMDebugger.setInstrumentationBreakpoint

Sets breakpoint on particular native event.
}
