% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMDebugger_commands.R
\name{DOMDebugger.setXHRBreakpoint}
\alias{DOMDebugger.setXHRBreakpoint}
\title{Send the command DOMDebugger.setXHRBreakpoint}
\usage{
DOMDebugger.setXHRBreakpoint(promise, url)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{url}{A character string.
Resource URL substring. All XHRs having this substring in the URL will get stopped upon.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sets breakpoint on XMLHttpRequest.
}
