% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMSnapshot_commands.R
\name{DOMSnapshot.captureSnapshot}
\alias{DOMSnapshot.captureSnapshot}
\title{Send the command DOMSnapshot.captureSnapshot

Returns a document snapshot, including the full DOM tree of the root node (including iframes,
template contents, and imported documents) in a flattened array, as well as layout and
white-listed computed style information for the nodes. Shadow DOM in the returned DOM tree is
flattened.}
\usage{
DOMSnapshot.captureSnapshot(promise, computedStyles)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{computedStyles}{A list of string.
Whitelist of computed styles to return.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command DOMSnapshot.captureSnapshot

Returns a document snapshot, including the full DOM tree of the root node (including iframes,
template contents, and imported documents) in a flattened array, as well as layout and
white-listed computed style information for the nodes. Shadow DOM in the returned DOM tree is
flattened.
}
