% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMSnapshot_commands.R
\name{DOMSnapshot.disable}
\alias{DOMSnapshot.disable}
\title{Send the command DOMSnapshot.disable}
\usage{
DOMSnapshot.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Disables DOM snapshot agent for the given page.
}
