% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMStorage_commands.R
\name{DOMStorage.clear}
\alias{DOMStorage.clear}
\title{Send the command DOMStorage.clear}
\usage{
DOMStorage.clear(promise, storageId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{storageId}{A StorageId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOMStorage.clear
}
