% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMStorage_commands.R
\name{DOMStorage.disable}
\alias{DOMStorage.disable}
\title{Send the command DOMStorage.disable}
\usage{
DOMStorage.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Disables storage tracking, prevents storage events from being sent to the client.
}
