% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMStorage_commands.R
\name{DOMStorage.enable}
\alias{DOMStorage.enable}
\title{Send the command DOMStorage.enable}
\usage{
DOMStorage.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables storage tracking, storage events will now be delivered to the client.
}
