% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMStorage_commands.R
\name{DOMStorage.removeDOMStorageItem}
\alias{DOMStorage.removeDOMStorageItem}
\title{Send the command DOMStorage.removeDOMStorageItem}
\usage{
DOMStorage.removeDOMStorageItem(promise, storageId, key)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{storageId}{A StorageId.}

\item{key}{A character string.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOMStorage.removeDOMStorageItem
}
