% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMStorage_commands.R
\name{DOMStorage.setDOMStorageItem}
\alias{DOMStorage.setDOMStorageItem}
\title{Send the command DOMStorage.setDOMStorageItem}
\usage{
DOMStorage.setDOMStorageItem(promise, storageId, key, value)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{storageId}{A StorageId.}

\item{key}{A character string.}

\item{value}{A character string.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOMStorage.setDOMStorageItem
}
