% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.canEmulate}
\alias{Emulation.canEmulate}
\title{Send the command Emulation.canEmulate}
\usage{
Emulation.canEmulate(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Tells whether emulation is supported.
}
