% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.clearGeolocationOverride}
\alias{Emulation.clearGeolocationOverride}
\title{Send the command Emulation.clearGeolocationOverride

Clears the overriden Geolocation Position and Error.}
\usage{
Emulation.clearGeolocationOverride(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Emulation.clearGeolocationOverride

Clears the overriden Geolocation Position and Error.
}
