% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setCPUThrottlingRate}
\alias{Emulation.setCPUThrottlingRate}
\title{Send the command Emulation.setCPUThrottlingRate

Enables CPU throttling to emulate slow CPUs.}
\usage{
Emulation.setCPUThrottlingRate(promise, rate)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{rate}{A numeric.
Throttling rate as a slowdown factor (1 is no throttle, 2 is 2x slowdown, etc).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Emulation.setCPUThrottlingRate

Enables CPU throttling to emulate slow CPUs.
}
