% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setDefaultBackgroundColorOverride}
\alias{Emulation.setDefaultBackgroundColorOverride}
\title{Send the command Emulation.setDefaultBackgroundColorOverride

Sets or clears an override of the default background color of the frame. This override is used
if the content does not specify one.}
\usage{
Emulation.setDefaultBackgroundColorOverride(promise, color = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{color}{Optional. A DOM.RGBA.
RGBA of the default background color. If not specified, any existing override will be
cleared.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Emulation.setDefaultBackgroundColorOverride

Sets or clears an override of the default background color of the frame. This override is used
if the content does not specify one.
}
