% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setDocumentCookieDisabled}
\alias{Emulation.setDocumentCookieDisabled}
\title{Send the command Emulation.setDocumentCookieDisabled}
\usage{
Emulation.setDocumentCookieDisabled(promise, disabled)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{disabled}{A logical.
Whether document.coookie API should be disabled.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Emulation.setDocumentCookieDisabled
}
