% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setEmitTouchEventsForMouse}
\alias{Emulation.setEmitTouchEventsForMouse}
\title{Send the command Emulation.setEmitTouchEventsForMouse}
\usage{
Emulation.setEmitTouchEventsForMouse(promise, enabled,
  configuration = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{enabled}{A logical.
Whether touch emulation based on mouse input should be enabled.}

\item{configuration}{Optional. A character string.
Touch/gesture events configuration. Default: current platform. Accepted values: mobile, desktop.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Emulation.setEmitTouchEventsForMouse
}
