% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setEmulatedMedia}
\alias{Emulation.setEmulatedMedia}
\title{Send the command Emulation.setEmulatedMedia}
\usage{
Emulation.setEmulatedMedia(promise, media)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{media}{A character string.
Media type to emulate. Empty string disables the override.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Emulates the given media for CSS media queries.
}
