% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setFocusEmulationEnabled}
\alias{Emulation.setFocusEmulationEnabled}
\title{Send the command Emulation.setFocusEmulationEnabled

Enables or disables simulating a focused and active page.}
\usage{
Emulation.setFocusEmulationEnabled(promise, enabled)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{enabled}{A logical.
Whether to enable to disable focus emulation.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Emulation.setFocusEmulationEnabled

Enables or disables simulating a focused and active page.
}
