% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setPageScaleFactor}
\alias{Emulation.setPageScaleFactor}
\title{Send the command Emulation.setPageScaleFactor}
\usage{
Emulation.setPageScaleFactor(promise, pageScaleFactor)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{pageScaleFactor}{A numeric.
Page scale factor.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sets a specified page scale factor.
}
