% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setScriptExecutionDisabled}
\alias{Emulation.setScriptExecutionDisabled}
\title{Send the command Emulation.setScriptExecutionDisabled}
\usage{
Emulation.setScriptExecutionDisabled(promise, value)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{value}{A logical.
Whether script execution should be disabled in the page.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Switches script execution in the page.
}
