% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setScrollbarsHidden}
\alias{Emulation.setScrollbarsHidden}
\title{Send the command Emulation.setScrollbarsHidden}
\usage{
Emulation.setScrollbarsHidden(promise, hidden)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{hidden}{A logical.
Whether scrollbars should be always hidden.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Emulation.setScrollbarsHidden
}
