% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setTouchEmulationEnabled}
\alias{Emulation.setTouchEmulationEnabled}
\title{Send the command Emulation.setTouchEmulationEnabled}
\usage{
Emulation.setTouchEmulationEnabled(promise, enabled,
  maxTouchPoints = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{enabled}{A logical.
Whether the touch event emulation should be enabled.}

\item{maxTouchPoints}{Optional. An integer.
Maximum touch points supported. Defaults to one.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables touch on platforms which do not support them.
}
