% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setVirtualTimePolicy}
\alias{Emulation.setVirtualTimePolicy}
\title{Send the command Emulation.setVirtualTimePolicy

Turns on virtual time for all frames (replacing real-time with a synthetic time source) and sets
the current virtual time policy.  Note this supersedes any previous time budget.}
\usage{
Emulation.setVirtualTimePolicy(promise, policy, budget = NULL,
  maxVirtualTimeTaskStarvationCount = NULL, waitForNavigation = NULL,
  initialVirtualTime = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{policy}{A VirtualTimePolicy.}

\item{budget}{Optional. A numeric.
If set, after this many virtual milliseconds have elapsed virtual time will be paused and a
virtualTimeBudgetExpired event is sent.}

\item{maxVirtualTimeTaskStarvationCount}{Optional. An integer.
If set this specifies the maximum number of tasks that can be run before virtual is forced
forwards to prevent deadlock.}

\item{waitForNavigation}{Optional. A logical.
If set the virtual time policy change should be deferred until any frame starts navigating.
Note any previous deferred policy change is superseded.}

\item{initialVirtualTime}{Optional. A Network.TimeSinceEpoch.
If set, base::Time::Now will be overriden to initially return this value.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Emulation.setVirtualTimePolicy

Turns on virtual time for all frames (replacing real-time with a synthetic time source) and sets
the current virtual time policy.  Note this supersedes any previous time budget.
}
