% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch_commands.R
\name{Fetch.continueRequest}
\alias{Fetch.continueRequest}
\title{Send the command Fetch.continueRequest

Continues the request, optionally modifying some of its parameters.}
\usage{
Fetch.continueRequest(promise, requestId, url = NULL, method = NULL,
  postData = NULL, headers = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{requestId}{A RequestId.
An id the client received in requestPaused event.}

\item{url}{Optional. A character string.
If set, the request url will be modified in a way that's not observable by page.}

\item{method}{Optional. A character string.
If set, the request method is overridden.}

\item{postData}{Optional. A character string.
If set, overrides the post data in the request.}

\item{headers}{Optional. A list of HeaderEntry.
If set, overrides the request headrts.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Fetch.continueRequest

Continues the request, optionally modifying some of its parameters.
}
