% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch_commands.R
\name{Fetch.continueWithAuth}
\alias{Fetch.continueWithAuth}
\title{Send the command Fetch.continueWithAuth

Continues a request supplying authChallengeResponse following authRequired event.}
\usage{
Fetch.continueWithAuth(promise, requestId, authChallengeResponse)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{requestId}{A RequestId.
An id the client received in authRequired event.}

\item{authChallengeResponse}{A AuthChallengeResponse.
Response to  with an authChallenge.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Fetch.continueWithAuth

Continues a request supplying authChallengeResponse following authRequired event.
}
