% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch_commands.R
\name{Fetch.enable}
\alias{Fetch.enable}
\title{Send the command Fetch.enable}
\usage{
Fetch.enable(promise, patterns = NULL, handleAuthRequests = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{patterns}{Optional. A list of RequestPattern.
If specified, only requests matching any of these patterns will produce
fetchRequested event and will be paused until clients response. If not set,
all requests will be affected.}

\item{handleAuthRequests}{Optional. A logical.
If true, authRequired events will be issued and requests will be paused
expecting a call to continueWithAuth.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables issuing of requestPaused events. A request will be paused until client
calls one of failRequest, fulfillRequest or continueRequest/continueWithAuth.
}
