% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch_commands.R
\name{Fetch.failRequest}
\alias{Fetch.failRequest}
\title{Send the command Fetch.failRequest

Causes the request to fail with specified reason.}
\usage{
Fetch.failRequest(promise, requestId, errorReason)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{requestId}{A RequestId.
An id the client received in requestPaused event.}

\item{errorReason}{A Network.ErrorReason.
Causes the request to fail with the given reason.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Fetch.failRequest

Causes the request to fail with specified reason.
}
