% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch_commands.R
\name{Fetch.fulfillRequest}
\alias{Fetch.fulfillRequest}
\title{Send the command Fetch.fulfillRequest

Provides response to the request.}
\usage{
Fetch.fulfillRequest(promise, requestId, responseCode, responseHeaders,
  body = NULL, responsePhrase = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{requestId}{A RequestId.
An id the client received in requestPaused event.}

\item{responseCode}{An integer.
An HTTP response code.}

\item{responseHeaders}{A list of HeaderEntry.
Response headers.}

\item{body}{Optional. A character string.
A response body.}

\item{responsePhrase}{Optional. A character string.
A textual representation of responseCode.
If absent, a standard phrase mathcing responseCode is used.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Fetch.fulfillRequest

Provides response to the request.
}
