% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chrome.R
\name{find_chrome_binary}
\alias{find_chrome_binary}
\title{Find Google Chrome binary in the system}
\usage{
find_chrome_binary()
}
\value{
A character string. The path the chrome binary that will
be used by \code{crrri}.
}
\description{
If the chrome binary path has not already been set in \var{HEADLESS_CHROME}
environment variable, the function will try to find the chrome binary
on your system using a some hints.
}
\details{
\subsection{Windows}{

It will look in the registry for an installed version
}

\subsection{macOS,}{

It will return a hard-coded path of Chrome under \file{/Applications}.
}

\subsection{Linux,}{

It will search for \command{chromium-browser} and \command{google-chrome} from
the system's \var{PATH} variable.
}
}
\note{
From \code{pagedown} R package, licence MIT.
}
\references{
\href{https://github.com/rstudio/pagedown/blob/b93f46fc1ad70182e5dd3d9fc843f752fd12f780/R/chrome.R#L213}{Source on Github}
}
\author{
Yihui Xie, Romain Lesur, Christophe Dervieux
}
