% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeadlessExperimental_commands.R
\name{HeadlessExperimental.disable}
\alias{HeadlessExperimental.disable}
\title{Send the command HeadlessExperimental.disable

Disables headless events for the target.}
\usage{
HeadlessExperimental.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command HeadlessExperimental.disable

Disables headless events for the target.
}
