% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapProfiler_commands.R
\name{HeapProfiler.addInspectedHeapObject}
\alias{HeapProfiler.addInspectedHeapObject}
\title{Send the command HeapProfiler.addInspectedHeapObject}
\usage{
HeapProfiler.addInspectedHeapObject(promise, heapObjectId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{heapObjectId}{A HeapSnapshotObjectId.
Heap snapshot object id to be accessible by means of $x command line API.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables console to refer to the node with given id via $x (see Command Line API for more details
$x functions).
}
