% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapProfiler_commands.R
\name{HeapProfiler.getHeapObjectId}
\alias{HeapProfiler.getHeapObjectId}
\title{Send the command HeapProfiler.getHeapObjectId}
\usage{
HeapProfiler.getHeapObjectId(promise, objectId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{objectId}{A Runtime.RemoteObjectId.
Identifier of the object to get heap object id for.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command HeapProfiler.getHeapObjectId
}
