% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapProfiler_commands.R
\name{HeapProfiler.getObjectByHeapObjectId}
\alias{HeapProfiler.getObjectByHeapObjectId}
\title{Send the command HeapProfiler.getObjectByHeapObjectId}
\usage{
HeapProfiler.getObjectByHeapObjectId(promise, objectId,
  objectGroup = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{objectId}{A HeapSnapshotObjectId.}

\item{objectGroup}{Optional. A character string.
Symbolic group name that can be used to release multiple objects.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command HeapProfiler.getObjectByHeapObjectId
}
