% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapProfiler_commands.R
\name{HeapProfiler.startSampling}
\alias{HeapProfiler.startSampling}
\title{Send the command HeapProfiler.startSampling}
\usage{
HeapProfiler.startSampling(promise, samplingInterval = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{samplingInterval}{Optional. A numeric.
Average sample interval in bytes. Poisson distribution is used for the intervals. The
default value is 32768 bytes.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command HeapProfiler.startSampling
}
