% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapProfiler_commands.R
\name{HeapProfiler.startTrackingHeapObjects}
\alias{HeapProfiler.startTrackingHeapObjects}
\title{Send the command HeapProfiler.startTrackingHeapObjects}
\usage{
HeapProfiler.startTrackingHeapObjects(promise, trackAllocations = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{trackAllocations}{Optional. A logical.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command HeapProfiler.startTrackingHeapObjects
}
