% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapProfiler_commands.R
\name{HeapProfiler.stopSampling}
\alias{HeapProfiler.stopSampling}
\title{Send the command HeapProfiler.stopSampling}
\usage{
HeapProfiler.stopSampling(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command HeapProfiler.stopSampling
}
