% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapProfiler_commands.R
\name{HeapProfiler.takeHeapSnapshot}
\alias{HeapProfiler.takeHeapSnapshot}
\title{Send the command HeapProfiler.takeHeapSnapshot}
\usage{
HeapProfiler.takeHeapSnapshot(promise, reportProgress = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{reportProgress}{Optional. A logical.
If true 'reportHeapSnapshotProgress' events will be generated while snapshot is being taken.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command HeapProfiler.takeHeapSnapshot
}
