% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_methods.R
\name{http-methods}
\alias{http-methods}
\alias{fetch_protocol}
\title{Chrome Debugging Protocol HTTP methods}
\usage{
fetch_protocol(host = "localhost", port = 9222, secure = FALSE,
  local = FALSE)
}
\arguments{
\item{host}{Character scalar, the host name of the application.}

\item{port}{The remote debugging port (a numeric or a character scalar).}

\item{secure}{A logical indicating whether a secure protocol shall be used.}

\item{local}{Logical scalar, indicating whether the local version of the
protocol (embedded in \code{crrri}) must be used or the protocol must be
fetched \emph{remotely}.}
}
\value{
A list.
}
\description{
Chrome Debugging Protocol HTTP methods
}
\examples{
\donttest{
# fetch the local protocol
protocol <- fetch_protocol(local = TRUE)

# fetch the protocol of a remote application
remote <- Chrome$new()
protocol <- fetch_protocol()
remote$close()
}
}
\keyword{internal}
