% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedDB_commands.R
\name{IndexedDB.clearObjectStore}
\alias{IndexedDB.clearObjectStore}
\title{Send the command IndexedDB.clearObjectStore

Clears all entries from an object store.}
\usage{
IndexedDB.clearObjectStore(promise, securityOrigin, databaseName,
  objectStoreName)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{securityOrigin}{A character string.
Security origin.}

\item{databaseName}{A character string.
Database name.}

\item{objectStoreName}{A character string.
Object store name.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command IndexedDB.clearObjectStore

Clears all entries from an object store.
}
