% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedDB_commands.R
\name{IndexedDB.deleteDatabase}
\alias{IndexedDB.deleteDatabase}
\title{Send the command IndexedDB.deleteDatabase

Deletes a database.}
\usage{
IndexedDB.deleteDatabase(promise, securityOrigin, databaseName)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{securityOrigin}{A character string.
Security origin.}

\item{databaseName}{A character string.
Database name.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command IndexedDB.deleteDatabase

Deletes a database.
}
