% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedDB_commands.R
\name{IndexedDB.disable}
\alias{IndexedDB.disable}
\title{Send the command IndexedDB.disable

Disables events from backend.}
\usage{
IndexedDB.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command IndexedDB.disable

Disables events from backend.
}
