% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedDB_commands.R
\name{IndexedDB.enable}
\alias{IndexedDB.enable}
\title{Send the command IndexedDB.enable

Enables events from backend.}
\usage{
IndexedDB.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command IndexedDB.enable

Enables events from backend.
}
