% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedDB_commands.R
\name{IndexedDB.requestData}
\alias{IndexedDB.requestData}
\title{Send the command IndexedDB.requestData

Requests data from object store or index.}
\usage{
IndexedDB.requestData(promise, securityOrigin, databaseName,
  objectStoreName, indexName, skipCount, pageSize, keyRange = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{securityOrigin}{A character string.
Security origin.}

\item{databaseName}{A character string.
Database name.}

\item{objectStoreName}{A character string.
Object store name.}

\item{indexName}{A character string.
Index name, empty string for object store data requests.}

\item{skipCount}{An integer.
Number of records to skip.}

\item{pageSize}{An integer.
Number of records to fetch.}

\item{keyRange}{Optional. A KeyRange.
Key range.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command IndexedDB.requestData

Requests data from object store or index.
}
