% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.dispatchKeyEvent}
\alias{Input.dispatchKeyEvent}
\title{Send the command Input.dispatchKeyEvent}
\usage{
Input.dispatchKeyEvent(promise, type, modifiers = NULL,
  timestamp = NULL, text = NULL, unmodifiedText = NULL,
  keyIdentifier = NULL, code = NULL, key = NULL,
  windowsVirtualKeyCode = NULL, nativeVirtualKeyCode = NULL,
  autoRepeat = NULL, isKeypad = NULL, isSystemKey = NULL,
  location = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{type}{A character string.
Type of the key event. Accepted values: keyDown, keyUp, rawKeyDown, char.}

\item{modifiers}{Optional. An integer.
Bit field representing pressed modifier keys. Alt=1, Ctrl=2, Meta/Command=4, Shift=8
(default: 0).}

\item{timestamp}{Optional. A TimeSinceEpoch.
Time at which the event occurred.}

\item{text}{Optional. A character string.
Text as generated by processing a virtual key code with a keyboard layout. Not needed for
for \code{keyUp} and \code{rawKeyDown} events (default: "")}

\item{unmodifiedText}{Optional. A character string.
Text that would have been generated by the keyboard if no modifiers were pressed (except for
shift). Useful for shortcut (accelerator) key handling (default: "").}

\item{keyIdentifier}{Optional. A character string.
Unique key identifier (e.g., 'U+0041') (default: "").}

\item{code}{Optional. A character string.
Unique DOM defined string value for each physical key (e.g., 'KeyA') (default: "").}

\item{key}{Optional. A character string.
Unique DOM defined string value describing the meaning of the key in the context of active
modifiers, keyboard layout, etc (e.g., 'AltGr') (default: "").}

\item{windowsVirtualKeyCode}{Optional. An integer.
Windows virtual key code (default: 0).}

\item{nativeVirtualKeyCode}{Optional. An integer.
Native virtual key code (default: 0).}

\item{autoRepeat}{Optional. A logical.
Whether the event was generated from auto repeat (default: false).}

\item{isKeypad}{Optional. A logical.
Whether the event was generated from the keypad (default: false).}

\item{isSystemKey}{Optional. A logical.
Whether the event was a system key event (default: false).}

\item{location}{Optional. An integer.
Whether the event was from the left or right side of the keyboard. 1=Left, 2=Right (default:
0).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Dispatches a key event to the page.
}
