% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.dispatchMouseEvent}
\alias{Input.dispatchMouseEvent}
\title{Send the command Input.dispatchMouseEvent}
\usage{
Input.dispatchMouseEvent(promise, type, x, y, modifiers = NULL,
  timestamp = NULL, button = NULL, clickCount = NULL,
  deltaX = NULL, deltaY = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{type}{A character string.
Type of the mouse event. Accepted values: mousePressed, mouseReleased, mouseMoved, mouseWheel.}

\item{x}{A numeric.
X coordinate of the event relative to the main frame's viewport in CSS pixels.}

\item{y}{A numeric.
Y coordinate of the event relative to the main frame's viewport in CSS pixels. 0 refers to
the top of the viewport and Y increases as it proceeds towards the bottom of the viewport.}

\item{modifiers}{Optional. An integer.
Bit field representing pressed modifier keys. Alt=1, Ctrl=2, Meta/Command=4, Shift=8
(default: 0).}

\item{timestamp}{Optional. A TimeSinceEpoch.
Time at which the event occurred.}

\item{button}{Optional. A character string.
Mouse button (default: "none"). Accepted values: none, left, middle, right.}

\item{clickCount}{Optional. An integer.
Number of times the mouse button was clicked (default: 0).}

\item{deltaX}{Optional. A numeric.
X delta in CSS pixels for mouse wheel event (default: 0).}

\item{deltaY}{Optional. A numeric.
Y delta in CSS pixels for mouse wheel event (default: 0).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Dispatches a mouse event to the page.
}
