% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.dispatchTouchEvent}
\alias{Input.dispatchTouchEvent}
\title{Send the command Input.dispatchTouchEvent}
\usage{
Input.dispatchTouchEvent(promise, type, touchPoints, modifiers = NULL,
  timestamp = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{type}{A character string.
Type of the touch event. TouchEnd and TouchCancel must not contain any touch points, while
TouchStart and TouchMove must contains at least one. Accepted values: touchStart, touchEnd, touchMove, touchCancel.}

\item{touchPoints}{A list of TouchPoint.
Active touch points on the touch device. One event per any changed point (compared to
previous touch event in a sequence) is generated, emulating pressing/moving/releasing points
one by one.}

\item{modifiers}{Optional. An integer.
Bit field representing pressed modifier keys. Alt=1, Ctrl=2, Meta/Command=4, Shift=8
(default: 0).}

\item{timestamp}{Optional. A TimeSinceEpoch.
Time at which the event occurred.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Dispatches a touch event to the page.
}
