% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.emulateTouchFromMouseEvent}
\alias{Input.emulateTouchFromMouseEvent}
\title{Send the command Input.emulateTouchFromMouseEvent}
\usage{
Input.emulateTouchFromMouseEvent(promise, type, x, y, button,
  timestamp = NULL, deltaX = NULL, deltaY = NULL, modifiers = NULL,
  clickCount = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{type}{A character string.
Type of the mouse event. Accepted values: mousePressed, mouseReleased, mouseMoved, mouseWheel.}

\item{x}{An integer.
X coordinate of the mouse pointer in DIP.}

\item{y}{An integer.
Y coordinate of the mouse pointer in DIP.}

\item{button}{A character string.
Mouse button. Accepted values: none, left, middle, right.}

\item{timestamp}{Optional. A TimeSinceEpoch.
Time at which the event occurred (default: current time).}

\item{deltaX}{Optional. A numeric.
X delta in DIP for mouse wheel event (default: 0).}

\item{deltaY}{Optional. A numeric.
Y delta in DIP for mouse wheel event (default: 0).}

\item{modifiers}{Optional. An integer.
Bit field representing pressed modifier keys. Alt=1, Ctrl=2, Meta/Command=4, Shift=8
(default: 0).}

\item{clickCount}{Optional. An integer.
Number of times the mouse button was clicked (default: 0).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Emulates touch event from the mouse event parameters.
}
