% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.insertText}
\alias{Input.insertText}
\title{Send the command Input.insertText}
\usage{
Input.insertText(promise, text)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{text}{A character string.
The text to insert.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
This method emulates inserting text that doesn't come from a key press,
for example an emoji keyboard or an IME.
}
