% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.setIgnoreInputEvents}
\alias{Input.setIgnoreInputEvents}
\title{Send the command Input.setIgnoreInputEvents

Ignores input events (useful while auditing page).}
\usage{
Input.setIgnoreInputEvents(promise, ignore)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{ignore}{A logical.
Ignores input events processing when set to true.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Input.setIgnoreInputEvents

Ignores input events (useful while auditing page).
}
