% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.synthesizePinchGesture}
\alias{Input.synthesizePinchGesture}
\title{Send the command Input.synthesizePinchGesture}
\usage{
Input.synthesizePinchGesture(promise, x, y, scaleFactor,
  relativeSpeed = NULL, gestureSourceType = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{x}{A numeric.
X coordinate of the start of the gesture in CSS pixels.}

\item{y}{A numeric.
Y coordinate of the start of the gesture in CSS pixels.}

\item{scaleFactor}{A numeric.
Relative scale factor after zooming (>1.0 zooms in, <1.0 zooms out).}

\item{relativeSpeed}{Optional. An integer.
Relative pointer speed in pixels per second (default: 800).}

\item{gestureSourceType}{Optional. A GestureSourceType.
Which type of input events to be generated (default: 'default', which queries the platform
for the preferred input type).}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Synthesizes a pinch gesture over a time period by issuing appropriate touch events.
}
