% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.synthesizeScrollGesture}
\alias{Input.synthesizeScrollGesture}
\title{Send the command Input.synthesizeScrollGesture}
\usage{
Input.synthesizeScrollGesture(promise, x, y, xDistance = NULL,
  yDistance = NULL, xOverscroll = NULL, yOverscroll = NULL,
  preventFling = NULL, speed = NULL, gestureSourceType = NULL,
  repeatCount = NULL, repeatDelayMs = NULL,
  interactionMarkerName = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{x}{A numeric.
X coordinate of the start of the gesture in CSS pixels.}

\item{y}{A numeric.
Y coordinate of the start of the gesture in CSS pixels.}

\item{xDistance}{Optional. A numeric.
The distance to scroll along the X axis (positive to scroll left).}

\item{yDistance}{Optional. A numeric.
The distance to scroll along the Y axis (positive to scroll up).}

\item{xOverscroll}{Optional. A numeric.
The number of additional pixels to scroll back along the X axis, in addition to the given
distance.}

\item{yOverscroll}{Optional. A numeric.
The number of additional pixels to scroll back along the Y axis, in addition to the given
distance.}

\item{preventFling}{Optional. A logical.
Prevent fling (default: true).}

\item{speed}{Optional. An integer.
Swipe speed in pixels per second (default: 800).}

\item{gestureSourceType}{Optional. A GestureSourceType.
Which type of input events to be generated (default: 'default', which queries the platform
for the preferred input type).}

\item{repeatCount}{Optional. An integer.
The number of times to repeat the gesture (default: 0).}

\item{repeatDelayMs}{Optional. An integer.
The number of milliseconds delay between each repeat. (default: 250).}

\item{interactionMarkerName}{Optional. A character string.
The name of the interaction markers to generate, if not empty (default: "").}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Synthesizes a scroll gesture over a time period by issuing appropriate touch events.
}
