% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.synthesizeTapGesture}
\alias{Input.synthesizeTapGesture}
\title{Send the command Input.synthesizeTapGesture

Synthesizes a tap gesture over a time period by issuing appropriate touch events.}
\usage{
Input.synthesizeTapGesture(promise, x, y, duration = NULL,
  tapCount = NULL, gestureSourceType = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{x}{A numeric.
X coordinate of the start of the gesture in CSS pixels.}

\item{y}{A numeric.
Y coordinate of the start of the gesture in CSS pixels.}

\item{duration}{Optional. An integer.
Duration between touchdown and touchup events in ms (default: 50).}

\item{tapCount}{Optional. An integer.
Number of times to perform the tap (e.g. 2 for double tap, default: 1).}

\item{gestureSourceType}{Optional. A GestureSourceType.
Which type of input events to be generated (default: 'default', which queries the platform
for the preferred input type).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Input.synthesizeTapGesture

Synthesizes a tap gesture over a time period by issuing appropriate touch events.
}
