% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_commands.R
\name{IO.close}
\alias{IO.close}
\title{Send the command IO.close}
\usage{
IO.close(promise, handle)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{handle}{A StreamHandle.
Handle of the stream to close.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Close the stream, discard any temporary backing storage.
}
