% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_commands.R
\name{IO.resolveBlob}
\alias{IO.resolveBlob}
\title{Send the command IO.resolveBlob

Return UUID of Blob object specified by a remote object id.}
\usage{
IO.resolveBlob(promise, objectId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{objectId}{A Runtime.RemoteObjectId.
Object id of a Blob object wrapper.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command IO.resolveBlob

Return UUID of Blob object specified by a remote object id.
}
