% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerTree_commands.R
\name{LayerTree.disable}
\alias{LayerTree.disable}
\title{Send the command LayerTree.disable

Disables compositing tree inspection.}
\usage{
LayerTree.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command LayerTree.disable

Disables compositing tree inspection.
}
