% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerTree_commands.R
\name{LayerTree.enable}
\alias{LayerTree.enable}
\title{Send the command LayerTree.enable}
\usage{
LayerTree.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables compositing tree inspection.
}
