% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerTree_commands.R
\name{LayerTree.loadSnapshot}
\alias{LayerTree.loadSnapshot}
\title{Send the command LayerTree.loadSnapshot

Returns the snapshot identifier.}
\usage{
LayerTree.loadSnapshot(promise, tiles)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{tiles}{A list of PictureTile.
An array of tiles composing the snapshot.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command LayerTree.loadSnapshot

Returns the snapshot identifier.
}
