% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerTree_commands.R
\name{LayerTree.makeSnapshot}
\alias{LayerTree.makeSnapshot}
\title{Send the command LayerTree.makeSnapshot}
\usage{
LayerTree.makeSnapshot(promise, layerId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{layerId}{A LayerId.
The id of the layer.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns the layer snapshot identifier.
}
