% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerTree_commands.R
\name{LayerTree.profileSnapshot}
\alias{LayerTree.profileSnapshot}
\title{Send the command LayerTree.profileSnapshot}
\usage{
LayerTree.profileSnapshot(promise, snapshotId, minRepeatCount = NULL,
  minDuration = NULL, clipRect = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{snapshotId}{A SnapshotId.
The id of the layer snapshot.}

\item{minRepeatCount}{Optional. An integer.
The maximum number of times to replay the snapshot (1, if not specified).}

\item{minDuration}{Optional. A numeric.
The minimum duration (in seconds) to replay the snapshot.}

\item{clipRect}{Optional. A DOM.Rect.
The clip rectangle to apply when replaying the snapshot.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command LayerTree.profileSnapshot
}
