% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerTree_commands.R
\name{LayerTree.replaySnapshot}
\alias{LayerTree.replaySnapshot}
\title{Send the command LayerTree.replaySnapshot}
\usage{
LayerTree.replaySnapshot(promise, snapshotId, fromStep = NULL,
  toStep = NULL, scale = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{snapshotId}{A SnapshotId.
The id of the layer snapshot.}

\item{fromStep}{Optional. An integer.
The first step to replay from (replay from the very start if not specified).}

\item{toStep}{Optional. An integer.
The last step to replay to (replay till the end if not specified).}

\item{scale}{Optional. A numeric.
The scale to apply while replaying (defaults to 1).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Replays the layer snapshot and returns the resulting bitmap.
}
