% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerTree_commands.R
\name{LayerTree.snapshotCommandLog}
\alias{LayerTree.snapshotCommandLog}
\title{Send the command LayerTree.snapshotCommandLog}
\usage{
LayerTree.snapshotCommandLog(promise, snapshotId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{snapshotId}{A SnapshotId.
The id of the layer snapshot.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Replays the layer snapshot and returns canvas log.
}
