% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_commands.R
\name{Log.clear}
\alias{Log.clear}
\title{Send the command Log.clear

Clears the log.}
\usage{
Log.clear(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Log.clear

Clears the log.
}
