% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_commands.R
\name{Log.disable}
\alias{Log.disable}
\title{Send the command Log.disable}
\usage{
Log.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Disables log domain, prevents further log entries from being reported to the client.
}
