% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_commands.R
\name{Log.enable}
\alias{Log.enable}
\title{Send the command Log.enable

Enables log domain, sends the entries collected so far to the client by means of the
\code{entryAdded} notification.}
\usage{
Log.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Log.enable

Enables log domain, sends the entries collected so far to the client by means of the
\code{entryAdded} notification.
}
