% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_commands.R
\name{Log.startViolationsReport}
\alias{Log.startViolationsReport}
\title{Send the command Log.startViolationsReport

start violation reporting.}
\usage{
Log.startViolationsReport(promise, config)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{config}{A list of ViolationSetting.
Configuration for violations.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Log.startViolationsReport

start violation reporting.
}
