% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_commands.R
\name{Log.stopViolationsReport}
\alias{Log.stopViolationsReport}
\title{Send the command Log.stopViolationsReport

Stop violation reporting.}
\usage{
Log.stopViolationsReport(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Log.stopViolationsReport

Stop violation reporting.
}
