% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory_commands.R
\name{Memory.getAllTimeSamplingProfile}
\alias{Memory.getAllTimeSamplingProfile}
\title{Send the command Memory.getAllTimeSamplingProfile

Retrieve native memory allocations profile
collected since renderer process startup.}
\usage{
Memory.getAllTimeSamplingProfile(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Memory.getAllTimeSamplingProfile

Retrieve native memory allocations profile
collected since renderer process startup.
}
