% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory_commands.R
\name{Memory.getBrowserSamplingProfile}
\alias{Memory.getBrowserSamplingProfile}
\title{Send the command Memory.getBrowserSamplingProfile

Retrieve native memory allocations profile
collected since browser process startup.}
\usage{
Memory.getBrowserSamplingProfile(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Memory.getBrowserSamplingProfile

Retrieve native memory allocations profile
collected since browser process startup.
}
