% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory_commands.R
\name{Memory.getDOMCounters}
\alias{Memory.getDOMCounters}
\title{Send the command Memory.getDOMCounters}
\usage{
Memory.getDOMCounters(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 3.
}
\description{
Send the command Memory.getDOMCounters
}
