% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory_commands.R
\name{Memory.setPressureNotificationsSuppressed}
\alias{Memory.setPressureNotificationsSuppressed}
\title{Send the command Memory.setPressureNotificationsSuppressed}
\usage{
Memory.setPressureNotificationsSuppressed(promise, suppressed)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{suppressed}{A logical.
If true, memory pressure notifications will be suppressed.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enable/disable suppressing memory pressure notifications in all processes.
}
