% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory_commands.R
\name{Memory.simulatePressureNotification}
\alias{Memory.simulatePressureNotification}
\title{Send the command Memory.simulatePressureNotification

Simulate a memory pressure notification in all processes.}
\usage{
Memory.simulatePressureNotification(promise, level)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{level}{A PressureLevel.
Memory pressure level of the notification.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Memory.simulatePressureNotification

Simulate a memory pressure notification in all processes.
}
