% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory_commands.R
\name{Memory.startSampling}
\alias{Memory.startSampling}
\title{Send the command Memory.startSampling}
\usage{
Memory.startSampling(promise, samplingInterval = NULL,
  suppressRandomness = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{samplingInterval}{Optional. An integer.
Average number of bytes between samples.}

\item{suppressRandomness}{Optional. A logical.
Do not randomize intervals between samples.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Start collecting native memory profile.
}
