% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.clearBrowserCache}
\alias{Network.clearBrowserCache}
\title{Send the command Network.clearBrowserCache

Clears browser cache.}
\usage{
Network.clearBrowserCache(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Network.clearBrowserCache

Clears browser cache.
}
