% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.deleteCookies}
\alias{Network.deleteCookies}
\title{Send the command Network.deleteCookies

Deletes browser cookies with matching name and url or domain/path pair.}
\usage{
Network.deleteCookies(promise, name, url = NULL, domain = NULL,
  path = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{name}{A character string.
Name of the cookies to remove.}

\item{url}{Optional. A character string.
If specified, deletes all the cookies with the given name where domain and path match
provided URL.}

\item{domain}{Optional. A character string.
If specified, deletes only cookies with the exact domain.}

\item{path}{Optional. A character string.
If specified, deletes only cookies with the exact path.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Network.deleteCookies

Deletes browser cookies with matching name and url or domain/path pair.
}
