% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.disable}
\alias{Network.disable}
\title{Send the command Network.disable}
\usage{
Network.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Disables network tracking, prevents network events from being sent to the client.
}
