% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Network.R
\name{Network.emulateNetworkConditions}
\alias{Network.emulateNetworkConditions}
\title{Send the command Network.emulateNetworkConditions}
\usage{
Network.emulateNetworkConditions(promise, offline, latency,
  downloadThroughput, uploadThroughput, connectionType = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{offline}{A logical.
True to emulate internet disconnection.}

\item{latency}{A numeric.
Minimum latency from request sent to response headers received (ms).}

\item{downloadThroughput}{A numeric.
Maximal aggregated download throughput (bytes/sec). -1 disables download throttling.}

\item{uploadThroughput}{A numeric.
Maximal aggregated upload throughput (bytes/sec).  -1 disables upload throttling.}

\item{connectionType}{Optional. A ConnectionType.
Connection type if known.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Activates emulation of network conditions.
}
