% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.enable}
\alias{Network.enable}
\title{Send the command Network.enable}
\usage{
Network.enable(promise, maxTotalBufferSize = NULL,
  maxResourceBufferSize = NULL, maxPostDataSize = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{maxTotalBufferSize}{Experimental. Optional. An integer.
Buffer size in bytes to use when preserving network payloads (XHRs, etc).}

\item{maxResourceBufferSize}{Experimental. Optional. An integer.
Per-resource buffer size in bytes to use when preserving network payloads (XHRs, etc).}

\item{maxPostDataSize}{Optional. An integer.
Longest post body size (in bytes) that would be included in requestWillBeSent notification}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Enables network tracking, network events will now be delivered to the client.
}
