% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.getAllCookies}
\alias{Network.getAllCookies}
\title{Send the command Network.getAllCookies

Returns all browser cookies. Depending on the backend support, will return detailed cookie
information in the \code{cookies} field.}
\usage{
Network.getAllCookies(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Network.getAllCookies

Returns all browser cookies. Depending on the backend support, will return detailed cookie
information in the \code{cookies} field.
}
