% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.getCertificate}
\alias{Network.getCertificate}
\title{Send the command Network.getCertificate

Returns the DER-encoded certificate.}
\usage{
Network.getCertificate(promise, origin)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{origin}{A character string.
Origin to get certificate for.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Network.getCertificate

Returns the DER-encoded certificate.
}
