% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.getCookies}
\alias{Network.getCookies}
\title{Send the command Network.getCookies

Returns all browser cookies for the current URL. Depending on the backend support, will return
detailed cookie information in the \code{cookies} field.}
\usage{
Network.getCookies(promise, urls = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{urls}{Optional. A list of string.
The list of URLs for which applicable cookies will be fetched}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Network.getCookies

Returns all browser cookies for the current URL. Depending on the backend support, will return
detailed cookie information in the \code{cookies} field.
}
