% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.getRequestPostData}
\alias{Network.getRequestPostData}
\title{Send the command Network.getRequestPostData

Returns post data sent with the request. Returns an error when no data was sent with the request.}
\usage{
Network.getRequestPostData(promise, requestId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{requestId}{A RequestId.
Identifier of the network request to get content for.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Network.getRequestPostData

Returns post data sent with the request. Returns an error when no data was sent with the request.
}
