% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.getResponseBody}
\alias{Network.getResponseBody}
\title{Send the command Network.getResponseBody}
\usage{
Network.getResponseBody(promise, requestId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{requestId}{A RequestId.
Identifier of the network request to get content for.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Returns content served for the given request.
}
