% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.getResponseBodyForInterception}
\alias{Network.getResponseBodyForInterception}
\title{Send the command Network.getResponseBodyForInterception

Returns content served for the given currently intercepted request.}
\usage{
Network.getResponseBodyForInterception(promise, interceptionId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{interceptionId}{A InterceptionId.
Identifier for the intercepted request to get body for.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command Network.getResponseBodyForInterception

Returns content served for the given currently intercepted request.
}
