% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.replayXHR}
\alias{Network.replayXHR}
\title{Send the command Network.replayXHR}
\usage{
Network.replayXHR(promise, requestId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{requestId}{A RequestId.
Identifier of XHR to replay.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
This method sends a new XMLHttpRequest which is identical to the original one. The following
parameters should be identical: method, url, async, request body, extra headers, withCredentials
attribute, user, password.
}
