% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_events.R
\name{Network.requestWillBeSent}
\alias{Network.requestWillBeSent}
\title{Await the event Network.requestWillBeSent or create a callback}
\usage{
Network.requestWillBeSent(promise, requestId = NULL, loaderId = NULL,
  documentURL = NULL, request = NULL, timestamp = NULL,
  wallTime = NULL, initiator = NULL, redirectResponse = NULL,
  type = NULL, frameId = NULL, hasUserGesture = NULL,
  .callback = NULL)
}
\arguments{
\item{promise}{An asynchronous result object.}

\item{requestId}{A RequestId.
Request identifier. Accepted values: \code{~ .res$requestId} (to refer to the previous result).}

\item{loaderId}{A LoaderId.
Loader identifier. Empty string if the request is fetched from worker. Accepted values: \code{~ .res$loaderId} (to refer to the previous result).}

\item{documentURL}{A character string.
URL of the document this request is loaded for. Accepted values: \code{~ .res$documentURL} (to refer to the previous result).}

\item{request}{A Request.
Request data. Accepted values: \code{~ .res$request} (to refer to the previous result).}

\item{timestamp}{A MonotonicTime.
Timestamp. Accepted values: \code{~ .res$timestamp} (to refer to the previous result).}

\item{wallTime}{A TimeSinceEpoch.
Timestamp. Accepted values: \code{~ .res$wallTime} (to refer to the previous result).}

\item{initiator}{A Initiator.
Request initiator. Accepted values: \code{~ .res$initiator} (to refer to the previous result).}

\item{redirectResponse}{A Response.
Redirect response data. Accepted values: \code{~ .res$redirectResponse} (to refer to the previous result).}

\item{type}{A ResourceType.
Type of this resource. Accepted values: \code{~ .res$type} (to refer to the previous result).}

\item{frameId}{A Page.FrameId.
Frame identifier. Accepted values: \code{~ .res$frameId} (to refer to the previous result).}

\item{hasUserGesture}{A logical.
Whether the request is initiated by a user gesture. Defaults to false. Accepted values: \code{~ .res$hasUserGesture} (to refer to the previous result).}

\item{.callback}{A callback function taking one argument. The object passed to
this function is the message received from Chrome: this is a named list
with an element \code{method} (that is equal to \code{"requestWillBeSent"})
and an element \code{params} which is a named list.
The \code{params} list is composed of
the following element(s): \code{requestId}, \code{loaderId}, \code{documentURL}, \code{request}, \code{timestamp}, \code{wallTime}, \code{initiator}, \code{redirectResponse} (optional) , \code{type} (optional) , \code{frameId} (optional) , \code{hasUserGesture} (optional) .}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the use of a callback function.
When \code{.callback} is \code{NULL}, the promise is fulfilled when the event is received.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
\code{result} is a named list: its elements are the parameters sended by Chrome.
You can chain this promise with another command or event listener.
When \code{.callback} is not \code{NULL}, the promise is fulfilled as soon as the callback is created; the value
is a function without any argument that can be called to cancel the callback. When you use the
\code{.callback} argument, you cannot send the result to any other command or event listener.
}
\description{
\strong{Event description}:
Fired when page is about to send HTTP request.
}
