% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.searchInResponseBody}
\alias{Network.searchInResponseBody}
\title{Send the command Network.searchInResponseBody

Searches for given string in response content.}
\usage{
Network.searchInResponseBody(promise, requestId, query,
  caseSensitive = NULL, isRegex = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{requestId}{A RequestId.
Identifier of the network response to search.}

\item{query}{A character string.
String to search for.}

\item{caseSensitive}{Optional. A logical.
If true, search is case sensitive.}

\item{isRegex}{Optional. A logical.
If true, treats string parameter as regex.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Network.searchInResponseBody

Searches for given string in response content.
}
