% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.setBlockedURLs}
\alias{Network.setBlockedURLs}
\title{Send the command Network.setBlockedURLs}
\usage{
Network.setBlockedURLs(promise, urls)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{urls}{A list of string.
URL patterns to block. Wildcards ('*') are allowed.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Blocks URLs from loading.
}
