% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.setCookie}
\alias{Network.setCookie}
\title{Send the command Network.setCookie}
\usage{
Network.setCookie(promise, name, value, url = NULL, domain = NULL,
  path = NULL, secure = NULL, httpOnly = NULL, sameSite = NULL,
  expires = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{name}{A character string.
Cookie name.}

\item{value}{A character string.
Cookie value.}

\item{url}{Optional. A character string.
The request-URI to associate with the setting of the cookie. This value can affect the
default domain and path values of the created cookie.}

\item{domain}{Optional. A character string.
Cookie domain.}

\item{path}{Optional. A character string.
Cookie path.}

\item{secure}{Optional. A logical.
True if cookie is secure.}

\item{httpOnly}{Optional. A logical.
True if cookie is http-only.}

\item{sameSite}{Optional. A CookieSameSite.
Cookie SameSite type.}

\item{expires}{Optional. A TimeSinceEpoch.
Cookie expiration date, session cookie if not set}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Sets a cookie with the given cookie data; may overwrite equivalent cookies if they exist.
}
