% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.setCookies}
\alias{Network.setCookies}
\title{Send the command Network.setCookies

Sets given cookies.}
\usage{
Network.setCookies(promise, cookies)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{cookies}{A list of CookieParam.
Cookies to be set.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Network.setCookies

Sets given cookies.
}
