% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.setDataSizeLimitsForTest}
\alias{Network.setDataSizeLimitsForTest}
\title{Send the command Network.setDataSizeLimitsForTest

For testing.}
\usage{
Network.setDataSizeLimitsForTest(promise, maxTotalSize, maxResourceSize)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{maxTotalSize}{An integer.
Maximum total buffer size.}

\item{maxResourceSize}{An integer.
Maximum per-resource size.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Network.setDataSizeLimitsForTest

For testing.
}
