% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.setExtraHTTPHeaders}
\alias{Network.setExtraHTTPHeaders}
\title{Send the command Network.setExtraHTTPHeaders}
\usage{
Network.setExtraHTTPHeaders(promise, headers)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{headers}{A Headers.
Map with extra HTTP headers.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Specifies whether to always send extra HTTP headers with the requests from this page.
}
