% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.setRequestInterception}
\alias{Network.setRequestInterception}
\title{Send the command Network.setRequestInterception}
\usage{
Network.setRequestInterception(promise, patterns)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{patterns}{A list of RequestPattern.
Requests matching any of these patterns will be forwarded and wait for the corresponding
continueInterceptedRequest call.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sets the requests to intercept that match a the provided patterns and optionally resource types.
}
