% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.setUserAgentOverride}
\alias{Network.setUserAgentOverride}
\title{Send the command Network.setUserAgentOverride

Allows overriding user agent with the given string.}
\usage{
Network.setUserAgentOverride(promise, userAgent, acceptLanguage = NULL,
  platform = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{userAgent}{A character string.
User agent to use.}

\item{acceptLanguage}{Optional. A character string.
Browser langugage to emulate.}

\item{platform}{Optional. A character string.
The platform navigator.platform should return.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Network.setUserAgentOverride

Allows overriding user agent with the given string.
}
