% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.takeResponseBodyForInterceptionAsStream}
\alias{Network.takeResponseBodyForInterceptionAsStream}
\title{Send the command Network.takeResponseBodyForInterceptionAsStream}
\usage{
Network.takeResponseBodyForInterceptionAsStream(promise, interceptionId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{interceptionId}{A InterceptionId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns a handle to the stream representing the response body. Note that after this command,
the intercepted request can't be continued as is -- you either need to cancel it or to provide
the response body. The stream only supports sequential read, IO.read will fail if the position
is specified.
}
