% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_Network.R
\name{Network.webSocketWillSendHandshakeRequest}
\alias{Network.webSocketWillSendHandshakeRequest}
\title{Await the event Network.webSocketWillSendHandshakeRequest or create a callback}
\usage{
Network.webSocketWillSendHandshakeRequest(promise, requestId = NULL,
  timestamp = NULL, wallTime = NULL, request = NULL,
  .callback = NULL)
}
\arguments{
\item{promise}{An asynchronous result object.}

\item{requestId}{A RequestId.
Request identifier. Accepted values: \code{~ .res$requestId} (to refer to the previous result).}

\item{timestamp}{A MonotonicTime.
Timestamp. Accepted values: \code{~ .res$timestamp} (to refer to the previous result).}

\item{wallTime}{A TimeSinceEpoch.
UTC Timestamp. Accepted values: \code{~ .res$wallTime} (to refer to the previous result).}

\item{request}{A WebSocketRequest.
WebSocket request data. Accepted values: \code{~ .res$request} (to refer to the previous result).}

\item{.callback}{A callback function taking one argument. The object passed to
this function is the message received from Chrome: this is a named list
with an element \code{method} (that is equal to \code{"webSocketWillSendHandshakeRequest"})
and an element \code{params} which is a named list.
The \code{params} list is composed of
the following element(s): \code{requestId}, \code{timestamp}, \code{wallTime}, \code{request}.}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the use of a callback function.
When \code{.callback} is \code{NULL}, the promise is fulfilled when the event is received.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
\code{result} is a named list: its elements are the parameters sended by Chrome.
You can chain this promise with another command or event listener.
When \code{.callback} is not \code{NULL}, the promise is fulfilled as soon as the callback is created; the value
is a function without any argument that can be called to cancel the callback. When you use the
\code{.callback} argument, you cannot send the result to any other command or event listener.
}
\description{
\strong{Event description}:
Fired when WebSocket is about to initiate handshake.
}
