% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.enable}
\alias{Overlay.enable}
\title{Send the command Overlay.enable

Enables domain notifications.}
\usage{
Overlay.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Overlay.enable

Enables domain notifications.
}
