% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.getHighlightObjectForTest}
\alias{Overlay.getHighlightObjectForTest}
\title{Send the command Overlay.getHighlightObjectForTest

For testing.}
\usage{
Overlay.getHighlightObjectForTest(promise, nodeId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A DOM.NodeId.
Id of the node to get highlight object for.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Overlay.getHighlightObjectForTest

For testing.
}
