% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.hideHighlight}
\alias{Overlay.hideHighlight}
\title{Send the command Overlay.hideHighlight

Hides any highlight.}
\usage{
Overlay.hideHighlight(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Overlay.hideHighlight

Hides any highlight.
}
