% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.highlightFrame}
\alias{Overlay.highlightFrame}
\title{Send the command Overlay.highlightFrame

Highlights owner element of the frame with given id.}
\usage{
Overlay.highlightFrame(promise, frameId, contentColor = NULL,
  contentOutlineColor = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{frameId}{A Page.FrameId.
Identifier of the frame to highlight.}

\item{contentColor}{Optional. A DOM.RGBA.
The content box highlight fill color (default: transparent).}

\item{contentOutlineColor}{Optional. A DOM.RGBA.
The content box highlight outline color (default: transparent).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Overlay.highlightFrame

Highlights owner element of the frame with given id.
}
