% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.highlightNode}
\alias{Overlay.highlightNode}
\title{Send the command Overlay.highlightNode}
\usage{
Overlay.highlightNode(promise, highlightConfig, nodeId = NULL,
  backendNodeId = NULL, objectId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{highlightConfig}{A HighlightConfig.
A descriptor for the highlight appearance.}

\item{nodeId}{Optional. A DOM.NodeId.
Identifier of the node to highlight.}

\item{backendNodeId}{Optional. A DOM.BackendNodeId.
Identifier of the backend node to highlight.}

\item{objectId}{Optional. A Runtime.RemoteObjectId.
JavaScript object id of the node to be highlighted.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Highlights DOM node with given id or with the given JavaScript object wrapper. Either nodeId or
objectId must be specified.
}
