% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.highlightQuad}
\alias{Overlay.highlightQuad}
\title{Send the command Overlay.highlightQuad}
\usage{
Overlay.highlightQuad(promise, quad, color = NULL, outlineColor = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{quad}{A DOM.Quad.
Quad to highlight}

\item{color}{Optional. A DOM.RGBA.
The highlight fill color (default: transparent).}

\item{outlineColor}{Optional. A DOM.RGBA.
The highlight outline color (default: transparent).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Highlights given quad. Coordinates are absolute with respect to the main frame viewport.
}
