% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.highlightRect}
\alias{Overlay.highlightRect}
\title{Send the command Overlay.highlightRect

Highlights given rectangle. Coordinates are absolute with respect to the main frame viewport.}
\usage{
Overlay.highlightRect(promise, x, y, width, height, color = NULL,
  outlineColor = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{x}{An integer.
X coordinate}

\item{y}{An integer.
Y coordinate}

\item{width}{An integer.
Rectangle width}

\item{height}{An integer.
Rectangle height}

\item{color}{Optional. A DOM.RGBA.
The highlight fill color (default: transparent).}

\item{outlineColor}{Optional. A DOM.RGBA.
The highlight outline color (default: transparent).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Overlay.highlightRect

Highlights given rectangle. Coordinates are absolute with respect to the main frame viewport.
}
