% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.setInspectMode}
\alias{Overlay.setInspectMode}
\title{Send the command Overlay.setInspectMode}
\usage{
Overlay.setInspectMode(promise, mode, highlightConfig = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{mode}{A InspectMode.
Set an inspection mode.}

\item{highlightConfig}{Optional. A HighlightConfig.
A descriptor for the highlight appearance of hovered-over nodes. May be omitted if \code{enabled == false}.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enters the 'inspect' mode. In this mode, elements that user is hovering over are highlighted.
Backend then generates 'inspectNodeRequested' event upon element selection.
}
