% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.setPausedInDebuggerMessage}
\alias{Overlay.setPausedInDebuggerMessage}
\title{Send the command Overlay.setPausedInDebuggerMessage}
\usage{
Overlay.setPausedInDebuggerMessage(promise, message = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{message}{Optional. A character string.
The message to display, also triggers resume and step over controls.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Overlay.setPausedInDebuggerMessage
}
