% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.setShowDebugBorders}
\alias{Overlay.setShowDebugBorders}
\title{Send the command Overlay.setShowDebugBorders

Requests that backend shows debug borders on layers}
\usage{
Overlay.setShowDebugBorders(promise, show)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{show}{A logical.
True for showing debug borders}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Overlay.setShowDebugBorders

Requests that backend shows debug borders on layers
}
