% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.setShowFPSCounter}
\alias{Overlay.setShowFPSCounter}
\title{Send the command Overlay.setShowFPSCounter

Requests that backend shows the FPS counter}
\usage{
Overlay.setShowFPSCounter(promise, show)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{show}{A logical.
True for showing the FPS counter}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Overlay.setShowFPSCounter

Requests that backend shows the FPS counter
}
