% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.setShowPaintRects}
\alias{Overlay.setShowPaintRects}
\title{Send the command Overlay.setShowPaintRects

Requests that backend shows paint rectangles}
\usage{
Overlay.setShowPaintRects(promise, result)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{result}{A logical.
True for showing paint rectangles}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Overlay.setShowPaintRects

Requests that backend shows paint rectangles
}
