% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.setShowScrollBottleneckRects}
\alias{Overlay.setShowScrollBottleneckRects}
\title{Send the command Overlay.setShowScrollBottleneckRects

Requests that backend shows scroll bottleneck rects}
\usage{
Overlay.setShowScrollBottleneckRects(promise, show)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{show}{A logical.
True for showing scroll bottleneck rects}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Overlay.setShowScrollBottleneckRects

Requests that backend shows scroll bottleneck rects
}
