% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.setShowViewportSizeOnResize}
\alias{Overlay.setShowViewportSizeOnResize}
\title{Send the command Overlay.setShowViewportSizeOnResize}
\usage{
Overlay.setShowViewportSizeOnResize(promise, show)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{show}{A logical.
Whether to paint size or not.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Paints viewport size upon main frame resize.
}
