% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.setSuspended}
\alias{Overlay.setSuspended}
\title{Send the command Overlay.setSuspended}
\usage{
Overlay.setSuspended(promise, suspended)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{suspended}{A logical.
Whether overlay should be suspended and not consume any resources until resumed.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Overlay.setSuspended
}
