% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.addCompilationCache}
\alias{Page.addCompilationCache}
\title{Send the command Page.addCompilationCache}
\usage{
Page.addCompilationCache(promise, url, data)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{url}{A character string.}

\item{data}{A character string.
Base64-encoded data}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Seeds compilation cache for given url. Compilation cache does not survive
cross-process navigation.
}
