% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.addScriptToEvaluateOnNewDocument}
\alias{Page.addScriptToEvaluateOnNewDocument}
\title{Send the command Page.addScriptToEvaluateOnNewDocument}
\usage{
Page.addScriptToEvaluateOnNewDocument(promise, source, worldName = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{source}{A character string.}

\item{worldName}{Experimental. Optional. A character string.
If specified, creates an isolated world with the given name and evaluates given script in it.
This world name will be used as the ExecutionContextDescription::name when the corresponding
event is emitted.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Evaluates given script in every frame upon creation (before loading frame's scripts).
}
