% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.captureScreenshot}
\alias{Page.captureScreenshot}
\title{Send the command Page.captureScreenshot}
\usage{
Page.captureScreenshot(promise, format = NULL, quality = NULL,
  clip = NULL, fromSurface = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{format}{Optional. A character string.
Image compression format (defaults to png). Accepted values: jpeg, png.}

\item{quality}{Optional. An integer.
Compression quality from range \link{0..100} (jpeg only).}

\item{clip}{Optional. A Viewport.
Capture the screenshot of a given region only.}

\item{fromSurface}{Experimental. Optional. A logical.
Capture the screenshot from the surface, rather than the view. Defaults to true.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Capture page screenshot.
}
