% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.captureSnapshot}
\alias{Page.captureSnapshot}
\title{Send the command Page.captureSnapshot}
\usage{
Page.captureSnapshot(promise, format = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{format}{Optional. A character string.
Format (defaults to mhtml). Accepted values: mhtml.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns a snapshot of the page as a string. For MHTML format, the serialization includes
iframes, shadow DOM, external resources, and element-inline styles.
}
