% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.close}
\alias{Page.close}
\title{Send the command Page.close

Tries to close page, running its beforeunload hooks, if any.}
\usage{
Page.close(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Page.close

Tries to close page, running its beforeunload hooks, if any.
}
