% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.crash}
\alias{Page.crash}
\title{Send the command Page.crash}
\usage{
Page.crash(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Crashes renderer on the IO thread, generates minidumps.
}
