% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.createIsolatedWorld}
\alias{Page.createIsolatedWorld}
\title{Send the command Page.createIsolatedWorld

Creates an isolated world for the given frame.}
\usage{
Page.createIsolatedWorld(promise, frameId, worldName = NULL,
  grantUniveralAccess = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{frameId}{A FrameId.
Id of the frame in which the isolated world should be created.}

\item{worldName}{Optional. A character string.
An optional name which is reported in the Execution Context.}

\item{grantUniveralAccess}{Optional. A logical.
Whether or not universal access should be granted to the isolated world. This is a powerful
option, use with caution.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Page.createIsolatedWorld

Creates an isolated world for the given frame.
}
