% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.enable}
\alias{Page.enable}
\title{Send the command Page.enable}
\usage{
Page.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables page domain notifications.
}
