% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.getAppManifest}
\alias{Page.getAppManifest}
\title{Send the command Page.getAppManifest}
\usage{
Page.getAppManifest(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 3.
}
\description{
Send the command Page.getAppManifest
}
