% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.getLayoutMetrics}
\alias{Page.getLayoutMetrics}
\title{Send the command Page.getLayoutMetrics}
\usage{
Page.getLayoutMetrics(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 3.
}
\description{
Returns metrics relating to the layouting of the page, such as viewport bounds/scale.
}
