% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.getResourceContent}
\alias{Page.getResourceContent}
\title{Send the command Page.getResourceContent

Returns content of the given resource.}
\usage{
Page.getResourceContent(promise, frameId, url)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{frameId}{A FrameId.
Frame id to get resource for.}

\item{url}{A character string.
URL of the resource to get content for.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command Page.getResourceContent

Returns content of the given resource.
}
