% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.getResourceTree}
\alias{Page.getResourceTree}
\title{Send the command Page.getResourceTree}
\usage{
Page.getResourceTree(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns present frame / resource tree structure.
}
