% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.handleJavaScriptDialog}
\alias{Page.handleJavaScriptDialog}
\title{Send the command Page.handleJavaScriptDialog

Accepts or dismisses a JavaScript initiated dialog (alert, confirm, prompt, or onbeforeunload).}
\usage{
Page.handleJavaScriptDialog(promise, accept, promptText = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{accept}{A logical.
Whether to accept or dismiss the dialog.}

\item{promptText}{Optional. A character string.
The text to enter into the dialog prompt before accepting. Used only if this is a prompt
dialog.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Page.handleJavaScriptDialog

Accepts or dismisses a JavaScript initiated dialog (alert, confirm, prompt, or onbeforeunload).
}
