% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_events.R
\name{Page.javascriptDialogOpening}
\alias{Page.javascriptDialogOpening}
\title{Await the event Page.javascriptDialogOpening or create a callback}
\usage{
Page.javascriptDialogOpening(promise, url = NULL, message = NULL,
  type = NULL, hasBrowserHandler = NULL, defaultPrompt = NULL,
  .callback = NULL)
}
\arguments{
\item{promise}{An asynchronous result object.}

\item{url}{A character string.
Frame url. Accepted values: \code{~ .res$url} (to refer to the previous result).}

\item{message}{A character string.
Message that will be displayed by the dialog. Accepted values: \code{~ .res$message} (to refer to the previous result).}

\item{type}{A DialogType.
Dialog type. Accepted values: \code{~ .res$type} (to refer to the previous result).}

\item{hasBrowserHandler}{A logical.
True iff browser is capable showing or acting on the given dialog. When browser has no
dialog handler for given target, calling alert while Page domain is engaged will stall
the page execution. Execution can be resumed via calling Page.handleJavaScriptDialog. Accepted values: \code{~ .res$hasBrowserHandler} (to refer to the previous result).}

\item{defaultPrompt}{A character string.
Default dialog prompt. Accepted values: \code{~ .res$defaultPrompt} (to refer to the previous result).}

\item{.callback}{A callback function taking one argument. The object passed to
this function is the message received from Chrome: this is a named list
with an element \code{method} (that is equal to \code{"javascriptDialogOpening"})
and an element \code{params} which is a named list.
The \code{params} list is composed of
the following element(s): \code{url}, \code{message}, \code{type}, \code{hasBrowserHandler}, \code{defaultPrompt} (optional) .}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the use of a callback function.
When \code{.callback} is \code{NULL}, the promise is fulfilled when the event is received.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
\code{result} is a named list: its elements are the parameters sended by Chrome.
You can chain this promise with another command or event listener.
When \code{.callback} is not \code{NULL}, the promise is fulfilled as soon as the callback is created; the value
is a function without any argument that can be called to cancel the callback. When you use the
\code{.callback} argument, you cannot send the result to any other command or event listener.
}
\description{
\strong{Event description}:
Fired when a JavaScript initiated dialog (alert, confirm, prompt, or onbeforeunload) is about to
open.
}
