% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_events.R
\name{Page.lifecycleEvent}
\alias{Page.lifecycleEvent}
\title{Await the event Page.lifecycleEvent or create a callback}
\usage{
Page.lifecycleEvent(promise, frameId = NULL, loaderId = NULL,
  name = NULL, timestamp = NULL, .callback = NULL)
}
\arguments{
\item{promise}{An asynchronous result object.}

\item{frameId}{A FrameId.
Id of the frame. Accepted values: \code{~ .res$frameId} (to refer to the previous result).}

\item{loaderId}{A Network.LoaderId.
Loader identifier. Empty string if the request is fetched from worker. Accepted values: \code{~ .res$loaderId} (to refer to the previous result).}

\item{name}{A character string.
Accepted values: \code{~ .res$name} (to refer to the previous result).}

\item{timestamp}{A Network.MonotonicTime.
Accepted values: \code{~ .res$timestamp} (to refer to the previous result).}

\item{.callback}{A callback function taking one argument. The object passed to
this function is the message received from Chrome: this is a named list
with an element \code{method} (that is equal to \code{"lifecycleEvent"})
and an element \code{params} which is a named list.
The \code{params} list is composed of
the following element(s): \code{frameId}, \code{loaderId}, \code{name}, \code{timestamp}.}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the use of a callback function.
When \code{.callback} is \code{NULL}, the promise is fulfilled when the event is received.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
\code{result} is a named list: its elements are the parameters sended by Chrome.
You can chain this promise with another command or event listener.
When \code{.callback} is not \code{NULL}, the promise is fulfilled as soon as the callback is created; the value
is a function without any argument that can be called to cancel the callback. When you use the
\code{.callback} argument, you cannot send the result to any other command or event listener.
}
\description{
\strong{Event description}:
Fired for top level page lifecycle events such as navigation, load, paint, etc.
}
