% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.navigate}
\alias{Page.navigate}
\title{Send the command Page.navigate

Navigates current page to the given URL.}
\usage{
Page.navigate(promise, url, referrer = NULL, transitionType = NULL,
  frameId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{url}{A character string.
URL to navigate the page to.}

\item{referrer}{Optional. A character string.
Referrer URL.}

\item{transitionType}{Optional. A TransitionType.
Intended transition type.}

\item{frameId}{Optional. A FrameId.
Frame id to navigate, if not specified navigates the top frame.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 3.
}
\description{
Send the command Page.navigate

Navigates current page to the given URL.
}
