% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_events.R
\name{Page.navigatedWithinDocument}
\alias{Page.navigatedWithinDocument}
\title{Await the event Page.navigatedWithinDocument or create a callback}
\usage{
Page.navigatedWithinDocument(promise, frameId = NULL, url = NULL,
  .callback = NULL)
}
\arguments{
\item{promise}{An asynchronous result object.}

\item{frameId}{A FrameId.
Id of the frame. Accepted values: \code{~ .res$frameId} (to refer to the previous result).}

\item{url}{A character string.
Frame's new url. Accepted values: \code{~ .res$url} (to refer to the previous result).}

\item{.callback}{A callback function taking one argument. The object passed to
this function is the message received from Chrome: this is a named list
with an element \code{method} (that is equal to \code{"navigatedWithinDocument"})
and an element \code{params} which is a named list.
The \code{params} list is composed of
the following element(s): \code{frameId}, \code{url}.}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the use of a callback function.
When \code{.callback} is \code{NULL}, the promise is fulfilled when the event is received.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
\code{result} is a named list: its elements are the parameters sended by Chrome.
You can chain this promise with another command or event listener.
When \code{.callback} is not \code{NULL}, the promise is fulfilled as soon as the callback is created; the value
is a function without any argument that can be called to cancel the callback. When you use the
\code{.callback} argument, you cannot send the result to any other command or event listener.
}
\description{
\strong{Event description}:
Fired when same-document navigation happens, e.g. due to history API usage or anchor navigation.
}
