% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.navigateToHistoryEntry}
\alias{Page.navigateToHistoryEntry}
\title{Send the command Page.navigateToHistoryEntry

Navigates current page to the given history entry.}
\usage{
Page.navigateToHistoryEntry(promise, entryId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{entryId}{An integer.
Unique id of the entry to navigate to.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Page.navigateToHistoryEntry

Navigates current page to the given history entry.
}
