% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.reload}
\alias{Page.reload}
\title{Send the command Page.reload}
\usage{
Page.reload(promise, ignoreCache = NULL, scriptToEvaluateOnLoad = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{ignoreCache}{Optional. A logical.
If true, browser cache is ignored (as if the user pressed Shift+refresh).}

\item{scriptToEvaluateOnLoad}{Optional. A character string.
If set, the script will be injected into all frames of the inspected page after reload.
Argument will be ignored if reloading dataURL origin.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Reloads given page optionally ignoring the cache.
}
