% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.removeScriptToEvaluateOnNewDocument}
\alias{Page.removeScriptToEvaluateOnNewDocument}
\title{Send the command Page.removeScriptToEvaluateOnNewDocument}
\usage{
Page.removeScriptToEvaluateOnNewDocument(promise, identifier)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{identifier}{A ScriptIdentifier.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Removes given script from the list.
}
