% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.requestAppBanner}
\alias{Page.requestAppBanner}
\title{Send the command Page.requestAppBanner}
\usage{
Page.requestAppBanner(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Page.requestAppBanner
}
