% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.screencastFrameAck}
\alias{Page.screencastFrameAck}
\title{Send the command Page.screencastFrameAck

Acknowledges that a screencast frame has been received by the frontend.}
\usage{
Page.screencastFrameAck(promise, sessionId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{sessionId}{An integer.
Frame number.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Page.screencastFrameAck

Acknowledges that a screencast frame has been received by the frontend.
}
