% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.searchInResource}
\alias{Page.searchInResource}
\title{Send the command Page.searchInResource

Searches for given string in resource content.}
\usage{
Page.searchInResource(promise, frameId, url, query, caseSensitive = NULL,
  isRegex = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{frameId}{A FrameId.
Frame id for resource to search in.}

\item{url}{A character string.
URL of the resource to search in.}

\item{query}{A character string.
String to search for.}

\item{caseSensitive}{Optional. A logical.
If true, search is case sensitive.}

\item{isRegex}{Optional. A logical.
If true, treats string parameter as regex.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Page.searchInResource

Searches for given string in resource content.
}
