% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.setAdBlockingEnabled}
\alias{Page.setAdBlockingEnabled}
\title{Send the command Page.setAdBlockingEnabled}
\usage{
Page.setAdBlockingEnabled(promise, enabled, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{enabled}{A logical.
Whether to block ads.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Enable Chrome's experimental ad filter on all sites.
}
