% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.setBypassCSP}
\alias{Page.setBypassCSP}
\title{Send the command Page.setBypassCSP}
\usage{
Page.setBypassCSP(promise, enabled)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{enabled}{A logical.
Whether to bypass page CSP.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enable page Content Security Policy by-passing.
}
