% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.setDocumentContent}
\alias{Page.setDocumentContent}
\title{Send the command Page.setDocumentContent}
\usage{
Page.setDocumentContent(promise, frameId, html)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{frameId}{A FrameId.
Frame id to set HTML for.}

\item{html}{A character string.
HTML content to set.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sets given markup as the document's HTML.
}
