% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.setDownloadBehavior}
\alias{Page.setDownloadBehavior}
\title{Send the command Page.setDownloadBehavior}
\usage{
Page.setDownloadBehavior(promise, behavior, downloadPath = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{behavior}{A character string.
Whether to allow all or deny all download requests, or use default Chrome behavior if
available (otherwise deny). Accepted values: deny, allow, default.}

\item{downloadPath}{Optional. A character string.
The default path to save downloaded files to. This is requred if behavior is set to 'allow'}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Set the behavior when downloading a file.
}
