% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.setFontSizes}
\alias{Page.setFontSizes}
\title{Send the command Page.setFontSizes

Set default font sizes.}
\usage{
Page.setFontSizes(promise, fontSizes)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{fontSizes}{A FontSizes.
Specifies font sizes to set. If a font size is not specified, it won't be changed.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Page.setFontSizes

Set default font sizes.
}
