% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.setLifecycleEventsEnabled}
\alias{Page.setLifecycleEventsEnabled}
\title{Send the command Page.setLifecycleEventsEnabled}
\usage{
Page.setLifecycleEventsEnabled(promise, enabled)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{enabled}{A logical.
If true, starts emitting lifecycle events.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Controls whether page will emit lifecycle events.
}
