% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.setProduceCompilationCache}
\alias{Page.setProduceCompilationCache}
\title{Send the command Page.setProduceCompilationCache

Forces compilation cache to be generated for every subresource script.}
\usage{
Page.setProduceCompilationCache(promise, enabled)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{enabled}{A logical.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Page.setProduceCompilationCache

Forces compilation cache to be generated for every subresource script.
}
