% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.setWebLifecycleState}
\alias{Page.setWebLifecycleState}
\title{Send the command Page.setWebLifecycleState}
\usage{
Page.setWebLifecycleState(promise, state)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{state}{A character string.
Target lifecycle state Accepted values: frozen, active.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Tries to update the web lifecycle state of the page.
It will transition the page to the given state according to:
https://github.com/WICG/web-lifecycle/
}
