% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.startScreencast}
\alias{Page.startScreencast}
\title{Send the command Page.startScreencast}
\usage{
Page.startScreencast(promise, format = NULL, quality = NULL,
  maxWidth = NULL, maxHeight = NULL, everyNthFrame = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{format}{Optional. A character string.
Image compression format. Accepted values: jpeg, png.}

\item{quality}{Optional. An integer.
Compression quality from range \link{0..100}.}

\item{maxWidth}{Optional. An integer.
Maximum screenshot width.}

\item{maxHeight}{Optional. An integer.
Maximum screenshot height.}

\item{everyNthFrame}{Optional. An integer.
Send every n-th frame.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Starts sending each frame using the \code{screencastFrame} event.
}
