% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.stopScreencast}
\alias{Page.stopScreencast}
\title{Send the command Page.stopScreencast}
\usage{
Page.stopScreencast(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Stops sending each frame in the \code{screencastFrame}.
}
