% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_commands.R
\name{Performance.disable}
\alias{Performance.disable}
\title{Send the command Performance.disable

Disable collecting and reporting metrics.}
\usage{
Performance.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Performance.disable

Disable collecting and reporting metrics.
}
