% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_commands.R
\name{Performance.getMetrics}
\alias{Performance.getMetrics}
\title{Send the command Performance.getMetrics}
\usage{
Performance.getMetrics(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Retrieve current values of run-time metrics.
}
