% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_events.R
\name{Performance.metrics}
\alias{Performance.metrics}
\title{Await the event Performance.metrics or create a callback}
\usage{
Performance.metrics(promise, metrics = NULL, title = NULL,
  .callback = NULL)
}
\arguments{
\item{promise}{An asynchronous result object.}

\item{metrics}{A list of Metric.
Current values of the metrics. Accepted values: \code{~ .res$metrics} (to refer to the previous result).}

\item{title}{A character string.
Timestamp title. Accepted values: \code{~ .res$title} (to refer to the previous result).}

\item{.callback}{A callback function taking one argument. The object passed to
this function is the message received from Chrome: this is a named list
with an element \code{method} (that is equal to \code{"metrics"})
and an element \code{params} which is a named list.
The \code{params} list is composed of
the following element(s): \code{metrics}, \code{title}.}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the use of a callback function.
When \code{.callback} is \code{NULL}, the promise is fulfilled when the event is received.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
\code{result} is a named list: its elements are the parameters sended by Chrome.
You can chain this promise with another command or event listener.
When \code{.callback} is not \code{NULL}, the promise is fulfilled as soon as the callback is created; the value
is a function without any argument that can be called to cancel the callback. When you use the
\code{.callback} argument, you cannot send the result to any other command or event listener.
}
\description{
\strong{Event description}:
Current values of the metrics.
}
