% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance_commands.R
\name{Performance.setTimeDomain}
\alias{Performance.setTimeDomain}
\title{Send the command Performance.setTimeDomain}
\usage{
Performance.setTimeDomain(promise, timeDomain)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{timeDomain}{A character string.
Time domain Accepted values: timeTicks, threadTicks.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sets time domain to use for collecting and reporting duration metrics.
Note that this must be called before enabling metrics collection. Calling
this method while metrics collection is enabled returns an error.
}
