% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{\%...>\%}
\alias{\%...T>\%}
\alias{\%...!\%}
\alias{\%...T!\%}
\title{Pipe operators}
\usage{
lhs \%>\% rhs

lhs \%T>\% rhs

lhs \%...>\% rhs

lhs \%...T>\% rhs

lhs \%...!\% rhs

lhs \%...T!\% rhs
}
\description{
See \code{promises::\link[promises]{pipes}} for details.
}
\keyword{internal}
