% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profiler_events.R
\name{Profiler.consoleProfileFinished}
\alias{Profiler.consoleProfileFinished}
\title{Await the event Profiler.consoleProfileFinished or create a callback}
\usage{
Profiler.consoleProfileFinished(promise, id = NULL, location = NULL,
  profile = NULL, title = NULL, .callback = NULL)
}
\arguments{
\item{promise}{An asynchronous result object.}

\item{id}{A character string.
Accepted values: \code{~ .res$id} (to refer to the previous result).}

\item{location}{A Debugger.Location.
Location of console.profileEnd(). Accepted values: \code{~ .res$location} (to refer to the previous result).}

\item{profile}{A Profile.
Accepted values: \code{~ .res$profile} (to refer to the previous result).}

\item{title}{A character string.
Profile title passed as an argument to console.profile(). Accepted values: \code{~ .res$title} (to refer to the previous result).}

\item{.callback}{A callback function taking one argument. The object passed to
this function is the message received from Chrome: this is a named list
with an element \code{method} (that is equal to \code{"consoleProfileFinished"})
and an element \code{params} which is a named list.
The \code{params} list is composed of
the following element(s): \code{id}, \code{location}, \code{profile}, \code{title} (optional) .}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the use of a callback function.
When \code{.callback} is \code{NULL}, the promise is fulfilled when the event is received.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
\code{result} is a named list: its elements are the parameters sended by Chrome.
You can chain this promise with another command or event listener.
When \code{.callback} is not \code{NULL}, the promise is fulfilled as soon as the callback is created; the value
is a function without any argument that can be called to cancel the callback. When you use the
\code{.callback} argument, you cannot send the result to any other command or event listener.
}
\description{
\strong{Event description}:
}
