% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profiler_commands.R
\name{Profiler.getBestEffortCoverage}
\alias{Profiler.getBestEffortCoverage}
\title{Send the command Profiler.getBestEffortCoverage

Collect coverage data for the current isolate. The coverage data may be incomplete due to
garbage collection.}
\usage{
Profiler.getBestEffortCoverage(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Profiler.getBestEffortCoverage

Collect coverage data for the current isolate. The coverage data may be incomplete due to
garbage collection.
}
