% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profiler_commands.R
\name{Profiler.setSamplingInterval}
\alias{Profiler.setSamplingInterval}
\title{Send the command Profiler.setSamplingInterval}
\usage{
Profiler.setSamplingInterval(promise, interval)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{interval}{An integer.
New sampling interval in microseconds.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Changes CPU profiler sampling interval. Must be called before CPU profiles recording started.
}
