% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profiler_commands.R
\name{Profiler.stop}
\alias{Profiler.stop}
\title{Send the command Profiler.stop}
\usage{
Profiler.stop(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Profiler.stop
}
