% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profiler_commands.R
\name{Profiler.stopPreciseCoverage}
\alias{Profiler.stopPreciseCoverage}
\title{Send the command Profiler.stopPreciseCoverage

Disable precise code coverage. Disabling releases unnecessary execution count records and allows
executing optimized code.}
\usage{
Profiler.stopPreciseCoverage(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Profiler.stopPreciseCoverage

Disable precise code coverage. Disabling releases unnecessary execution count records and allows
executing optimized code.
}
