% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profiler_commands.R
\name{Profiler.stopTypeProfile}
\alias{Profiler.stopTypeProfile}
\title{Send the command Profiler.stopTypeProfile}
\usage{
Profiler.stopTypeProfile(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Disable type profile. Disabling releases type profile data collected so far.
}
