% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profiler_commands.R
\name{Profiler.takePreciseCoverage}
\alias{Profiler.takePreciseCoverage}
\title{Send the command Profiler.takePreciseCoverage}
\usage{
Profiler.takePreciseCoverage(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Collect coverage data for the current isolate, and resets execution counters. Precise code
coverage needs to have started.
}
