% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-promises.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{promise}
\alias{promise_all}
\alias{promise_map}
\alias{promise_race}
\alias{promise_reduce}
\alias{promise_reject}
\alias{promise_resolve}
\alias{then}
\alias{catch}
\alias{finally}
\alias{as.promise}
\alias{is.promise}
\alias{is.promising}
\alias{\%>\%}
\alias{\%T>\%}
\alias{\%...>\%}
\alias{\%...T>\%}
\alias{\%...!\%}
\alias{\%...T!\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{promises}{\code{\link[promises]{promise}}, \code{\link[promises]{promise_all}}, \code{\link[promises]{promise_map}}, \code{\link[promises]{promise_race}}, \code{\link[promises]{promise_reduce}}, \code{\link[promises]{promise_reject}}, \code{\link[promises]{promise_resolve}}, \code{\link[promises]{then}}, \code{\link[promises]{catch}}, \code{\link[promises]{finally}}, \code{\link[promises]{as.promise}}, \code{\link[promises]{is.promise}}, \code{\link[promises]{is.promising}}, \code{\link[promises]{\%>\%}}, \code{\link[promises]{\%T>\%}}, \code{\link[promises]{\%...>\%}}, \code{\link[promises]{\%...T>\%}}, \code{\link[promises]{\%...!\%}}, \code{\link[promises]{\%...T!\%}}}
}}

