% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.awaitPromise}
\alias{Runtime.awaitPromise}
\title{Send the command Runtime.awaitPromise

Add handler to promise with given promise object id.}
\usage{
Runtime.awaitPromise(promise, promiseObjectId, returnByValue = NULL,
  generatePreview = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{promiseObjectId}{A RemoteObjectId.
Identifier of the promise.}

\item{returnByValue}{Optional. A logical.
Whether the result is expected to be a JSON object that should be sent by value.}

\item{generatePreview}{Optional. A logical.
Whether preview should be generated for the result.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command Runtime.awaitPromise

Add handler to promise with given promise object id.
}
