% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.callFunctionOn}
\alias{Runtime.callFunctionOn}
\title{Send the command Runtime.callFunctionOn

Calls function with given declaration on the given object. Object group of the result is
inherited from the target object.}
\usage{
Runtime.callFunctionOn(promise, functionDeclaration, objectId = NULL,
  arguments = NULL, silent = NULL, returnByValue = NULL,
  generatePreview = NULL, userGesture = NULL, awaitPromise = NULL,
  executionContextId = NULL, objectGroup = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{functionDeclaration}{A character string.
Declaration of the function to call.}

\item{objectId}{Optional. A RemoteObjectId.
Identifier of the object to call function on. Either objectId or executionContextId should
be specified.}

\item{arguments}{Optional. A list of CallArgument.
Call arguments. All call arguments must belong to the same JavaScript world as the target
object.}

\item{silent}{Optional. A logical.
In silent mode exceptions thrown during evaluation are not reported and do not pause
execution. Overrides \code{setPauseOnException} state.}

\item{returnByValue}{Optional. A logical.
Whether the result is expected to be a JSON object which should be sent by value.}

\item{generatePreview}{Experimental. Optional. A logical.
Whether preview should be generated for the result.}

\item{userGesture}{Optional. A logical.
Whether execution should be treated as initiated by user in the UI.}

\item{awaitPromise}{Optional. A logical.
Whether execution should \code{await} for resulting value and return once awaited promise is
resolved.}

\item{executionContextId}{Optional. A ExecutionContextId.
Specifies execution context which global object will be used to call function on. Either
executionContextId or objectId should be specified.}

\item{objectGroup}{Optional. A character string.
Symbolic group name that can be used to release multiple objects. If objectGroup is not
specified and objectId is, objectGroup will be inherited from object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command Runtime.callFunctionOn

Calls function with given declaration on the given object. Object group of the result is
inherited from the target object.
}
