% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.compileScript}
\alias{Runtime.compileScript}
\title{Send the command Runtime.compileScript}
\usage{
Runtime.compileScript(promise, expression, sourceURL, persistScript,
  executionContextId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{expression}{A character string.
Expression to compile.}

\item{sourceURL}{A character string.
Source url to be set for the script.}

\item{persistScript}{A logical.
Specifies whether the compiled script should be persisted.}

\item{executionContextId}{Optional. A ExecutionContextId.
Specifies in which execution context to perform script run. If the parameter is omitted the
evaluation will be performed in the context of the inspected page.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Compiles expression.
}
