% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.disable}
\alias{Runtime.disable}
\title{Send the command Runtime.disable}
\usage{
Runtime.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Disables reporting of execution contexts creation.
}
