% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.discardConsoleEntries}
\alias{Runtime.discardConsoleEntries}
\title{Send the command Runtime.discardConsoleEntries

Discards collected exceptions and console API calls.}
\usage{
Runtime.discardConsoleEntries(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Runtime.discardConsoleEntries

Discards collected exceptions and console API calls.
}
