% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.enable}
\alias{Runtime.enable}
\title{Send the command Runtime.enable}
\usage{
Runtime.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables reporting of execution contexts creation by means of \code{executionContextCreated} event.
When the reporting gets enabled the event will be sent immediately for each existing execution
context.
}
