% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.getHeapUsage}
\alias{Runtime.getHeapUsage}
\title{Send the command Runtime.getHeapUsage

Returns the JavaScript heap usage.
It is the total usage of the corresponding isolate not scoped to a particular Runtime.}
\usage{
Runtime.getHeapUsage(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command Runtime.getHeapUsage

Returns the JavaScript heap usage.
It is the total usage of the corresponding isolate not scoped to a particular Runtime.
}
