% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.getIsolateId}
\alias{Runtime.getIsolateId}
\title{Send the command Runtime.getIsolateId

Returns the isolate id.}
\usage{
Runtime.getIsolateId(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Runtime.getIsolateId

Returns the isolate id.
}
