% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.getProperties}
\alias{Runtime.getProperties}
\title{Send the command Runtime.getProperties

Returns properties of a given object. Object group of the result is inherited from the target
object.}
\usage{
Runtime.getProperties(promise, objectId, ownProperties = NULL,
  accessorPropertiesOnly = NULL, generatePreview = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{objectId}{A RemoteObjectId.
Identifier of the object to return properties for.}

\item{ownProperties}{Optional. A logical.
If true, returns properties belonging only to the element itself, not to its prototype
chain.}

\item{accessorPropertiesOnly}{Experimental. Optional. A logical.
If true, returns accessor properties (with getter/setter) only; internal properties are not
returned either.}

\item{generatePreview}{Experimental. Optional. A logical.
Whether preview should be generated for the results.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 3.
}
\description{
Send the command Runtime.getProperties

Returns properties of a given object. Object group of the result is inherited from the target
object.
}
