% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.globalLexicalScopeNames}
\alias{Runtime.globalLexicalScopeNames}
\title{Send the command Runtime.globalLexicalScopeNames}
\usage{
Runtime.globalLexicalScopeNames(promise, executionContextId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{executionContextId}{Optional. A ExecutionContextId.
Specifies in which execution context to lookup global scope variables.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns all let, const and class variables from global scope.
}
