% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.queryObjects}
\alias{Runtime.queryObjects}
\title{Send the command Runtime.queryObjects}
\usage{
Runtime.queryObjects(promise, prototypeObjectId, objectGroup = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{prototypeObjectId}{A RemoteObjectId.
Identifier of the prototype to return objects for.}

\item{objectGroup}{Optional. A character string.
Symbolic group name that can be used to release the results.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Runtime.queryObjects
}
