% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.releaseObjectGroup}
\alias{Runtime.releaseObjectGroup}
\title{Send the command Runtime.releaseObjectGroup}
\usage{
Runtime.releaseObjectGroup(promise, objectGroup)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{objectGroup}{A character string.
Symbolic object group name.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Releases all remote objects that belong to a given group.
}
