% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.runIfWaitingForDebugger}
\alias{Runtime.runIfWaitingForDebugger}
\title{Send the command Runtime.runIfWaitingForDebugger}
\usage{
Runtime.runIfWaitingForDebugger(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Tells inspected instance to run if it was waiting for debugger to attach.
}
