% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.runScript}
\alias{Runtime.runScript}
\title{Send the command Runtime.runScript

Runs script with given id in a given context.}
\usage{
Runtime.runScript(promise, scriptId, executionContextId = NULL,
  objectGroup = NULL, silent = NULL, includeCommandLineAPI = NULL,
  returnByValue = NULL, generatePreview = NULL, awaitPromise = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{scriptId}{A ScriptId.
Id of the script to run.}

\item{executionContextId}{Optional. A ExecutionContextId.
Specifies in which execution context to perform script run. If the parameter is omitted the
evaluation will be performed in the context of the inspected page.}

\item{objectGroup}{Optional. A character string.
Symbolic group name that can be used to release multiple objects.}

\item{silent}{Optional. A logical.
In silent mode exceptions thrown during evaluation are not reported and do not pause
execution. Overrides \code{setPauseOnException} state.}

\item{includeCommandLineAPI}{Optional. A logical.
Determines whether Command Line API should be available during the evaluation.}

\item{returnByValue}{Optional. A logical.
Whether the result is expected to be a JSON object which should be sent by value.}

\item{generatePreview}{Optional. A logical.
Whether preview should be generated for the result.}

\item{awaitPromise}{Optional. A logical.
Whether execution should \code{await} for resulting value and return once awaited promise is
resolved.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command Runtime.runScript

Runs script with given id in a given context.
}
