% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.setAsyncCallStackDepth}
\alias{Runtime.setAsyncCallStackDepth}
\title{Send the command Runtime.setAsyncCallStackDepth

Enables or disables async call stacks tracking.}
\usage{
Runtime.setAsyncCallStackDepth(promise, maxDepth)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{maxDepth}{An integer.
Maximum depth of async call stacks. Setting to \code{0} will effectively disable collecting async
call stacks (default).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Runtime.setAsyncCallStackDepth

Enables or disables async call stacks tracking.
}
