% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.setCustomObjectFormatterEnabled}
\alias{Runtime.setCustomObjectFormatterEnabled}
\title{Send the command Runtime.setCustomObjectFormatterEnabled}
\usage{
Runtime.setCustomObjectFormatterEnabled(promise, enabled)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{enabled}{A logical.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Runtime.setCustomObjectFormatterEnabled
}
