% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.setMaxCallStackSizeToCapture}
\alias{Runtime.setMaxCallStackSizeToCapture}
\title{Send the command Runtime.setMaxCallStackSizeToCapture}
\usage{
Runtime.setMaxCallStackSizeToCapture(promise, size)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{size}{An integer.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Runtime.setMaxCallStackSizeToCapture
}
