% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.terminateExecution}
\alias{Runtime.terminateExecution}
\title{Send the command Runtime.terminateExecution}
\usage{
Runtime.terminateExecution(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Terminate current or next JavaScript execution.
Will cancel the termination when the outer-most script execution ends.
}
