% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schema_commands.R
\name{Schema.getDomains}
\alias{Schema.getDomains}
\title{Send the command Schema.getDomains}
\usage{
Schema.getDomains(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns supported domains.
}
