% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_Security.R
\name{Security.certificateError}
\alias{Security.certificateError}
\title{Await the event Security.certificateError or create a callback}
\usage{
Security.certificateError(promise, eventId = NULL, errorType = NULL,
  requestURL = NULL, .callback = NULL)
}
\arguments{
\item{promise}{An asynchronous result object.}

\item{eventId}{An integer.
The ID of the event. Accepted values: \code{~ .res$eventId} (to refer to the previous result).}

\item{errorType}{A character string.
The type of the error. Accepted values: \code{~ .res$errorType} (to refer to the previous result).}

\item{requestURL}{A character string.
The url that was requested. Accepted values: \code{~ .res$requestURL} (to refer to the previous result).}

\item{.callback}{A callback function taking one argument. The object passed to
this function is the message received from Chrome: this is a named list
with an element \code{method} (that is equal to \code{"certificateError"})
and an element \code{params} which is a named list.
The \code{params} list is composed of
the following element(s): \code{eventId}, \code{errorType}, \code{requestURL}.}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the use of a callback function.
When \code{.callback} is \code{NULL}, the promise is fulfilled when the event is received.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
\code{result} is a named list: its elements are the parameters sended by Chrome.
You can chain this promise with another command or event listener.
When \code{.callback} is not \code{NULL}, the promise is fulfilled as soon as the callback is created; the value
is a function without any argument that can be called to cancel the callback. When you use the
\code{.callback} argument, you cannot send the result to any other command or event listener.
}
\description{
\strong{Event description}:
There is a certificate error. If overriding certificate errors is enabled, then it should be
handled with the \code{handleCertificateError} command. Note: this event does not fire if the
certificate error has been allowed internally. Only one client per target should override
certificate errors at the same time.
}
