% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Security_commands.R
\name{Security.disable}
\alias{Security.disable}
\title{Send the command Security.disable}
\usage{
Security.disable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Disables tracking security state changes.
}
