% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Security_commands.R
\name{Security.enable}
\alias{Security.enable}
\title{Send the command Security.enable}
\usage{
Security.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables tracking security state changes.
}
