% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Security_commands.R
\name{Security.setIgnoreCertificateErrors}
\alias{Security.setIgnoreCertificateErrors}
\title{Send the command Security.setIgnoreCertificateErrors}
\usage{
Security.setIgnoreCertificateErrors(promise, ignore)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{ignore}{A logical.
If true, all certificate errors will be ignored.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enable/disable whether all certificate errors should be ignored.
}
