% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServiceWorker_commands.R
\name{ServiceWorker.deliverPushMessage}
\alias{ServiceWorker.deliverPushMessage}
\title{Send the command ServiceWorker.deliverPushMessage}
\usage{
ServiceWorker.deliverPushMessage(promise, origin, registrationId, data)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{origin}{A character string.}

\item{registrationId}{A character string.}

\item{data}{A character string.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command ServiceWorker.deliverPushMessage
}
