% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServiceWorker_commands.R
\name{ServiceWorker.dispatchSyncEvent}
\alias{ServiceWorker.dispatchSyncEvent}
\title{Send the command ServiceWorker.dispatchSyncEvent}
\usage{
ServiceWorker.dispatchSyncEvent(promise, origin, registrationId, tag,
  lastChance)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{origin}{A character string.}

\item{registrationId}{A character string.}

\item{tag}{A character string.}

\item{lastChance}{A logical.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command ServiceWorker.dispatchSyncEvent
}
