% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServiceWorker_commands.R
\name{ServiceWorker.setForceUpdateOnPageLoad}
\alias{ServiceWorker.setForceUpdateOnPageLoad}
\title{Send the command ServiceWorker.setForceUpdateOnPageLoad}
\usage{
ServiceWorker.setForceUpdateOnPageLoad(promise, forceUpdateOnPageLoad)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{forceUpdateOnPageLoad}{A logical.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command ServiceWorker.setForceUpdateOnPageLoad
}
