% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServiceWorker_commands.R
\name{ServiceWorker.unregister}
\alias{ServiceWorker.unregister}
\title{Send the command ServiceWorker.unregister}
\usage{
ServiceWorker.unregister(promise, scopeURL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{scopeURL}{A character string.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command ServiceWorker.unregister
}
