% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage_commands.R
\name{Storage.clearDataForOrigin}
\alias{Storage.clearDataForOrigin}
\title{Send the command Storage.clearDataForOrigin}
\usage{
Storage.clearDataForOrigin(promise, origin, storageTypes)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{origin}{A character string.
Security origin.}

\item{storageTypes}{A character string.
Comma separated origin names.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Clears storage for origin.
}
