% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage_commands.R
\name{Storage.getUsageAndQuota}
\alias{Storage.getUsageAndQuota}
\title{Send the command Storage.getUsageAndQuota}
\usage{
Storage.getUsageAndQuota(promise, origin)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{origin}{A character string.
Security origin.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 3.
}
\description{
Returns usage and quota in bytes.
}
