% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage_commands.R
\name{Storage.trackCacheStorageForOrigin}
\alias{Storage.trackCacheStorageForOrigin}
\title{Send the command Storage.trackCacheStorageForOrigin}
\usage{
Storage.trackCacheStorageForOrigin(promise, origin)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{origin}{A character string.
Security origin.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Registers origin to be notified when an update occurs to its cache storage list.
}
