% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage_commands.R
\name{Storage.untrackIndexedDBForOrigin}
\alias{Storage.untrackIndexedDBForOrigin}
\title{Send the command Storage.untrackIndexedDBForOrigin

Unregisters origin from receiving notifications for IndexedDB.}
\usage{
Storage.untrackIndexedDBForOrigin(promise, origin)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{origin}{A character string.
Security origin.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Storage.untrackIndexedDBForOrigin

Unregisters origin from receiving notifications for IndexedDB.
}
