% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemInfo_commands.R
\name{SystemInfo.getInfo}
\alias{SystemInfo.getInfo}
\title{Send the command SystemInfo.getInfo}
\usage{
SystemInfo.getInfo(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 4.
}
\description{
Returns information about the system.
}
