% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemInfo_commands.R
\name{SystemInfo.getProcessInfo}
\alias{SystemInfo.getProcessInfo}
\title{Send the command SystemInfo.getProcessInfo}
\usage{
SystemInfo.getProcessInfo(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns information about all running processes.
}
