% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.activateTarget}
\alias{Target.activateTarget}
\title{Send the command Target.activateTarget

Activates (focuses) the target.}
\usage{
Target.activateTarget(promise, targetId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{targetId}{A TargetID.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Target.activateTarget

Activates (focuses) the target.
}
