% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.attachToBrowserTarget}
\alias{Target.attachToBrowserTarget}
\title{Send the command Target.attachToBrowserTarget

Attaches to the browser target, only uses flat sessionId mode.}
\usage{
Target.attachToBrowserTarget(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Target.attachToBrowserTarget

Attaches to the browser target, only uses flat sessionId mode.
}
