% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.attachToTarget}
\alias{Target.attachToTarget}
\title{Send the command Target.attachToTarget

Attaches to the target with given id.}
\usage{
Target.attachToTarget(promise, targetId, flatten = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{targetId}{A TargetID.}

\item{flatten}{Experimental. Optional. A logical.
Enables "flat" access to the session via specifying sessionId attribute in the commands.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Target.attachToTarget

Attaches to the target with given id.
}
