% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.createBrowserContext}
\alias{Target.createBrowserContext}
\title{Send the command Target.createBrowserContext}
\usage{
Target.createBrowserContext(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Creates a new empty BrowserContext. Similar to an incognito profile but you can have more than
one.
}
