% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_Target.R
\name{Target.detachedFromTarget}
\alias{Target.detachedFromTarget}
\title{Await the event Target.detachedFromTarget or create a callback}
\usage{
Target.detachedFromTarget(promise, sessionId = NULL, targetId = NULL,
  .callback = NULL)
}
\arguments{
\item{promise}{An asynchronous result object.}

\item{sessionId}{A SessionID.
Detached session identifier. Accepted values: \code{~ .res$sessionId} (to refer to the previous result).}

\item{targetId}{Deprecated. A TargetID.
Deprecated. Accepted values: \code{~ .res$targetId} (to refer to the previous result).}

\item{.callback}{A callback function taking one argument. The object passed to
this function is the message received from Chrome: this is a named list
with an element \code{method} (that is equal to \code{"detachedFromTarget"})
and an element \code{params} which is a named list.
The \code{params} list is composed of
the following element(s): \code{sessionId}, \code{targetId} (optional) .}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the use of a callback function.
When \code{.callback} is \code{NULL}, the promise is fulfilled when the event is received.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
\code{result} is a named list: its elements are the parameters sended by Chrome.
You can chain this promise with another command or event listener.
When \code{.callback} is not \code{NULL}, the promise is fulfilled as soon as the callback is created; the value
is a function without any argument that can be called to cancel the callback. When you use the
\code{.callback} argument, you cannot send the result to any other command or event listener.
}
\description{
\strong{Event description}:
Issued when detached from target for any reason (including \code{detachFromTarget} command). Can be
issued multiple times per target if multiple sessions have been attached to it.
}
