% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.detachFromTarget}
\alias{Target.detachFromTarget}
\title{Send the command Target.detachFromTarget}
\usage{
Target.detachFromTarget(promise, sessionId = NULL, targetId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{sessionId}{Optional. A SessionID.
Session to detach.}

\item{targetId}{Deprecated. Optional. A TargetID.
Deprecated.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Detaches session with given id.
}
