% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.getTargetInfo}
\alias{Target.getTargetInfo}
\title{Send the command Target.getTargetInfo}
\usage{
Target.getTargetInfo(promise, targetId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{targetId}{Optional. A TargetID.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns information about a target.
}
