% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.getTargets}
\alias{Target.getTargets}
\title{Send the command Target.getTargets

Retrieves a list of available targets.}
\usage{
Target.getTargets(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Target.getTargets

Retrieves a list of available targets.
}
