% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.sendMessageToTarget}
\alias{Target.sendMessageToTarget}
\title{Send the command Target.sendMessageToTarget}
\usage{
Target.sendMessageToTarget(promise, message, sessionId = NULL,
  targetId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{message}{A character string.}

\item{sessionId}{Optional. A SessionID.
Identifier of the session.}

\item{targetId}{Deprecated. Optional. A TargetID.
Deprecated.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Sends protocol message over session with given id.
}
