% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.setAutoAttach}
\alias{Target.setAutoAttach}
\title{Send the command Target.setAutoAttach

Controls whether to automatically attach to new targets which are considered to be related to
this one. When turned on, attaches to all existing related targets as well. When turned off,
automatically detaches from all currently attached targets.}
\usage{
Target.setAutoAttach(promise, autoAttach, waitForDebuggerOnStart,
  flatten = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{autoAttach}{A logical.
Whether to auto-attach to related targets.}

\item{waitForDebuggerOnStart}{A logical.
Whether to pause new targets when attaching to them. Use \code{Runtime.runIfWaitingForDebugger}
to run paused targets.}

\item{flatten}{Experimental. Optional. A logical.
Enables "flat" access to the session via specifying sessionId attribute in the commands.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Target.setAutoAttach

Controls whether to automatically attach to new targets which are considered to be related to
this one. When turned on, attaches to all existing related targets as well. When turned off,
automatically detaches from all currently attached targets.
}
