% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.setDiscoverTargets}
\alias{Target.setDiscoverTargets}
\title{Send the command Target.setDiscoverTargets

Controls whether to discover available targets and notify via
\code{targetCreated/targetInfoChanged/targetDestroyed} events.}
\usage{
Target.setDiscoverTargets(promise, discover)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{discover}{A logical.
Whether to discover available targets.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Target.setDiscoverTargets

Controls whether to discover available targets and notify via
\code{targetCreated/targetInfoChanged/targetDestroyed} events.
}
