% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.setRemoteLocations}
\alias{Target.setRemoteLocations}
\title{Send the command Target.setRemoteLocations}
\usage{
Target.setRemoteLocations(promise, locations)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{locations}{A list of RemoteLocation.
List of remote locations.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables target discovery for the specified locations, when \code{setDiscoverTargets} was set to
\code{true}.
}
