% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing_commands.R
\name{Testing.generateTestReport}
\alias{Testing.generateTestReport}
\title{Send the command Testing.generateTestReport

Generates a report for testing.}
\usage{
Testing.generateTestReport(promise, message, group = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{message}{A character string.
Message to be displayed in the report.}

\item{group}{Optional. A character string.
Specifies the endpoint group to deliver the report to.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Testing.generateTestReport

Generates a report for testing.
}
