% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tethering_commands.R
\name{Tethering.bind}
\alias{Tethering.bind}
\title{Send the command Tethering.bind}
\usage{
Tethering.bind(promise, port)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{port}{An integer.
Port number to bind.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Request browser port binding.
}
