% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracing_commands.R
\name{Tracing.end}
\alias{Tracing.end}
\title{Send the command Tracing.end}
\usage{
Tracing.end(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Stop trace events collection.
}
