% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracing_commands.R
\name{Tracing.getCategories}
\alias{Tracing.getCategories}
\title{Send the command Tracing.getCategories

Gets supported tracing categories.}
\usage{
Tracing.getCategories(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Tracing.getCategories

Gets supported tracing categories.
}
