% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracing_commands.R
\name{Tracing.recordClockSyncMarker}
\alias{Tracing.recordClockSyncMarker}
\title{Send the command Tracing.recordClockSyncMarker}
\usage{
Tracing.recordClockSyncMarker(promise, syncId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{syncId}{A character string.
The ID of this clock sync marker}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Record a clock sync marker in the trace.
}
