% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracing_commands.R
\name{Tracing.requestMemoryDump}
\alias{Tracing.requestMemoryDump}
\title{Send the command Tracing.requestMemoryDump}
\usage{
Tracing.requestMemoryDump(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Request a global memory dump.
}
