% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracing_commands.R
\name{Tracing.start}
\alias{Tracing.start}
\title{Send the command Tracing.start

Start trace events collection.}
\usage{
Tracing.start(promise, categories = NULL, options = NULL,
  bufferUsageReportingInterval = NULL, transferMode = NULL,
  streamCompression = NULL, traceConfig = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{categories}{Deprecated. Optional. A character string.
Category/tag filter}

\item{options}{Deprecated. Optional. A character string.
Tracing options}

\item{bufferUsageReportingInterval}{Optional. A numeric.
If set, the agent will issue bufferUsage events at this interval, specified in milliseconds}

\item{transferMode}{Optional. A character string.
Whether to report trace events as series of dataCollected events or to save trace to a
stream (defaults to \code{ReportEvents}). Accepted values: ReportEvents, ReturnAsStream.}

\item{streamCompression}{Optional. A StreamCompression.
Compression format to use. This only applies when using \code{ReturnAsStream}
transfer mode (defaults to \code{none})}

\item{traceConfig}{Optional. A TraceConfig.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Tracing.start

Start trace events collection.
}
