% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait.R
\name{wait}
\alias{wait}
\title{Wait to resolve a promise}
\usage{
wait(promise, delay = 0)
}
\arguments{
\item{promise}{A promise.}

\item{delay}{Number of seconds before resolving the promise.}
}
\value{
A promise. The value of the promise is the same of the lhs promise.
}
\description{
This is a helper function to add a waiting time after a promise resolution.
It is built on top of the \code{promises::then()} function. So, it can be used with
the \code{magrittr} pipe operators. The value of the resolved promise is the value
of the lhs: it can be used with \code{\%>\%} or \code{\%T>\%} with the same result.
}
\examples{
library(promises)

pr <- promise_resolve(runif(1))

pr \%>\%
  wait(2) \%>\%
  then(~cat(., "\\n"))
}
