% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{write_base64}
\alias{write_base64}
\title{write base64 result to file}
\usage{
write_base64(res, con)
}
\arguments{
\item{res}{A resulting list object with some \code{data} field containing base64
encoded data. (like \href{https://chromedevtools.github.io/devtools-protocol/tot/Page#method-printToPDF}{Page$printToPDF()})}

\item{con}{A \link[base:connections]{connection} object. See \link[base:readBin]{base::writeBin()}.}
}
\description{
This is a helper function to write some resulting data from chrome to file.
It will:
\itemize{
\item decode the base64 encoded raw data
\item write to file using \link[base:readBin]{base::writeBin()}
It will be useful for function like \href{https://chromedevtools.github.io/devtools-protocol/tot/Page#method-printToPDF}{Page$printToPDF()}
}
}
