% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessibility_commands.R
\name{Accessibility.enable}
\alias{Accessibility.enable}
\title{Send the command Accessibility.enable}
\usage{
Accessibility.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables the accessibility domain which causes \code{AXNodeId}s to remain consistent between method calls.
This turns on accessibility for the page, which can impact performance until accessibility is disabled.
}
