% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessibility_commands.R
\name{Accessibility.getPartialAXTree}
\alias{Accessibility.getPartialAXTree}
\title{Send the command Accessibility.getPartialAXTree}
\usage{
Accessibility.getPartialAXTree(promise, nodeId = NULL,
  backendNodeId = NULL, objectId = NULL, fetchRelatives = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{nodeId}{Optional. A DOM.NodeId.
Identifier of the node to get the partial accessibility tree for.}

\item{backendNodeId}{Optional. A DOM.BackendNodeId.
Identifier of the backend node to get the partial accessibility tree for.}

\item{objectId}{Optional. A Runtime.RemoteObjectId.
JavaScript object id of the node wrapper to get the partial accessibility tree for.}

\item{fetchRelatives}{Optional. A logical.
Whether to fetch this nodes ancestors, siblings and children. Defaults to true.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Fetches the accessibility node and partial accessibility tree for this DOM node, if it exists.
}
