% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animation_commands.R
\name{Animation.getCurrentTime}
\alias{Animation.getCurrentTime}
\title{Send the command Animation.getCurrentTime

Returns the current time of the an animation.}
\usage{
Animation.getCurrentTime(promise, id)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{id}{A character string.
Id of animation.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Animation.getCurrentTime

Returns the current time of the an animation.
}
