% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Audits.R
\name{Audits.getEncodedResponse}
\alias{Audits.getEncodedResponse}
\title{Send the command Audits.getEncodedResponse}
\usage{
Audits.getEncodedResponse(promise, requestId, encoding, quality = NULL,
  sizeOnly = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{requestId}{A Network.RequestId.
Identifier of the network request to get content for.}

\item{encoding}{A character string.
The encoding to use. Accepted values: webp, jpeg, png.}

\item{quality}{Optional. A numeric.
The quality of the encoding (0-1). (defaults to 1)}

\item{sizeOnly}{Optional. A logical.
Whether to only return the size information (defaults to false).}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 3.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Returns the response body and size if it were re-encoded with the specified settings. Only
applies to images.
}
