% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.getWindowForTarget}
\alias{Browser.getWindowForTarget}
\title{Send the command Browser.getWindowForTarget}
\usage{
Browser.getWindowForTarget(promise, targetId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{targetId}{Optional. A Target.TargetID.
Devtools agent host id. If called as a part of the session, associated targetId is used.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Get the browser window that contains the devtools target.
}
