% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browser_commands.R
\name{Browser.setWindowBounds}
\alias{Browser.setWindowBounds}
\title{Send the command Browser.setWindowBounds}
\usage{
Browser.setWindowBounds(promise, windowId, bounds, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{windowId}{A WindowID.
Browser window id.}

\item{bounds}{A Bounds.
New window bounds. The 'minimized', 'maximized' and 'fullscreen' states cannot be combined
with 'left', 'top', 'width' or 'height'. Leaves unspecified fields unchanged.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Set position and/or size of the browser window.
}
