% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_connect.R
\name{chr_connect}
\alias{chr_connect}
\title{Connect R to a new Chrome instance}
\usage{
chr_connect(bin = "google-chrome", debug_port = 9222,
  extra_args = NULL, headless = TRUE, work_dir = "remote-profile")
}
\arguments{
\item{bin}{Command or path to the Chrome exe.}

\item{debug_port}{Port of the Chrome instance.}

\item{extra_args}{Extra arguments passed to command line.}

\item{headless}{Is Chrome launch in headless mode?}

\item{work_dir}{Chrome working directory.}
}
\value{
An asynchronous result object that can be passed to any command or event listeners.
}
\description{
Connect R to a new Chrome instance
}
