% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_CSS.R
\name{CSS.enable}
\alias{CSS.enable}
\title{Send the command CSS.enable}
\usage{
CSS.enable(promise, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Enables the CSS agent for the given page. Clients should not assume that the CSS agent has been
enabled until the result of this command is received.
}
