% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_CSS.R
\name{CSS.forcePseudoState}
\alias{CSS.forcePseudoState}
\title{Send the command CSS.forcePseudoState}
\usage{
CSS.forcePseudoState(promise, nodeId, forcedPseudoClasses,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{nodeId}{A DOM.NodeId.
The element id for which to force the pseudo state.}

\item{forcedPseudoClasses}{A list of string.
Element pseudo classes to force when computing the element's style.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Ensures that the given node will have specified pseudo-classes whenever its style is computed by
the browser.
}
