% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.continueToLocation}
\alias{Debugger.continueToLocation}
\title{Send the command Debugger.continueToLocation}
\usage{
Debugger.continueToLocation(promise, location, targetCallFrames = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{location}{A Location.
Location to continue to.}

\item{targetCallFrames}{Optional. A character string.
Accepted values: any, current.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Continues execution until specific location is reached.
}
