% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.getPossibleBreakpoints}
\alias{Debugger.getPossibleBreakpoints}
\title{Send the command Debugger.getPossibleBreakpoints

Returns possible locations for breakpoint. scriptId in start and end range locations should be
the same.}
\usage{
Debugger.getPossibleBreakpoints(promise, start, end = NULL,
  restrictToFunction = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{start}{A Location.
Start of range to search possible breakpoint locations in.}

\item{end}{Optional. A Location.
End of range to search possible breakpoint locations in (excluding). When not specified, end
of scripts is used as end of range.}

\item{restrictToFunction}{Optional. A logical.
Only consider locations which are in the same (non-nested) function as start.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Debugger.getPossibleBreakpoints

Returns possible locations for breakpoint. scriptId in start and end range locations should be
the same.
}
