% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.getStackTrace}
\alias{Debugger.getStackTrace}
\title{Send the command Debugger.getStackTrace

Returns stack trace with given \code{stackTraceId}.}
\usage{
Debugger.getStackTrace(promise, stackTraceId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{stackTraceId}{A Runtime.StackTraceId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Debugger.getStackTrace

Returns stack trace with given \code{stackTraceId}.
}
