% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Debugger.R
\name{Debugger.setAsyncCallStackDepth}
\alias{Debugger.setAsyncCallStackDepth}
\title{Send the command Debugger.setAsyncCallStackDepth}
\usage{
Debugger.setAsyncCallStackDepth(promise, maxDepth, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{maxDepth}{An integer.
Maximum depth of async call stacks. Setting to \code{0} will effectively disable collecting async
call stacks (default).}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Enables or disables async call stacks tracking.
}
