% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Debugger.R
\name{Debugger.setBlackboxPatterns}
\alias{Debugger.setBlackboxPatterns}
\title{Send the command Debugger.setBlackboxPatterns}
\usage{
Debugger.setBlackboxPatterns(promise, patterns, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{patterns}{A list of string.
Array of regexps that will be used to check script url for blackbox state.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Replace previous blackbox patterns with passed ones. Forces backend to skip stepping/pausing in
scripts with url matching one of the patterns. VM will try to leave blackboxed script by
performing 'step in' several times, finally resorting to 'step out' if unsuccessful.
}
