% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Debugger.R
\name{Debugger.setBreakpointByUrl}
\alias{Debugger.setBreakpointByUrl}
\title{Send the command Debugger.setBreakpointByUrl}
\usage{
Debugger.setBreakpointByUrl(promise, lineNumber, url = NULL,
  urlRegex = NULL, scriptHash = NULL, columnNumber = NULL,
  condition = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{lineNumber}{An integer.
Line number to set breakpoint at.}

\item{url}{Optional. A character string.
URL of the resources to set breakpoint on.}

\item{urlRegex}{Optional. A character string.
Regex pattern for the URLs of the resources to set breakpoints on. Either \code{url} or
\code{urlRegex} must be specified.}

\item{scriptHash}{Optional. A character string.
Script hash of the resources to set breakpoint on.}

\item{columnNumber}{Optional. An integer.
Offset in the line to set breakpoint at.}

\item{condition}{Optional. A character string.
Expression to use as a breakpoint condition. When specified, debugger will only stop on the
breakpoint if this expression evaluates to true.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 2.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Sets JavaScript breakpoint at given location specified either by URL or URL regex. Once this
command is issued, all existing parsed scripts will have breakpoints resolved and returned in
\code{locations} property. Further matching script parsing will result in subsequent
\code{breakpointResolved} events issued. This logical breakpoint will survive page reloads.
}
