% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.setBreakpointOnFunctionCall}
\alias{Debugger.setBreakpointOnFunctionCall}
\title{Send the command Debugger.setBreakpointOnFunctionCall

Sets JavaScript breakpoint before each call to the given function.
If another function was created from the same source as a given one,
calling it will also trigger the breakpoint.}
\usage{
Debugger.setBreakpointOnFunctionCall(promise, objectId, condition = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{objectId}{A Runtime.RemoteObjectId.
Function object id.}

\item{condition}{Optional. A character string.
Expression to use as a breakpoint condition. When specified, debugger will
stop on the breakpoint if this expression evaluates to true.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Debugger.setBreakpointOnFunctionCall

Sets JavaScript breakpoint before each call to the given function.
If another function was created from the same source as a given one,
calling it will also trigger the breakpoint.
}
