% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugger_commands.R
\name{Debugger.setVariableValue}
\alias{Debugger.setVariableValue}
\title{Send the command Debugger.setVariableValue}
\usage{
Debugger.setVariableValue(promise, scopeNumber, variableName, newValue,
  callFrameId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{scopeNumber}{An integer.
0-based number of scope as was listed in scope chain. Only 'local', 'closure' and 'catch'
scope types are allowed. Other scopes could be manipulated manually.}

\item{variableName}{A character string.
Variable name.}

\item{newValue}{A Runtime.CallArgument.
New variable value.}

\item{callFrameId}{A CallFrameId.
Id of callframe that holds variable.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Changes value of variable in a callframe. Object-based scopes are not supported and must be
mutated manually.
}
