% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeviceOrientation_commands.R
\name{DeviceOrientation.setDeviceOrientationOverride}
\alias{DeviceOrientation.setDeviceOrientationOverride}
\title{Send the command DeviceOrientation.setDeviceOrientationOverride}
\usage{
DeviceOrientation.setDeviceOrientationOverride(promise, alpha, beta, gamma)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{alpha}{A numeric.
Mock alpha}

\item{beta}{A numeric.
Mock beta}

\item{gamma}{A numeric.
Mock gamma}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Overrides the Device Orientation.
}
