% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.copyTo}
\alias{DOM.copyTo}
\title{Send the command DOM.copyTo}
\usage{
DOM.copyTo(promise, nodeId, targetNodeId, insertBeforeNodeId = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{nodeId}{A NodeId.
Id of the node to copy.}

\item{targetNodeId}{A NodeId.
Id of the element to drop the copy into.}

\item{insertBeforeNodeId}{Optional. A NodeId.
Drop the copy before this node (if absent, the copy becomes the last child of
\code{targetNodeId}).}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Creates a deep copy of the specified node and places it into the target container before the
given anchor.
}
