% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.getNodeForLocation}
\alias{DOM.getNodeForLocation}
\title{Send the command DOM.getNodeForLocation

Returns node id at given location. Depending on whether DOM domain is enabled, nodeId is
either returned or not.}
\usage{
DOM.getNodeForLocation(promise, x, y, includeUserAgentShadowDOM = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{x}{An integer.
X coordinate.}

\item{y}{An integer.
Y coordinate.}

\item{includeUserAgentShadowDOM}{Optional. A logical.
False to skip to the nearest non-UA shadow root ancestor (default: false).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command DOM.getNodeForLocation

Returns node id at given location. Depending on whether DOM domain is enabled, nodeId is
either returned or not.
}
