% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.performSearch}
\alias{DOM.performSearch}
\title{Send the command DOM.performSearch

Searches for a given string in the DOM tree. Use \code{getSearchResults} to access search results or
\code{cancelSearch} to end this search session.}
\usage{
DOM.performSearch(promise, query, includeUserAgentShadowDOM = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{query}{A character string.
Plain text or query selector or XPath search query.}

\item{includeUserAgentShadowDOM}{Optional. A logical.
True to search in user agent shadow DOM.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command DOM.performSearch

Searches for a given string in the DOM tree. Use \code{getSearchResults} to access search results or
\code{cancelSearch} to end this search session.
}
