% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.pushNodesByBackendIdsToFrontend}
\alias{DOM.pushNodesByBackendIdsToFrontend}
\title{Send the command DOM.pushNodesByBackendIdsToFrontend}
\usage{
DOM.pushNodesByBackendIdsToFrontend(promise, backendNodeIds)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{backendNodeIds}{A list of BackendNodeId.
The array of backend node ids.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Requests that a batch of nodes is sent to the caller given their backend node ids.
}
