% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.requestChildNodes}
\alias{DOM.requestChildNodes}
\title{Send the command DOM.requestChildNodes

Requests that children of the node with given id are returned to the caller in form of
\code{setChildNodes} events where not only immediate children are retrieved, but all children down to
the specified depth.}
\usage{
DOM.requestChildNodes(promise, nodeId, depth = NULL, pierce = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the node to get children for.}

\item{depth}{Optional. An integer.
The maximum depth at which children should be retrieved, defaults to 1. Use -1 for the
entire subtree or provide an integer larger than 0.}

\item{pierce}{Optional. A logical.
Whether or not iframes and shadow roots should be traversed when returning the sub-tree
(default is false).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOM.requestChildNodes

Requests that children of the node with given id are returned to the caller in form of
\code{setChildNodes} events where not only immediate children are retrieved, but all children down to
the specified depth.
}
