% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.setAttributesAsText}
\alias{DOM.setAttributesAsText}
\title{Send the command DOM.setAttributesAsText

Sets attributes on element with given id. This method is useful when user edits some existing
attribute value and types in several attribute name/value pairs.}
\usage{
DOM.setAttributesAsText(promise, nodeId, text, name = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{nodeId}{A NodeId.
Id of the element to set attributes for.}

\item{text}{A character string.
Text with a number of attributes. Will parse this text using HTML parser.}

\item{name}{Optional. A character string.
Attribute name to replace with new attributes derived from text in case text parsed
successfully.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOM.setAttributesAsText

Sets attributes on element with given id. This method is useful when user edits some existing
attribute value and types in several attribute name/value pairs.
}
