% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.setFileInputFiles}
\alias{DOM.setFileInputFiles}
\title{Send the command DOM.setFileInputFiles

Sets files for the given file input element.}
\usage{
DOM.setFileInputFiles(promise, files, nodeId = NULL,
  backendNodeId = NULL, objectId = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{files}{A list of string.
Array of file paths to set.}

\item{nodeId}{Optional. A NodeId.
Identifier of the node.}

\item{backendNodeId}{Optional. A BackendNodeId.
Identifier of the backend node.}

\item{objectId}{Optional. A Runtime.RemoteObjectId.
JavaScript object id of the node wrapper.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command DOM.setFileInputFiles

Sets files for the given file input element.
}
