% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOM_commands.R
\name{DOM.setNodeValue}
\alias{DOM.setNodeValue}
\title{Send the command DOM.setNodeValue}
\usage{
DOM.setNodeValue(promise, nodeId, value, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{nodeId}{A NodeId.
Id of the node to set value for.}

\item{value}{A character string.
New node's value.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Sets node value for a node with given id.
}
