% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_DOMDebugger.R
\name{DOMDebugger.getEventListeners}
\alias{DOMDebugger.getEventListeners}
\title{Send the command DOMDebugger.getEventListeners}
\usage{
DOMDebugger.getEventListeners(promise, objectId, depth = NULL,
  pierce = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{objectId}{A Runtime.RemoteObjectId.
Identifier of the object to return listeners for.}

\item{depth}{Optional. An integer.
The maximum depth at which Node children should be retrieved, defaults to 1. Use -1 for the
entire subtree or provide an integer larger than 0.}

\item{pierce}{Optional. A logical.
Whether or not iframes and shadow roots should be traversed when returning the subtree
(default is false). Reports listeners for all contexts if pierce is enabled.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Returns event listeners of the given object.
}
