% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setDeviceMetricsOverride}
\alias{Emulation.setDeviceMetricsOverride}
\title{Send the command Emulation.setDeviceMetricsOverride}
\usage{
Emulation.setDeviceMetricsOverride(promise, width, height,
  deviceScaleFactor, mobile, scale = NULL, screenWidth = NULL,
  screenHeight = NULL, positionX = NULL, positionY = NULL,
  dontSetVisibleSize = NULL, screenOrientation = NULL,
  viewport = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{width}{An integer.
Overriding width value in pixels (minimum 0, maximum 10000000). 0 disables the override.}

\item{height}{An integer.
Overriding height value in pixels (minimum 0, maximum 10000000). 0 disables the override.}

\item{deviceScaleFactor}{A numeric.
Overriding device scale factor value. 0 disables the override.}

\item{mobile}{A logical.
Whether to emulate mobile device. This includes viewport meta tag, overlay scrollbars, text
autosizing and more.}

\item{scale}{Experimental. Optional. A numeric.
Scale to apply to resulting view image.}

\item{screenWidth}{Experimental. Optional. An integer.
Overriding screen width value in pixels (minimum 0, maximum 10000000).}

\item{screenHeight}{Experimental. Optional. An integer.
Overriding screen height value in pixels (minimum 0, maximum 10000000).}

\item{positionX}{Experimental. Optional. An integer.
Overriding view X position on screen in pixels (minimum 0, maximum 10000000).}

\item{positionY}{Experimental. Optional. An integer.
Overriding view Y position on screen in pixels (minimum 0, maximum 10000000).}

\item{dontSetVisibleSize}{Experimental. Optional. A logical.
Do not set visible view size, rely upon explicit setVisibleSize call.}

\item{screenOrientation}{Optional. A ScreenOrientation.
Screen orientation override.}

\item{viewport}{Experimental. Optional. A Page.Viewport.
If set, the visible area of the page will be overridden to this viewport. This viewport
change is not observed by the page, e.g. viewport-relative elements do not change positions.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Overrides the values of device screen dimensions (window.screen.width, window.screen.height,
window.innerWidth, window.innerHeight, and "device-width"/"device-height"-related CSS media
query results).
}
