% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulation_commands.R
\name{Emulation.setGeolocationOverride}
\alias{Emulation.setGeolocationOverride}
\title{Send the command Emulation.setGeolocationOverride

Overrides the Geolocation Position or Error. Omitting any of the parameters emulates position
unavailable.}
\usage{
Emulation.setGeolocationOverride(promise, latitude = NULL,
  longitude = NULL, accuracy = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{latitude}{Optional. A numeric.
Mock latitude}

\item{longitude}{Optional. A numeric.
Mock longitude}

\item{accuracy}{Optional. A numeric.
Mock accuracy}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Emulation.setGeolocationOverride

Overrides the Geolocation Position or Error. Omitting any of the parameters emulates position
unavailable.
}
