% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Emulation.R
\name{Emulation.setUserAgentOverride}
\alias{Emulation.setUserAgentOverride}
\title{Send the command Emulation.setUserAgentOverride}
\usage{
Emulation.setUserAgentOverride(promise, userAgent, acceptLanguage = NULL,
  platform = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{userAgent}{A character string.
User agent to use.}

\item{acceptLanguage}{Optional. A character string.
Browser langugage to emulate.}

\item{platform}{Optional. A character string.
The platform navigator.platform should return.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Allows overriding user agent with the given string.
}
