% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Emulation.R
\name{Emulation.setVirtualTimePolicy}
\alias{Emulation.setVirtualTimePolicy}
\title{Send the command Emulation.setVirtualTimePolicy}
\usage{
Emulation.setVirtualTimePolicy(promise, policy, budget = NULL,
  maxVirtualTimeTaskStarvationCount = NULL, waitForNavigation = NULL,
  initialVirtualTime = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{policy}{A VirtualTimePolicy.}

\item{budget}{Optional. A numeric.
If set, after this many virtual milliseconds have elapsed virtual time will be paused and a
virtualTimeBudgetExpired event is sent.}

\item{maxVirtualTimeTaskStarvationCount}{Optional. An integer.
If set this specifies the maximum number of tasks that can be run before virtual is forced
forwards to prevent deadlock.}

\item{waitForNavigation}{Optional. A logical.
If set the virtual time policy change should be deferred until any frame starts navigating.
Note any previous deferred policy change is superseded.}

\item{initialVirtualTime}{Optional. A Network.TimeSinceEpoch.
If set, base::Time::Now will be overriden to initially return this value.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Turns on virtual time for all frames (replacing real-time with a synthetic time source) and sets
the current virtual time policy.  Note this supersedes any previous time budget.
}
