% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventemitters.R
\docType{data}
\name{EventEmitter}
\alias{EventEmitter}
\title{R6 class to emit some event}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventEmitter
}
\description{
This is a general purpose class to build on. It is inspired by node.js EventEmitter class.
See \link{https://nodejs.org/api/events.html}.
}
\examples{
myEmitter <- EventEmitter$new()
myEmitter$on('event',
    function() {
        message("an event occured")
    }
)
myEmitter$emit('event')
# Listener is still listening
myEmitter$emit('event')
myEmitter$once('uniqueevent',
    function() {
        message("an event occured - only once")
    }
)
# the listerner will be removed after first emission of the event
myEmitter$emit('uniqueevent')
# nothing happens as there is no more listener
myEmitter$emit('uniqueevent')

# an error in one listener will throws an error in R
myEmitter$on("eventwitherror", function(...) stop("An error in an event"))
# throws an error
\dontrun{
  myEmitter$emit("eventwitherror")
}
# you can catch an error with a special "error" event that always
# be emits when an error occured inside an eventEmitters
myEmitter$on("error", function(e) cat(conditionMessage(e)))
myEmitter$emit("eventwitherror")

# As in node.js class, a newListener event is emitted before each new event registration.
# newListener expect a function with eventName and callback argument
myEmitter <- EventEmitter$new()
# use once here to avoir infinite recursion#'
myEmitter$once("newListener", function(eventName, callback) {
   if(eventName == "event") {
      myEmitter$on("event", function(...) cat("B"))
   }
})
myEmitter$on("event", function(...) cat("A"))
myEmitter$emit("event") # BA
# newListener has been unregistered and is no more called
myEmitter$on("event", function(...) cat("C"))
myEmitter$emit("event") # BAC

# get the number of listeners for an event
myEmitter$listenerCount('event')

# get the event names which have been registered
myEmitter$eventNames()
}
\keyword{datasets}
