% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Fetch.R
\name{Fetch.fulfillRequest}
\alias{Fetch.fulfillRequest}
\title{Send the command Fetch.fulfillRequest}
\usage{
Fetch.fulfillRequest(promise, requestId, responseCode, responseHeaders,
  body = NULL, responsePhrase = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{requestId}{A RequestId.
An id the client received in requestPaused event.}

\item{responseCode}{An integer.
An HTTP response code.}

\item{responseHeaders}{A list of HeaderEntry.
Response headers.}

\item{body}{Optional. A character string.
A response body.}

\item{responsePhrase}{Optional. A character string.
A textual representation of responseCode.
If absent, a standard phrase mathcing responseCode is used.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Provides response to the request.
}
