% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch_commands.R
\name{Fetch.getResponseBody}
\alias{Fetch.getResponseBody}
\title{Send the command Fetch.getResponseBody}
\usage{
Fetch.getResponseBody(promise, requestId, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{requestId}{A RequestId.
Identifier for the intercepted request to get body for.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 2.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Causes the body of the response to be received from the server and
returned as a single string. May only be issued for a request that
is paused in the Response stage and is mutually exclusive with
takeResponseBodyForInterceptionAsStream. Calling other methods that
affect the request or disabling fetch domain before body is received
results in an undefined behavior.
}
