% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch_commands.R
\name{Fetch.takeResponseBodyAsStream}
\alias{Fetch.takeResponseBodyAsStream}
\title{Send the command Fetch.takeResponseBodyAsStream}
\usage{
Fetch.takeResponseBodyAsStream(promise, requestId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{requestId}{A RequestId.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Returns a handle to the stream representing the response body.
The request must be paused in the HeadersReceived stage.
Note that after this command the request can't be continued
as is -- client either needs to cancel it or to provide the
response body.
The stream only supports sequential read, IO.read will fail if the position
is specified.
This method is mutually exclusive with getResponseBody.
Calling other methods that affect the request or disabling fetch
domain before body is received results in an undefined behavior.
}
