% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeadlessExperimental_commands.R
\name{HeadlessExperimental.beginFrame}
\alias{HeadlessExperimental.beginFrame}
\title{Send the command HeadlessExperimental.beginFrame

Sends a BeginFrame to the target and returns when the frame was completed. Optionally captures a
screenshot from the resulting frame. Requires that the target was created with enabled
BeginFrameControl. Designed for use with --run-all-compositor-stages-before-draw, see also
https://goo.gl/3zHXhB for more background.}
\usage{
HeadlessExperimental.beginFrame(promise, frameTimeTicks = NULL,
  interval = NULL, noDisplayUpdates = NULL, screenshot = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{frameTimeTicks}{Optional. A numeric.
Timestamp of this BeginFrame in Renderer TimeTicks (milliseconds of uptime). If not set,
the current time will be used.}

\item{interval}{Optional. A numeric.
The interval between BeginFrames that is reported to the compositor, in milliseconds.
Defaults to a 60 frames/second interval, i.e. about 16.666 milliseconds.}

\item{noDisplayUpdates}{Optional. A logical.
Whether updates should not be committed and drawn onto the display. False by default. If
true, only side effects of the BeginFrame will be run, such as layout and animations, but
any visual updates may not be visible on the display or in screenshots.}

\item{screenshot}{Optional. A ScreenshotParams.
If set, a screenshot of the frame will be captured and returned in the response. Otherwise,
no screenshot will be captured. Note that capturing a screenshot can fail, for example,
during renderer initialization. In such a case, no screenshot data will be returned.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command HeadlessExperimental.beginFrame

Sends a BeginFrame to the target and returns when the frame was completed. Optionally captures a
screenshot from the resulting frame. Requires that the target was created with enabled
BeginFrameControl. Designed for use with --run-all-compositor-stages-before-draw, see also
https://goo.gl/3zHXhB for more background.
}
