% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapProfiler_commands.R
\name{HeapProfiler.addInspectedHeapObject}
\alias{HeapProfiler.addInspectedHeapObject}
\title{Send the command HeapProfiler.addInspectedHeapObject}
\usage{
HeapProfiler.addInspectedHeapObject(promise, heapObjectId,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{heapObjectId}{A HeapSnapshotObjectId.
Heap snapshot object id to be accessible by means of $x command line API.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Enables console to refer to the node with given id via $x (see Command Line API for more details
$x functions).
}
