% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapProfiler_commands.R
\name{HeapProfiler.getObjectByHeapObjectId}
\alias{HeapProfiler.getObjectByHeapObjectId}
\title{Send the command HeapProfiler.getObjectByHeapObjectId}
\usage{
HeapProfiler.getObjectByHeapObjectId(promise, objectId,
  objectGroup = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{objectId}{A HeapSnapshotObjectId.}

\item{objectGroup}{Optional. A character string.
Symbolic group name that can be used to release multiple objects.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Send the command HeapProfiler.getObjectByHeapObjectId
}
