% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapProfiler_commands.R
\name{HeapProfiler.stopTrackingHeapObjects}
\alias{HeapProfiler.stopTrackingHeapObjects}
\title{Send the command HeapProfiler.stopTrackingHeapObjects}
\usage{
HeapProfiler.stopTrackingHeapObjects(promise, reportProgress = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{reportProgress}{Optional. A logical.
If true 'reportHeapSnapshotProgress' events will be generated while snapshot is being taken
when the tracking is stopped.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command HeapProfiler.stopTrackingHeapObjects
}
