% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_methods.R
\name{http-methods}
\alias{http-methods}
\alias{fetch_version}
\alias{list_targets}
\alias{fetch_protocol}
\alias{new_tab}
\alias{activate_target}
\alias{close_target}
\alias{inspect_target}
\title{Chrome Debugging Protocol HTTP methods}
\usage{
fetch_version(host = "localhost", port = 9222, secure = FALSE)

list_targets(host = "localhost", port = 9222, secure = FALSE)

fetch_protocol(host = "localhost", port = 9222, secure = FALSE,
  local = FALSE)

new_tab(host = "localhost", port = 9222, secure = FALSE,
  url = NULL)

activate_target(host = "localhost", port = 9222, secure = FALSE,
  target_id)

close_target(host = "localhost", port = 9222, secure = FALSE,
  target_id)

inspect_target(host = "localhost", port = 9222, secure = FALSE,
  target_id)
}
\arguments{
\item{host}{Character scalar, the host name of the application.}

\item{port}{The remote debugging port (a numeric or a character scalar).}

\item{secure}{A logical indicating whether a secure protocol shall be used.}

\item{local}{Logical scalar, indicating whether the local version of the
protocol (embedded in \code{crrri}) must be used or the protocol must be
fetched \emph{remotely}.}

\item{url}{URL to open in a new tab.}

\item{target_id}{Target (or tab) identifier.}
}
\value{
\code{fetch_version()}, \code{list_targets()}, \code{fetch_protocols()} and
\code{new_tab()} return a list. \code{activate_target()} and \code{close_target()} returns
a logical: \code{TRUE} is returned when the command succeeds, \code{FALSE} is
returned if a wrong \code{target_id} is provided.
}
\description{
When Chromium/Chrome is started in debugging mode, several HTTP endpoints
are available. Using these endpoint, one can retrieve information from
Chromium/Chrome or send commands, see
\url{https://chromedevtools.github.io/devtools-protocol/#endpoints}.
}
\examples{
\donttest{
chrome <- Chrome$new()

fetch_version()
# fetch the Chromium/Chrome protocol
protocol <- fetch_protocol()

# get the list of opened tabs
list_targets()
# open an url in a new tab
new_target <- new_tab(url = "http://r-project.org")
# the new tab is referenced in the list of opened tabs
list_targets()
# inspect a target in a web browser
inspect_target(target_id = new_target$id)

chrome$close()
}
}
\keyword{internal}
