% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedDB_commands.R
\name{IndexedDB.deleteObjectStoreEntries}
\alias{IndexedDB.deleteObjectStoreEntries}
\title{Send the command IndexedDB.deleteObjectStoreEntries}
\usage{
IndexedDB.deleteObjectStoreEntries(promise, securityOrigin, databaseName,
  objectStoreName, keyRange)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{securityOrigin}{A character string.}

\item{databaseName}{A character string.}

\item{objectStoreName}{A character string.}

\item{keyRange}{A KeyRange.
Range of entry keys to delete}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Delete a range of entries from an object store
}
