% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedDB_commands.R
\name{IndexedDB.requestDatabaseNames}
\alias{IndexedDB.requestDatabaseNames}
\title{Send the command IndexedDB.requestDatabaseNames

Requests database names for given security origin.}
\usage{
IndexedDB.requestDatabaseNames(promise, securityOrigin)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{securityOrigin}{A character string.
Security origin.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command IndexedDB.requestDatabaseNames

Requests database names for given security origin.
}
