% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.dispatchKeyEvent}
\alias{Input.dispatchKeyEvent}
\title{Send the command Input.dispatchKeyEvent}
\usage{
Input.dispatchKeyEvent(promise, type, modifiers = NULL,
  timestamp = NULL, text = NULL, unmodifiedText = NULL,
  keyIdentifier = NULL, code = NULL, key = NULL,
  windowsVirtualKeyCode = NULL, nativeVirtualKeyCode = NULL,
  autoRepeat = NULL, isKeypad = NULL, isSystemKey = NULL,
  location = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{type}{A character string.
Type of the key event. Accepted values: keyDown, keyUp, rawKeyDown, char.}

\item{modifiers}{Optional. An integer.
Bit field representing pressed modifier keys. Alt=1, Ctrl=2, Meta/Command=4, Shift=8
(default: 0).}

\item{timestamp}{Optional. A TimeSinceEpoch.
Time at which the event occurred.}

\item{text}{Optional. A character string.
Text as generated by processing a virtual key code with a keyboard layout. Not needed for
for \code{keyUp} and \code{rawKeyDown} events (default: "")}

\item{unmodifiedText}{Optional. A character string.
Text that would have been generated by the keyboard if no modifiers were pressed (except for
shift). Useful for shortcut (accelerator) key handling (default: "").}

\item{keyIdentifier}{Optional. A character string.
Unique key identifier (e.g., 'U+0041') (default: "").}

\item{code}{Optional. A character string.
Unique DOM defined string value for each physical key (e.g., 'KeyA') (default: "").}

\item{key}{Optional. A character string.
Unique DOM defined string value describing the meaning of the key in the context of active
modifiers, keyboard layout, etc (e.g., 'AltGr') (default: "").}

\item{windowsVirtualKeyCode}{Optional. An integer.
Windows virtual key code (default: 0).}

\item{nativeVirtualKeyCode}{Optional. An integer.
Native virtual key code (default: 0).}

\item{autoRepeat}{Optional. A logical.
Whether the event was generated from auto repeat (default: false).}

\item{isKeypad}{Optional. A logical.
Whether the event was generated from the keypad (default: false).}

\item{isSystemKey}{Optional. A logical.
Whether the event was a system key event (default: false).}

\item{location}{Optional. An integer.
Whether the event was from the left or right side of the keyboard. 1=Left, 2=Right (default:
0).}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Dispatches a key event to the page.
}
