% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.dispatchTouchEvent}
\alias{Input.dispatchTouchEvent}
\title{Send the command Input.dispatchTouchEvent}
\usage{
Input.dispatchTouchEvent(promise, type, touchPoints, modifiers = NULL,
  timestamp = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{type}{A character string.
Type of the touch event. TouchEnd and TouchCancel must not contain any touch points, while
TouchStart and TouchMove must contains at least one. Accepted values: touchStart, touchEnd, touchMove, touchCancel.}

\item{touchPoints}{A list of TouchPoint.
Active touch points on the touch device. One event per any changed point (compared to
previous touch event in a sequence) is generated, emulating pressing/moving/releasing points
one by one.}

\item{modifiers}{Optional. An integer.
Bit field representing pressed modifier keys. Alt=1, Ctrl=2, Meta/Command=4, Shift=8
(default: 0).}

\item{timestamp}{Optional. A TimeSinceEpoch.
Time at which the event occurred.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Dispatches a touch event to the page.
}
