% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.synthesizePinchGesture}
\alias{Input.synthesizePinchGesture}
\title{Send the command Input.synthesizePinchGesture}
\usage{
Input.synthesizePinchGesture(promise, x, y, scaleFactor,
  relativeSpeed = NULL, gestureSourceType = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{x}{A numeric.
X coordinate of the start of the gesture in CSS pixels.}

\item{y}{A numeric.
Y coordinate of the start of the gesture in CSS pixels.}

\item{scaleFactor}{A numeric.
Relative scale factor after zooming (>1.0 zooms in, <1.0 zooms out).}

\item{relativeSpeed}{Optional. An integer.
Relative pointer speed in pixels per second (default: 800).}

\item{gestureSourceType}{Optional. A GestureSourceType.
Which type of input events to be generated (default: 'default', which queries the platform
for the preferred input type).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Synthesizes a pinch gesture over a time period by issuing appropriate touch events.
}
