% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_commands.R
\name{Input.synthesizeScrollGesture}
\alias{Input.synthesizeScrollGesture}
\title{Send the command Input.synthesizeScrollGesture}
\usage{
Input.synthesizeScrollGesture(promise, x, y, xDistance = NULL,
  yDistance = NULL, xOverscroll = NULL, yOverscroll = NULL,
  preventFling = NULL, speed = NULL, gestureSourceType = NULL,
  repeatCount = NULL, repeatDelayMs = NULL,
  interactionMarkerName = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{x}{A numeric.
X coordinate of the start of the gesture in CSS pixels.}

\item{y}{A numeric.
Y coordinate of the start of the gesture in CSS pixels.}

\item{xDistance}{Optional. A numeric.
The distance to scroll along the X axis (positive to scroll left).}

\item{yDistance}{Optional. A numeric.
The distance to scroll along the Y axis (positive to scroll up).}

\item{xOverscroll}{Optional. A numeric.
The number of additional pixels to scroll back along the X axis, in addition to the given
distance.}

\item{yOverscroll}{Optional. A numeric.
The number of additional pixels to scroll back along the Y axis, in addition to the given
distance.}

\item{preventFling}{Optional. A logical.
Prevent fling (default: true).}

\item{speed}{Optional. An integer.
Swipe speed in pixels per second (default: 800).}

\item{gestureSourceType}{Optional. A GestureSourceType.
Which type of input events to be generated (default: 'default', which queries the platform
for the preferred input type).}

\item{repeatCount}{Optional. An integer.
The number of times to repeat the gesture (default: 0).}

\item{repeatDelayMs}{Optional. An integer.
The number of milliseconds delay between each repeat. (default: 250).}

\item{interactionMarkerName}{Optional. A character string.
The name of the interaction markers to generate, if not empty (default: "").}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Synthesizes a scroll gesture over a time period by issuing appropriate touch events.
}
