% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inspector_commands.R
\name{Inspector.enable}
\alias{Inspector.enable}
\title{Send the command Inspector.enable

Enables inspector domain notifications.}
\usage{
Inspector.enable(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Inspector.enable

Enables inspector domain notifications.
}
