% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_commands.R
\name{IO.read}
\alias{IO.read}
\title{Send the command IO.read}
\usage{
IO.read(promise, handle, offset = NULL, size = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{handle}{A StreamHandle.
Handle of the stream to read.}

\item{offset}{Optional. An integer.
Seek to the specified offset before reading (if not specificed, proceed with offset
following the last read). Some types of streams may only support sequential reads.}

\item{size}{Optional. An integer.
Maximum number of bytes to read (left upon the agent discretion if not specified).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 3.
}
\description{
Read a chunk of the stream
}
