% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerTree_commands.R
\name{LayerTree.compositingReasons}
\alias{LayerTree.compositingReasons}
\title{Send the command LayerTree.compositingReasons}
\usage{
LayerTree.compositingReasons(promise, layerId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{layerId}{A LayerId.
The id of the layer for which we want to get the reasons it was composited.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Provides the reasons why the given layer was composited.
}
