% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_LayerTree.R
\name{LayerTree.profileSnapshot}
\alias{LayerTree.profileSnapshot}
\title{Send the command LayerTree.profileSnapshot}
\usage{
LayerTree.profileSnapshot(promise, snapshotId, minRepeatCount = NULL,
  minDuration = NULL, clipRect = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{snapshotId}{A SnapshotId.
The id of the layer snapshot.}

\item{minRepeatCount}{Optional. An integer.
The maximum number of times to replay the snapshot (1, if not specified).}

\item{minDuration}{Optional. A numeric.
The minimum duration (in seconds) to replay the snapshot.}

\item{clipRect}{Optional. A DOM.Rect.
The clip rectangle to apply when replaying the snapshot.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Send the command LayerTree.profileSnapshot
}
