% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerTree_commands.R
\name{LayerTree.replaySnapshot}
\alias{LayerTree.replaySnapshot}
\title{Send the command LayerTree.replaySnapshot}
\usage{
LayerTree.replaySnapshot(promise, snapshotId, fromStep = NULL,
  toStep = NULL, scale = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{snapshotId}{A SnapshotId.
The id of the layer snapshot.}

\item{fromStep}{Optional. An integer.
The first step to replay from (replay from the very start if not specified).}

\item{toStep}{Optional. An integer.
The last step to replay to (replay till the end if not specified).}

\item{scale}{Optional. A numeric.
The scale to apply while replaying (defaults to 1).}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Replays the layer snapshot and returns the resulting bitmap.
}
