% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory_commands.R
\name{Memory.getSamplingProfile}
\alias{Memory.getSamplingProfile}
\title{Send the command Memory.getSamplingProfile}
\usage{
Memory.getSamplingProfile(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Retrieve native memory allocations profile collected since last
\code{startSampling} call.
}
