% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Memory_commands.R
\name{Memory.startSampling}
\alias{Memory.startSampling}
\title{Send the command Memory.startSampling}
\usage{
Memory.startSampling(promise, samplingInterval = NULL,
  suppressRandomness = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{samplingInterval}{Optional. An integer.
Average number of bytes between samples.}

\item{suppressRandomness}{Optional. A logical.
Do not randomize intervals between samples.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Start collecting native memory profile.
}
