% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.continueInterceptedRequest}
\alias{Network.continueInterceptedRequest}
\title{Send the command Network.continueInterceptedRequest

Response to Network.requestIntercepted which either modifies the request to continue with any
modifications, or blocks it, or completes it with the provided response bytes. If a network
fetch occurs as a result which encounters a redirect an additional Network.requestIntercepted
event will be sent with the same InterceptionId.}
\usage{
Network.continueInterceptedRequest(promise, interceptionId,
  errorReason = NULL, rawResponse = NULL, url = NULL,
  method = NULL, postData = NULL, headers = NULL,
  authChallengeResponse = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{interceptionId}{A InterceptionId.}

\item{errorReason}{Optional. A ErrorReason.
If set this causes the request to fail with the given reason. Passing \code{Aborted} for requests
marked with \code{isNavigationRequest} also cancels the navigation. Must not be set in response
to an authChallenge.}

\item{rawResponse}{Optional. A character string.
If set the requests completes using with the provided base64 encoded raw response, including
HTTP status line and headers etc... Must not be set in response to an authChallenge.}

\item{url}{Optional. A character string.
If set the request url will be modified in a way that's not observable by page. Must not be
set in response to an authChallenge.}

\item{method}{Optional. A character string.
If set this allows the request method to be overridden. Must not be set in response to an
authChallenge.}

\item{postData}{Optional. A character string.
If set this allows postData to be set. Must not be set in response to an authChallenge.}

\item{headers}{Optional. A Headers.
If set this allows the request headers to be changed. Must not be set in response to an
authChallenge.}

\item{authChallengeResponse}{Optional. A AuthChallengeResponse.
Response to a requestIntercepted with an authChallenge. Must not be set otherwise.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Network.continueInterceptedRequest

Response to Network.requestIntercepted which either modifies the request to continue with any
modifications, or blocks it, or completes it with the provided response bytes. If a network
fetch occurs as a result which encounters a redirect an additional Network.requestIntercepted
event will be sent with the same InterceptionId.
}
