% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Network.R
\name{Network.deleteCookies}
\alias{Network.deleteCookies}
\title{Send the command Network.deleteCookies}
\usage{
Network.deleteCookies(promise, name, url = NULL, domain = NULL,
  path = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{name}{A character string.
Name of the cookies to remove.}

\item{url}{Optional. A character string.
If specified, deletes all the cookies with the given name where domain and path match
provided URL.}

\item{domain}{Optional. A character string.
If specified, deletes only cookies with the exact domain.}

\item{path}{Optional. A character string.
If specified, deletes only cookies with the exact path.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Deletes browser cookies with matching name and url or domain/path pair.
}
