% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.emulateNetworkConditions}
\alias{Network.emulateNetworkConditions}
\title{Send the command Network.emulateNetworkConditions}
\usage{
Network.emulateNetworkConditions(promise, offline, latency,
  downloadThroughput, uploadThroughput, connectionType = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{offline}{A logical.
True to emulate internet disconnection.}

\item{latency}{A numeric.
Minimum latency from request sent to response headers received (ms).}

\item{downloadThroughput}{A numeric.
Maximal aggregated download throughput (bytes/sec). -1 disables download throttling.}

\item{uploadThroughput}{A numeric.
Maximal aggregated upload throughput (bytes/sec).  -1 disables upload throttling.}

\item{connectionType}{Optional. A ConnectionType.
Connection type if known.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Activates emulation of network conditions.
}
