% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.enable}
\alias{Network.enable}
\title{Send the command Network.enable}
\usage{
Network.enable(promise, maxTotalBufferSize = NULL,
  maxResourceBufferSize = NULL, maxPostDataSize = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{maxTotalBufferSize}{Experimental. Optional. An integer.
Buffer size in bytes to use when preserving network payloads (XHRs, etc).}

\item{maxResourceBufferSize}{Experimental. Optional. An integer.
Per-resource buffer size in bytes to use when preserving network payloads (XHRs, etc).}

\item{maxPostDataSize}{Optional. An integer.
Longest post body size (in bytes) that would be included in requestWillBeSent notification}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enables network tracking, network events will now be delivered to the client.
}
