% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.setBypassServiceWorker}
\alias{Network.setBypassServiceWorker}
\title{Send the command Network.setBypassServiceWorker}
\usage{
Network.setBypassServiceWorker(promise, bypass)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{bypass}{A logical.
Bypass service worker and load from network.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Toggles ignoring of service worker for each request.
}
