% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_commands.R
\name{Network.setCacheDisabled}
\alias{Network.setCacheDisabled}
\title{Send the command Network.setCacheDisabled

Toggles ignoring cache for each request. If \code{true}, cache will not be used.}
\usage{
Network.setCacheDisabled(promise, cacheDisabled)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{cacheDisabled}{A logical.
Cache disabled state.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Network.setCacheDisabled

Toggles ignoring cache for each request. If \code{true}, cache will not be used.
}
