% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Network.R
\name{Network.setCookie}
\alias{Network.setCookie}
\title{Send the command Network.setCookie}
\usage{
Network.setCookie(promise, name, value, url = NULL, domain = NULL,
  path = NULL, secure = NULL, httpOnly = NULL, sameSite = NULL,
  expires = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{name}{A character string.
Cookie name.}

\item{value}{A character string.
Cookie value.}

\item{url}{Optional. A character string.
The request-URI to associate with the setting of the cookie. This value can affect the
default domain and path values of the created cookie.}

\item{domain}{Optional. A character string.
Cookie domain.}

\item{path}{Optional. A character string.
Cookie path.}

\item{secure}{Optional. A logical.
True if cookie is secure.}

\item{httpOnly}{Optional. A logical.
True if cookie is http-only.}

\item{sameSite}{Optional. A CookieSameSite.
Cookie SameSite type.}

\item{expires}{Optional. A TimeSinceEpoch.
Cookie expiration date, session cookie if not set}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Sets a cookie with the given cookie data; may overwrite equivalent cookies if they exist.
}
