% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.highlightFrame}
\alias{Overlay.highlightFrame}
\title{Send the command Overlay.highlightFrame}
\usage{
Overlay.highlightFrame(promise, frameId, contentColor = NULL,
  contentOutlineColor = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{frameId}{A Page.FrameId.
Identifier of the frame to highlight.}

\item{contentColor}{Optional. A DOM.RGBA.
The content box highlight fill color (default: transparent).}

\item{contentOutlineColor}{Optional. A DOM.RGBA.
The content box highlight outline color (default: transparent).}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Highlights owner element of the frame with given id.
}
