% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Overlay.R
\name{Overlay.highlightNode}
\alias{Overlay.highlightNode}
\title{Send the command Overlay.highlightNode}
\usage{
Overlay.highlightNode(promise, highlightConfig, nodeId = NULL,
  backendNodeId = NULL, objectId = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{highlightConfig}{A HighlightConfig.
A descriptor for the highlight appearance.}

\item{nodeId}{Optional. A DOM.NodeId.
Identifier of the node to highlight.}

\item{backendNodeId}{Optional. A DOM.BackendNodeId.
Identifier of the backend node to highlight.}

\item{objectId}{Optional. A Runtime.RemoteObjectId.
JavaScript object id of the node to be highlighted.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Highlights DOM node with given id or with the given JavaScript object wrapper. Either nodeId or
objectId must be specified.
}
