% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.highlightRect}
\alias{Overlay.highlightRect}
\title{Send the command Overlay.highlightRect}
\usage{
Overlay.highlightRect(promise, x, y, width, height, color = NULL,
  outlineColor = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{x}{An integer.
X coordinate}

\item{y}{An integer.
Y coordinate}

\item{width}{An integer.
Rectangle width}

\item{height}{An integer.
Rectangle height}

\item{color}{Optional. A DOM.RGBA.
The highlight fill color (default: transparent).}

\item{outlineColor}{Optional. A DOM.RGBA.
The highlight outline color (default: transparent).}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Highlights given rectangle. Coordinates are absolute with respect to the main frame viewport.
}
