% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay_commands.R
\name{Overlay.setShowHitTestBorders}
\alias{Overlay.setShowHitTestBorders}
\title{Send the command Overlay.setShowHitTestBorders}
\usage{
Overlay.setShowHitTestBorders(promise, show)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{show}{A logical.
True for showing hit-test borders}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Requests that backend shows hit-test borders on layers
}
