% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Page.R
\name{Page.addScriptToEvaluateOnNewDocument}
\alias{Page.addScriptToEvaluateOnNewDocument}
\title{Send the command Page.addScriptToEvaluateOnNewDocument}
\usage{
Page.addScriptToEvaluateOnNewDocument(promise, source, worldName = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{source}{A character string.}

\item{worldName}{Experimental. Optional. A character string.
If specified, creates an isolated world with the given name and evaluates given script in it.
This world name will be used as the ExecutionContextDescription::name when the corresponding
event is emitted.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Evaluates given script in every frame upon creation (before loading frame's scripts).
}
