% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.captureScreenshot}
\alias{Page.captureScreenshot}
\title{Send the command Page.captureScreenshot}
\usage{
Page.captureScreenshot(promise, format = NULL, quality = NULL,
  clip = NULL, fromSurface = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{format}{Optional. A character string.
Image compression format (defaults to png). Accepted values: jpeg, png.}

\item{quality}{Optional. An integer.
Compression quality from range \link{0..100} (jpeg only).}

\item{clip}{Optional. A Viewport.
Capture the screenshot of a given region only.}

\item{fromSurface}{Experimental. Optional. A logical.
Capture the screenshot from the surface, rather than the view. Defaults to true.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Capture page screenshot.
}
