% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Page.R
\name{Page.createIsolatedWorld}
\alias{Page.createIsolatedWorld}
\title{Send the command Page.createIsolatedWorld}
\usage{
Page.createIsolatedWorld(promise, frameId, worldName = NULL,
  grantUniveralAccess = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{frameId}{A FrameId.
Id of the frame in which the isolated world should be created.}

\item{worldName}{Optional. A character string.
An optional name which is reported in the Execution Context.}

\item{grantUniveralAccess}{Optional. A logical.
Whether or not universal access should be granted to the isolated world. This is a powerful
option, use with caution.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Creates an isolated world for the given frame.
}
