% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.getNavigationHistory}
\alias{Page.getNavigationHistory}
\title{Send the command Page.getNavigationHistory

Returns navigation history for the current page.}
\usage{
Page.getNavigationHistory(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Send the command Page.getNavigationHistory

Returns navigation history for the current page.
}
