% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.printToPDF}
\alias{Page.printToPDF}
\title{Send the command Page.printToPDF}
\usage{
Page.printToPDF(promise, landscape = NULL, displayHeaderFooter = NULL,
  printBackground = NULL, scale = NULL, paperWidth = NULL,
  paperHeight = NULL, marginTop = NULL, marginBottom = NULL,
  marginLeft = NULL, marginRight = NULL, pageRanges = NULL,
  ignoreInvalidPageRanges = NULL, headerTemplate = NULL,
  footerTemplate = NULL, preferCSSPageSize = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{landscape}{Optional. A logical.
Paper orientation. Defaults to false.}

\item{displayHeaderFooter}{Optional. A logical.
Display header and footer. Defaults to false.}

\item{printBackground}{Optional. A logical.
Print background graphics. Defaults to false.}

\item{scale}{Optional. A numeric.
Scale of the webpage rendering. Defaults to 1.}

\item{paperWidth}{Optional. A numeric.
Paper width in inches. Defaults to 8.5 inches.}

\item{paperHeight}{Optional. A numeric.
Paper height in inches. Defaults to 11 inches.}

\item{marginTop}{Optional. A numeric.
Top margin in inches. Defaults to 1cm (~0.4 inches).}

\item{marginBottom}{Optional. A numeric.
Bottom margin in inches. Defaults to 1cm (~0.4 inches).}

\item{marginLeft}{Optional. A numeric.
Left margin in inches. Defaults to 1cm (~0.4 inches).}

\item{marginRight}{Optional. A numeric.
Right margin in inches. Defaults to 1cm (~0.4 inches).}

\item{pageRanges}{Optional. A character string.
Paper ranges to print, e.g., '1-5, 8, 11-13'. Defaults to the empty string, which means
print all pages.}

\item{ignoreInvalidPageRanges}{Optional. A logical.
Whether to silently ignore invalid but successfully parsed page ranges, such as '3-2'.
Defaults to false.}

\item{headerTemplate}{Optional. A character string.
HTML template for the print header. Should be valid HTML markup with following
classes used to inject printing values into them:
- \code{date}: formatted print date
- \code{title}: document title
- \code{url}: document location
- \code{pageNumber}: current page number
- \code{totalPages}: total pages in the document\preformatted{   For example, `<span class=title></span>` would generate span containing the title.
}}

\item{footerTemplate}{Optional. A character string.
HTML template for the print footer. Should use the same format as the \code{headerTemplate}.}

\item{preferCSSPageSize}{Optional. A logical.
Whether or not to prefer page size as defined by css. Defaults to false,
in which case the content will be scaled to fit the paper size.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 1.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Print page as PDF.
}
