% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Page.R
\name{Page.reload}
\alias{Page.reload}
\title{Send the command Page.reload}
\usage{
Page.reload(promise, ignoreCache = NULL, scriptToEvaluateOnLoad = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{ignoreCache}{Optional. A logical.
If true, browser cache is ignored (as if the user pressed Shift+refresh).}

\item{scriptToEvaluateOnLoad}{Optional. A character string.
If set, the script will be injected into all frames of the inspected page after reload.
Argument will be ignored if reloading dataURL origin.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Reloads given page optionally ignoring the cache.
}
