% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.setAdBlockingEnabled}
\alias{Page.setAdBlockingEnabled}
\title{Send the command Page.setAdBlockingEnabled

Enable Chrome's experimental ad filter on all sites.}
\usage{
Page.setAdBlockingEnabled(promise, enabled)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{enabled}{A logical.
Whether to block ads.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Page.setAdBlockingEnabled

Enable Chrome's experimental ad filter on all sites.
}
