% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.setFontFamilies}
\alias{Page.setFontFamilies}
\title{Send the command Page.setFontFamilies}
\usage{
Page.setFontFamilies(promise, fontFamilies)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{fontFamilies}{A FontFamilies.
Specifies font families to set. If a font family is not specified, it won't be changed.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Set generic font families.
}
