% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Page.R
\name{Page.setWebLifecycleState}
\alias{Page.setWebLifecycleState}
\title{Send the command Page.setWebLifecycleState}
\usage{
Page.setWebLifecycleState(promise, state, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{state}{A character string.
Target lifecycle state Accepted values: frozen, active.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Tries to update the web lifecycle state of the page.
It will transition the page to the given state according to:
https://github.com/WICG/web-lifecycle/
}
