% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Page.R
\name{Page.startScreencast}
\alias{Page.startScreencast}
\title{Send the command Page.startScreencast}
\usage{
Page.startScreencast(promise, format = NULL, quality = NULL,
  maxWidth = NULL, maxHeight = NULL, everyNthFrame = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{format}{Optional. A character string.
Image compression format. Accepted values: jpeg, png.}

\item{quality}{Optional. An integer.
Compression quality from range \code{[0..100]}.}

\item{maxWidth}{Optional. An integer.
Maximum screenshot width.}

\item{maxHeight}{Optional. An integer.
Maximum screenshot height.}

\item{everyNthFrame}{Optional. An integer.
Send every n-th frame.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Starts sending each frame using the \code{screencastFrame} event.
}
