% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_commands.R
\name{Page.stopLoading}
\alias{Page.stopLoading}
\title{Send the command Page.stopLoading}
\usage{
Page.stopLoading(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Force the page stop all navigations and pending resource fetches.
}
