% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profiler_commands.R
\name{Profiler.startPreciseCoverage}
\alias{Profiler.startPreciseCoverage}
\title{Send the command Profiler.startPreciseCoverage}
\usage{
Profiler.startPreciseCoverage(promise, callCount = NULL,
  detailed = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{callCount}{Optional. A logical.
Collect accurate call counts beyond simple 'covered' or 'not covered'.}

\item{detailed}{Optional. A logical.
Collect block-based coverage.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Enable precise code coverage. Coverage data for JavaScript executed before enabling precise code
coverage may be incomplete. Enabling prevents running optimized code and resets execution
counters.
}
