% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Runtime.R
\name{Runtime.callFunctionOn}
\alias{Runtime.callFunctionOn}
\title{Send the command Runtime.callFunctionOn}
\usage{
Runtime.callFunctionOn(promise, functionDeclaration, objectId = NULL,
  arguments = NULL, silent = NULL, returnByValue = NULL,
  generatePreview = NULL, userGesture = NULL, awaitPromise = NULL,
  executionContextId = NULL, objectGroup = NULL, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{functionDeclaration}{A character string.
Declaration of the function to call.}

\item{objectId}{Optional. A RemoteObjectId.
Identifier of the object to call function on. Either objectId or executionContextId should
be specified.}

\item{arguments}{Optional. A list of CallArgument.
Call arguments. All call arguments must belong to the same JavaScript world as the target
object.}

\item{silent}{Optional. A logical.
In silent mode exceptions thrown during evaluation are not reported and do not pause
execution. Overrides \code{setPauseOnException} state.}

\item{returnByValue}{Optional. A logical.
Whether the result is expected to be a JSON object which should be sent by value.}

\item{generatePreview}{Experimental. Optional. A logical.
Whether preview should be generated for the result.}

\item{userGesture}{Optional. A logical.
Whether execution should be treated as initiated by user in the UI.}

\item{awaitPromise}{Optional. A logical.
Whether execution should \code{await} for resulting value and return once awaited promise is
resolved.}

\item{executionContextId}{Optional. A ExecutionContextId.
Specifies execution context which global object will be used to call function on. Either
executionContextId or objectId should be specified.}

\item{objectGroup}{Optional. A character string.
Symbolic group name that can be used to release multiple objects. If objectGroup is not
specified and objectId is, objectGroup will be inherited from object.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 2.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Calls function with given declaration on the given object. Object group of the result is
inherited from the target object.
}
