% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_Runtime.R
\name{Runtime.enable}
\alias{Runtime.enable}
\title{Send the command Runtime.enable}
\usage{
Runtime.enable(promise, awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Enables reporting of execution contexts creation by means of \code{executionContextCreated} event.
When the reporting gets enabled the event will be sent immediately for each existing execution
context.
}
