% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.evaluate}
\alias{Runtime.evaluate}
\title{Send the command Runtime.evaluate}
\usage{
Runtime.evaluate(promise, expression, objectGroup = NULL,
  includeCommandLineAPI = NULL, silent = NULL, contextId = NULL,
  returnByValue = NULL, generatePreview = NULL, userGesture = NULL,
  awaitPromise = NULL, throwOnSideEffect = NULL, timeout = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{expression}{A character string.
Expression to evaluate.}

\item{objectGroup}{Optional. A character string.
Symbolic group name that can be used to release multiple objects.}

\item{includeCommandLineAPI}{Optional. A logical.
Determines whether Command Line API should be available during the evaluation.}

\item{silent}{Optional. A logical.
In silent mode exceptions thrown during evaluation are not reported and do not pause
execution. Overrides \code{setPauseOnException} state.}

\item{contextId}{Optional. A ExecutionContextId.
Specifies in which execution context to perform evaluation. If the parameter is omitted the
evaluation will be performed in the context of the inspected page.}

\item{returnByValue}{Optional. A logical.
Whether the result is expected to be a JSON object that should be sent by value.}

\item{generatePreview}{Experimental. Optional. A logical.
Whether preview should be generated for the result.}

\item{userGesture}{Optional. A logical.
Whether execution should be treated as initiated by user in the UI.}

\item{awaitPromise}{Optional. A logical.
Whether execution should \code{await} for resulting value and return once awaited promise is
resolved.}

\item{throwOnSideEffect}{Experimental. Optional. A logical.
Whether to throw an exception if side effect cannot be ruled out during evaluation.}

\item{timeout}{Experimental. Optional. A TimeDelta.
Terminate execution after timing out (number of milliseconds).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 2.
}
\description{
Evaluates expression on global object.
}
