% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.getProperties}
\alias{Runtime.getProperties}
\title{Send the command Runtime.getProperties}
\usage{
Runtime.getProperties(promise, objectId, ownProperties = NULL,
  accessorPropertiesOnly = NULL, generatePreview = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{objectId}{A RemoteObjectId.
Identifier of the object to return properties for.}

\item{ownProperties}{Optional. A logical.
If true, returns properties belonging only to the element itself, not to its prototype
chain.}

\item{accessorPropertiesOnly}{Experimental. Optional. A logical.
If true, returns accessor properties (with getter/setter) only; internal properties are not
returned either.}

\item{generatePreview}{Experimental. Optional. A logical.
Whether preview should be generated for the results.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a named list of length 3.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Returns properties of a given object. Object group of the result is inherited from the target
object.
}
