% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtime_commands.R
\name{Runtime.releaseObject}
\alias{Runtime.releaseObject}
\title{Send the command Runtime.releaseObject

Releases remote object with given id.}
\usage{
Runtime.releaseObject(promise, objectId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{objectId}{A RemoteObjectId.
Identifier of the object to release.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Runtime.releaseObject

Releases remote object with given id.
}
