% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage_commands.R
\name{Storage.untrackCacheStorageForOrigin}
\alias{Storage.untrackCacheStorageForOrigin}
\title{Send the command Storage.untrackCacheStorageForOrigin

Unregisters origin from receiving notifications for cache storage.}
\usage{
Storage.untrackCacheStorageForOrigin(promise, origin)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{origin}{A character string.
Security origin.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Storage.untrackCacheStorageForOrigin

Unregisters origin from receiving notifications for cache storage.
}
