% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.closeTarget}
\alias{Target.closeTarget}
\title{Send the command Target.closeTarget}
\usage{
Target.closeTarget(promise, targetId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{targetId}{A TargetID.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Closes the target. If the target is a page that gets closed too.
}
