% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.createTarget}
\alias{Target.createTarget}
\title{Send the command Target.createTarget}
\usage{
Target.createTarget(promise, url, width = NULL, height = NULL,
  browserContextId = NULL, enableBeginFrameControl = NULL)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{url}{A character string.
The initial URL the page will be navigated to.}

\item{width}{Optional. An integer.
Frame width in DIP (headless chrome only).}

\item{height}{Optional. An integer.
Frame height in DIP (headless chrome only).}

\item{browserContextId}{Optional. A BrowserContextID.
The browser context to create the page in.}

\item{enableBeginFrameControl}{Experimental. Optional. A logical.
Whether BeginFrames for this target will be controlled via DevTools (headless chrome only,
not supported on MacOS yet, false by default).}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Creates a new page.
}
