% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.disposeBrowserContext}
\alias{Target.disposeBrowserContext}
\title{Send the command Target.disposeBrowserContext

Deletes a BrowserContext. All the belonging pages will be closed without calling their
beforeunload hooks.}
\usage{
Target.disposeBrowserContext(promise, browserContextId)
}
\arguments{
\item{promise}{An aynchronous result object.}

\item{browserContextId}{A BrowserContextID.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 0.
}
\description{
Send the command Target.disposeBrowserContext

Deletes a BrowserContext. All the belonging pages will be closed without calling their
beforeunload hooks.
}
