% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_commands.R
\name{Target.getBrowserContexts}
\alias{Target.getBrowserContexts}
\title{Send the command Target.getBrowserContexts

Returns all browser contexts created with \code{Target.createBrowserContext} method.}
\usage{
Target.getBrowserContexts(promise)
}
\arguments{
\item{promise}{An aynchronous result object.}
}
\value{
A promise (following the definition of the promises package).
The value of the fulfilled promise is a named list of length 1.
}
\description{
Send the command Target.getBrowserContexts

Returns all browser contexts created with \code{Target.createBrowserContext} method.
}
