% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing_commands.R
\name{Testing.generateTestReport}
\alias{Testing.generateTestReport}
\title{Send the command Testing.generateTestReport}
\usage{
Testing.generateTestReport(promise, message, group = NULL,
  awaitResult = TRUE)
}
\arguments{
\item{promise}{An asynchronous result.}

\item{message}{A character string.
Message to be displayed in the report.}

\item{group}{Optional. A character string.
Specifies the endpoint group to deliver the report to.}

\item{awaitResult}{Await for the command result?}
}
\value{
An async value of class \code{promise}.
The value and the completion of the promise differ according to the value of \code{awaitResult}.
Its value is a named list of two elements: \code{ws} (the websocket connexion) and \code{result}.
When \code{awaitResult} is \code{TRUE}, the promise is fulfilled once the result of the command is received. In this case,
\code{result} is a void named list.
When \code{awaitResult} is \code{FALSE}, the promise is fulfilled once the command is sent:
\code{result} is equal to the previous result (\code{promise$result}).
In both cases, you can chain this promise with another command or event listener.
}
\description{
Generates a report for testing.
}
