% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait.R
\name{timeout}
\alias{timeout}
\title{Set a timeout}
\usage{
timeout(delay = 0)
}
\arguments{
\item{delay}{Number of seconds before rejecting the promise.}
}
\value{
A promise that if rejected after a delay of \code{delay} seconds.
}
\description{
Set a timeout
}
