% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait.R
\name{wait}
\alias{wait}
\title{Return a promise after a delay}
\usage{
wait(x, delay = 0)
}
\arguments{
\item{x}{An object.}

\item{delay}{Number of seconds before resolving the promise.}
}
\value{
A \link[promises:promise]{promise}. See details for the value of the
fulfilled promise.
}
\description{
This is a helper function that returns a \link[promises:promise]{promise} after
a delay. It can be used with any pipe and any object (see examples).
}
\details{
The value of the returned promise depends on the class of \code{x}. If \code{x} can
be coerced to a \link[promises:promise]{promise} (using
\link[promises:is.promise]{promises::as.promise()}), the value of the returned
promise is identical to the value of \code{promises::as.promise(x)} once
fulfilled; otherwise the value of the returned promise is \code{x} after the
delay.
}
\examples{
\donttest{
library(promises)

value <- runif(1)
pr <- promise_resolve(value)

# works with `magrittr` pipe
pr \%>\%
  wait(1) \%>\%
  then(~ cat(., "\\n"))

# works with `promises` pipe
pr \%...>\%
  wait(1) \%...>\%
{ cat(., "\\n") }

# also works with any object
value \%>\%
  wait(1) \%>\%
  then(~cat(., "\\n"))
}
}
